Replace uterus 1 with uterus 2
