+++ src/scaffoldmaker/meshtypes/meshtype_3d_uterus1.py
@@ -3,74 +3,864 @@ Generates a 3-D uterus mesh from a 1-D network layout, with variable
 numbers of elements around, along and through wall.
 """
 
-import copy
-import math
-
-from cmlibs.maths.vectorops import add, cross, mult, normalize, sub, magnitude, set_magnitude
+from cmlibs.maths.vectorops import add, cross, mult, set_magnitude, sub, normalize, magnitude, \
+    axis_angle_to_rotation_matrix
 from cmlibs.utils.zinc.field import findOrCreateFieldCoordinates
-from cmlibs.utils.zinc.general import ChangeManager
 from cmlibs.zinc.element import Element
 from cmlibs.zinc.field import Field
 from cmlibs.zinc.node import Node
-from scaffoldmaker.utils.eft_utils import remapEftNodeValueLabel, scaleEftNodeValueLabels, setEftScaleFactorIds
 from scaffoldmaker.annotation.annotationgroup import AnnotationGroup, getAnnotationGroupForTerm, \
     findOrCreateAnnotationGroupForTerm
 from scaffoldmaker.annotation.uterus_terms import get_uterus_term
 from scaffoldmaker.meshtypes.meshtype_1d_network_layout1 import MeshType_1d_network_layout1
 from scaffoldmaker.meshtypes.scaffold_base import Scaffold_base
 from scaffoldmaker.scaffoldpackage import ScaffoldPackage
-from scaffoldmaker.utils import interpolation as interp
-from scaffoldmaker.utils import tubemesh
-from scaffoldmaker.utils.eftfactory_bicubichermitelinear import eftfactory_bicubichermitelinear
-from scaffoldmaker.utils.geometry import createEllipsePoints
-from scaffoldmaker.utils.interpolation import (
-    computeCubicHermiteDerivativeScaling, interpolateCubicHermite, interpolateLagrangeHermiteDerivative,
-    smoothCubicHermiteDerivativesLine)
-from scaffoldmaker.utils.zinc_utils import exnode_string_from_nodeset_field_parameters
-from scaffoldmaker.utils.zinc_utils import get_nodeset_path_ordered_field_parameters
+from scaffoldmaker.utils.interpolation import smoothCurveSideCrossDerivatives, smoothCubicHermiteDerivativesLine,\
+    interpolateCubicHermite, sampleCubicHermiteCurves
+from scaffoldmaker.utils.networkmesh import NetworkMesh, pathValueLabels
+from scaffoldmaker.utils.tubenetworkmesh import TubeNetworkMeshBuilder, TubeNetworkMeshGenerateData, \
+    PatchTubeNetworkMeshSegment
+from scaffoldmaker.utils.zinc_utils import group_add_connected_elements, get_nodeset_path_ordered_field_parameters
 
 
-def get_bifurcation_triple_point(p1x, p1d, p2x, p2d, p3x, p3d):
+import math
+
+class UterusTubeNetworkMeshGenerateData(TubeNetworkMeshGenerateData):
+
+    def __init__(self, region, meshDimension, coordinateFieldName="coordinates",
+                 startNodeIdentifier=1, startElementIdentifier=1, isLinearThroughShell=False, isShowTrimSurfaces=False):
+        """
+        :param isLinearThroughWall: Callers should only set if 3-D with no core.
+        :param isShowTrimSurfaces: Tells junction generateMesh to make 2-D trim surfaces.
+        """
+        super(UterusTubeNetworkMeshGenerateData, self).__init__(
+            region, meshDimension, coordinateFieldName, startNodeIdentifier, startElementIdentifier,
+            isLinearThroughShell, isShowTrimSurfaces)
+        self._fundusGroup = self.getOrCreateAnnotationGroup(get_uterus_term("fundus of uterus"))
+        self._bodyGroup = self.getOrCreateAnnotationGroup(get_uterus_term("body of uterus"))
+        self._upperCervixGroup = self.getOrCreateAnnotationGroup(("upper cervix", ""))
+        self._lowerCervixGroup = self.getOrCreateAnnotationGroup(("lower cervix", ""))
+        # force these annotation group names in base class
+        self._leftGroup = self.getOrCreateAnnotationGroup(("left uterus", ""))
+        self._rightGroup = self.getOrCreateAnnotationGroup(("right uterus", ""))
+        self._dorsalGroup = self.getOrCreateAnnotationGroup(("dorsal uterus", ""))
+        self._ventralGroup = self.getOrCreateAnnotationGroup(("ventral uterus", ""))
+
+    def getFundusMeshGroup(self):
+        return self._fundusGroup.getMeshGroup(self._mesh)
+
+    def getBodyMeshGroup(self):
+        return self._bodyGroup.getMeshGroup(self._mesh)
+
+    def getUpperCervixMeshGroup(self):
+        return self._upperCervixGroup.getMeshGroup(self._mesh)
+
+    def getLowerCervixMeshGroup(self):
+        return self._lowerCervixGroup.getMeshGroup(self._mesh)
+
+class UterusTubeNetworkMeshBuilder(TubeNetworkMeshBuilder):
     """
-    Get coordinates and derivatives of triple point between p1, p2 and p3 with derivatives.
-    :param p1x..p3d: Point coordinates and derivatives, numbered anticlockwise around triple point.
-    All derivatives point away from triple point.
-    Returned d1 points from triple point to p2, d2 points from triple point to p3.
-    :return: x, d1, d2
+    Adds left, right, dorsal, ventral, fundus, body annotations.
+    Future: derive from BodyTubeNetworkMeshBuilder to get left/right/dorsal/ventral.
     """
-    scaling12 = computeCubicHermiteDerivativeScaling(p1x, [-d for d in p1d], p2x, p2d)
-    scaling23 = computeCubicHermiteDerivativeScaling(p2x, [-d for d in p2d], p3x, p3d)
-    scaling31 = computeCubicHermiteDerivativeScaling(p3x, [-d for d in p3d], p1x, p1d)
-    trx1 = interpolateCubicHermite(p1x, mult(p1d, -scaling12), p2x, mult(p2d, scaling12), 0.5)
-    trx2 = interpolateCubicHermite(p2x, mult(p2d, -scaling23), p3x, mult(p3d, scaling23), 0.5)
-    trx3 = interpolateCubicHermite(p3x, mult(p3d, -scaling31), p1x, mult(p1d, scaling31), 0.5)
-    trx = [(trx1[c] + trx2[c] + trx3[c]) / 3.0 for c in range(3)]
-    td1 = interpolateLagrangeHermiteDerivative(trx, p1x, p1d, 0.0)
-    td2 = interpolateLagrangeHermiteDerivative(trx, p2x, p2d, 0.0)
-    td3 = interpolateLagrangeHermiteDerivative(trx, p3x, p3d, 0.0)
-    n12 = cross(td1, td2)
-    n23 = cross(td2, td3)
-    n31 = cross(td3, td1)
-    norm = normalize([(n12[c] + n23[c] + n31[c]) for c in range(3)])
-    sd1 = smoothCubicHermiteDerivativesLine([trx, p1x], [normalize(cross(norm, cross(td1, norm))), p1d],
-                                            fixStartDirection=True, fixEndDerivative=True)[0]
-    sd2 = smoothCubicHermiteDerivativesLine([trx, p2x], [normalize(cross(norm, cross(td2, norm))), p2d],
-                                            fixStartDirection=True, fixEndDerivative=True)[0]
-    sd3 = smoothCubicHermiteDerivativesLine([trx, p3x], [normalize(cross(norm, cross(td3, norm))), p3d],
-                                            fixStartDirection=True, fixEndDerivative=True)[0]
-    trd1 = mult(sub(sd2, add(sd3, sd1)), 0.5)
-    trd2 = mult(sub(sd3, add(sd1, sd2)), 0.5)
-    return trx, trd1, trd2
+
+    def createSegment(self, networkSegment):
+        if networkSegment.isPatch():
+            pathParametersList = [get_nodeset_path_ordered_field_parameters(
+                self._layoutNodes, self._layoutCoordinates, pathValueLabels,
+                networkSegment.getNodeIdentifiers(), networkSegment.getNodeVersions())]
+            if self._layoutInnerCoordinates:
+                pathParametersList.append(get_nodeset_path_ordered_field_parameters(
+                    self._layoutNodes, self._layoutInnerCoordinates, pathValueLabels,
+                    networkSegment.getNodeIdentifiers(), networkSegment.getNodeVersions()))
+            elementsCountAround = self._defaultElementsCountAround
+            elementsCountCoreBoxMinor = self._defaultElementsCountCoreBoxMinor
+            coreBoundaryScalingMode = self._defaultCoreBoundaryScalingMode
+
+            return PatchTubeNetworkMeshSegment(networkSegment, pathParametersList, elementsCountAround,
+                                               self._elementsCountThroughShell, self._isCore, elementsCountCoreBoxMinor,
+                                               self._elementsCountTransition, coreBoundaryScalingMode)
+
+        return super(UterusTubeNetworkMeshBuilder, self).createSegment(networkSegment)
+
+    def generateMesh(self, generateData):
+        super(UterusTubeNetworkMeshBuilder, self).generateMesh(generateData)
+        # build temporary left/right dorsal/ventral groups
+        mesh = generateData.getMesh()
+        fundusMeshGroup = generateData.getFundusMeshGroup()
+        bodyMeshGroup = generateData.getBodyMeshGroup()
+        upperCervixMeshGroup = generateData.getUpperCervixMeshGroup()
+        lowerCervixMeshGroup = generateData.getLowerCervixMeshGroup()
+        leftMeshGroup = generateData.getLeftMeshGroup()
+        rightMeshGroup = generateData.getRightMeshGroup()
+        dorsalMeshGroup = generateData.getDorsalMeshGroup()
+        ventralMeshGroup = generateData.getVentralMeshGroup()
+        for networkSegment in self._networkMesh.getNetworkSegments():
+            segment = self._segments[networkSegment]
+            annotationTerms = segment.getAnnotationTerms()
+            segment.addSideD3ElementsToMeshGroup(True, ventralMeshGroup)
+            segment.addSideD3ElementsToMeshGroup(False, dorsalMeshGroup)
+            for annotationTerm in annotationTerms:
+                if "fundus" in annotationTerm[0]:
+                    segment.addAllElementsToMeshGroup(fundusMeshGroup)
+                if "body" in annotationTerm[0]:
+                    segment.addAllElementsToMeshGroup(bodyMeshGroup)
+                if "cervix" in annotationTerm[0]:
+                    elementsCountRim = segment.getElementsCountRim()
+                    elementsCountAlong = segment.getSampledElementsCountAlong()
+                    elementsCountAround = segment.getElementsCountAround()
+                    for e1 in range(elementsCountAround):
+                        for e2 in range(elementsCountAlong):
+                            for e3 in range(elementsCountRim):
+                                elementIdentifier = segment.getRimElementId(e1, e2, e3)
+                                if elementIdentifier is not None:
+                                    element = mesh.findElementByIdentifier(elementIdentifier)
+                                    if e2 >= int(elementsCountAlong * 0.5):
+                                        lowerCervixMeshGroup.addElement(element)
+                                    else:
+                                        upperCervixMeshGroup.addElement(element)
+                if "left" in annotationTerm[0]:
+                    segment.addAllElementsToMeshGroup(leftMeshGroup)
+                    break
+                if "right" in annotationTerm[0]:
+                    segment.addAllElementsToMeshGroup(rightMeshGroup)
+                    break
+            else:
+                # segment on main axis
+                segment.addSideD2ElementsToMeshGroup(True, leftMeshGroup)
+                segment.addSideD2ElementsToMeshGroup(False, rightMeshGroup)
 
 
-def get_tube_bifurcation_connection_elements_counts(tCounts):
+class MeshType_1d_uterus_network_layout1(MeshType_1d_network_layout1):
     """
-    Get number of elements directly connecting tubes 1, 2 and 3 from the supplied number around.
-    :param tCounts: Number of elements around tubes in order.
-    :return: List of elements connect tube with its next neighbour, looping back to first.
+    Defines uterus network layout.
     """
-    assert len(tCounts) == 3
-    return [(tCounts[i] + tCounts[i - 2] - tCounts[i - 1]) // 2 for i in range(3)]
+
+    @classmethod
+    def getName(cls):
+        return "1D Uterus Network Layout 1"
+
+    @classmethod
+    def getParameterSetNames(cls):
+        return ["Default",
+                "Human 1",
+                "Human Pregnant 1",
+                "Mouse 1",
+                "Rat 1"]
+
+    @classmethod
+    def getDefaultOptions(cls, parameterSetName="Default"):
+        options = {}
+        options["Base parameter set"] = parameterSetName
+        if parameterSetName in ("Mouse 1", "Rat 1"):
+            options["Structure"] = (
+                "1-2-3-4-5-6-7-8-9-10-11-12-13-14-15-16-36.1,"
+                "17-18-19-20-21-22-23-24-25-26-27-28-29-30-31-32-36.2,"
+                "#-33-34-35-36.3,"
+                "36.4-37-38-39,"
+                "39-40,"
+                "40-41-42")
+            options["Oviduct diameter"] = 0.5
+            options["Oviduct length"] = 4.0
+            options["Body length"] = 0.75
+            options["Fundus width between oviducts"] = 0.8
+            options["Fundus depth between oviducts"] = 0.5
+            options["Cervical length"] = 0.25
+            options["Cervical width around internal os"] = 0.8
+            options["Cervical depth around internal os"] = 0.5
+            options["Cervical width around external os"] = 0.8
+            options["Cervical depth around external os"] = 0.5
+            options["Vagina length"] = 0.5
+            options["Vagina width around vagina orifice"] = 0.8
+            options["Vagina depth around vagina orifice"] = 0.5
+            options["Inner proportion body"] = 0.75
+            options["Inner proportion cervix"] = 0.75
+            options["Inner proportion vagina"] = 0.75
+            options["Angle of anteversion degrees"] = 0.0
+        elif parameterSetName == "Human Pregnant 1":
+            options["Structure"] = (
+                "1-2-3-4-5-6-7-8-9-31.1,"
+                "10-11-12-13-14-15-16-17-18-31.2,"                
+                "#-19-20-21-22-23-24-25-26-27-28-29-30-31.3,"
+                "31.4-32-33-34-35-36-37-38-39-40-41-42-43,"
+                "43-44,"
+                "44-45-46-47-48")
+            options["Oviduct diameter"] = 0.35
+            options["Oviduct length"] = 10.0
+            options["Body length"] = 30.0
+            options["Fundus width between oviducts"] = 28.0
+            options["Fundus depth between oviducts"] = 28.0
+            options["Cervical length"] = 1.0
+            options["Cervical width around internal os"] = 5.5
+            options["Cervical depth around internal os"] = 3.8
+            options["Cervical width around external os"] = 5.0
+            options["Cervical depth around external os"] = 3.0
+            options["Vagina length"] = 10.0
+            options["Vagina width around vagina orifice"] = 1.25
+            options["Vagina depth around vagina orifice"] = 1.25
+            options["Inner proportion body"] = 0.8
+            options["Inner proportion cervix"] = 0.15
+            options["Inner proportion vagina"] = 0.8
+            options["Angle of anteversion degrees"] = 70.0
+        else:
+            options["Structure"] = (
+                "1-2-3-4-5-6-7-8-23.1,"
+                "9-10-11-12-13-14-15-16-23.2,"
+                "#-17-18-19-20-21-22-23.3,"
+                "23.4-24-25-26-27-28-29,"
+                "29-30-31,"
+                "31-32-33-34-35-36-37-38")
+            options["Oviduct diameter"] = 0.35
+            options["Oviduct length"] = 10.0
+            options["Body length"] = 7.0
+            options["Fundus width between oviducts"] = 8.0
+            options["Fundus depth between oviducts"] = 6.0
+            options["Cervical length"] = 1.0
+            options["Cervical width around internal os"] = 5.5
+            options["Cervical depth around internal os"] = 3.8
+            options["Cervical width around external os"] = 5.0
+            options["Cervical depth around external os"] = 3.0
+            options["Vagina length"] = 10.0
+            options["Vagina width around vagina orifice"] = 1.25
+            options["Vagina depth around vagina orifice"] = 1.25
+            options["Inner proportion body"] = 0.75
+            options["Inner proportion cervix"] = 0.15
+            options["Inner proportion vagina"] = 0.8
+            options["Angle of anteversion degrees"] = 70.0
+
+        options["Define inner coordinates"] = True
+        options["Inner proportion oviducts"] = 0.5
+
+        return options
+
+    @classmethod
+    def getOrderedOptionNames(cls):
+        return [
+            "Oviduct diameter",
+            "Oviduct length",
+            "Body length",
+            "Fundus width between oviducts",
+            "Fundus depth between oviducts",
+            "Cervical length",
+            "Cervical width around internal os",
+            "Cervical depth around internal os",
+            "Cervical width around external os",
+            "Cervical depth around external os",
+            "Vagina length",
+            "Vagina width around vagina orifice",
+            "Vagina depth around vagina orifice",
+            "Inner proportion oviducts",
+            "Inner proportion body",
+            "Inner proportion cervix",
+            "Inner proportion vagina",
+            "Angle of anteversion degrees"
+        ]
+
+    @classmethod
+    def checkOptions(cls, options):
+        dependentChanges = False
+        for key in [
+            "Oviduct diameter",
+            "Oviduct length",
+            "Body length",
+            "Fundus width between oviducts",
+            "Fundus depth between oviducts",
+            "Cervical length",
+            "Cervical width around internal os",
+            "Cervical depth around internal os",
+            "Cervical width around external os",
+            "Cervical depth around external os",
+            "Vagina length",
+            "Vagina width around vagina orifice",
+            "Vagina depth around vagina orifice"
+        ]:
+            if options[key] < 0.1:
+                options[key] = 0.1
+        for key in [
+            "Inner proportion oviducts",
+            "Inner proportion body",
+            "Inner proportion cervix",
+            "Inner proportion vagina",
+        ]:
+            if options[key] < 0.1:
+                options[key] = 0.1
+            elif options[key] > 0.9:
+                options[key] = 0.9
+
+        for key, angleRange in {
+            "Angle of anteversion degrees": (-100.0, 100.0)
+        }.items():
+            if options[key] < angleRange[0]:
+                options[key] = angleRange[0]
+            elif options[key] > angleRange[1]:
+                options[key] = angleRange[1]
+        return dependentChanges
+
+    @classmethod
+    def generateBaseMesh(cls, region, options):
+        """
+        Generate the unrefined mesh.
+        :param region: Zinc region to define model in. Must be empty.
+        :param options: Dict containing options. See getDefaultOptions().
+        :return: [] empty list of AnnotationGroup, NetworkMesh
+        """
+        parameterSetName = options['Base parameter set']
+        structure = options["Structure"]
+        oviductLength = options["Oviduct length"]
+        oviductRadius = 0.5 * options["Oviduct diameter"]
+        bodyLength = options["Body length"]
+        halfFundusWidth = 0.5 * options["Fundus width between oviducts"]
+        halfFundusDepth = 0.5 * options["Fundus depth between oviducts"]
+        cervicalLength = options["Cervical length"]
+        halfCervicalWidthInternalOs = 0.5 * options["Cervical width around internal os"]
+        halfCervicalDepthInternalOs = 0.5 * options["Cervical depth around internal os"]
+        halfCervicalWidthExternalOs = 0.5 * options["Cervical width around external os"]
+        halfCervicalDepthExternalOs = 0.5 * options["Cervical depth around external os"]
+        vaginaLength = options["Vagina length"]
+        halfVaginaOrificeWidth = 0.5 * options["Vagina width around vagina orifice"]
+        halfVaginaOrificeDepth = 0.5 * options["Vagina depth around vagina orifice"]
+        anteversionAngleRad = math.radians(options["Angle of anteversion degrees"])
+        innerProportionOviducts = options["Inner proportion oviducts"]
+        innerProportionBody = options["Inner proportion body"]
+        innerProportionCervix = options["Inner proportion cervix"]
+        innerProportionVagina = options["Inner proportion vagina"]
+
+        isPregnant = parameterSetName in 'Human Pregnant 1'
+        isRat = parameterSetName in 'Rat 1'
+        isRodent = parameterSetName in ("Mouse 1", "Rat 1")
+
+        networkMesh = NetworkMesh(structure)
+        networkMesh.create1DLayoutMesh(region)
+
+        fieldmodule = region.getFieldmodule()
+        mesh = fieldmodule.findMeshByDimension(1)
+
+        # set up element annotations
+        uterusGroup = AnnotationGroup(region, get_uterus_term("uterus"))
+        if isRodent:
+            leftOviductGroup = AnnotationGroup(region, get_uterus_term("left uterine horn"))
+            rightOviductGroup = AnnotationGroup(region, get_uterus_term("right uterine horn"))
+        else:
+            leftOviductGroup = AnnotationGroup(region, get_uterus_term("left oviduct"))
+            rightOviductGroup = AnnotationGroup(region, get_uterus_term("right oviduct"))
+        bodyGroup = AnnotationGroup(region, get_uterus_term("body of uterus"))
+        cervixGroup = AnnotationGroup(region, ("cervix", ""))
+        vaginaGroup = AnnotationGroup(region, get_uterus_term("vagina"))
+        fundusPatchGroup = AnnotationGroup(region, get_uterus_term("fundus of uterus"))
+        annotationGroups = [uterusGroup, leftOviductGroup, rightOviductGroup, fundusPatchGroup, bodyGroup,
+                            cervixGroup, vaginaGroup]
+
+        uterusMeshGroup = uterusGroup.getMeshGroup(mesh)
+        elementIdentifier = 1
+
+        left = 0
+        right = 1
+
+        if isRodent:
+            oviductElementsCount = 16
+            fundusPatchElementsCount = 3
+            fundusPostBodyJunctionElementsCount = 3
+            cervixElementsCount = 1
+            vaginaElementsCount = 2
+        elif isPregnant:
+            oviductElementsCount = 9
+            fundusPatchElementsCount = 12
+            fundusPostBodyJunctionElementsCount = 12
+            cervixElementsCount = 1
+            vaginaElementsCount = 4
+        else:
+            oviductElementsCount = 8
+            fundusPatchElementsCount = 6
+            fundusPostBodyJunctionElementsCount = 6
+            cervixElementsCount = 2
+            vaginaElementsCount = 7
+
+        for side in (left, right):
+            sideOviductGroup = leftOviductGroup if (side == left) else rightOviductGroup
+            meshGroups = [uterusMeshGroup, sideOviductGroup.getMeshGroup(mesh)]
+            for e in range(oviductElementsCount):
+                element = mesh.findElementByIdentifier(elementIdentifier)
+                for meshGroup in meshGroups:
+                    meshGroup.addElement(element)
+                elementIdentifier += 1
+
+        meshGroups = [uterusMeshGroup, fundusPatchGroup.getMeshGroup(mesh)]
+        for e in range(fundusPatchElementsCount):
+            element = mesh.findElementByIdentifier(elementIdentifier)
+            for meshGroup in meshGroups:
+                meshGroup.addElement(element)
+            elementIdentifier += 1
+
+        meshGroups = [uterusMeshGroup, bodyGroup.getMeshGroup(mesh)]
+        for e in range(fundusPostBodyJunctionElementsCount):
+            element = mesh.findElementByIdentifier(elementIdentifier)
+            for meshGroup in meshGroups:
+                meshGroup.addElement(element)
+            elementIdentifier += 1
+
+        meshGroups = [cervixGroup.getMeshGroup(mesh)]
+        for e in range(cervixElementsCount):
+            element = mesh.findElementByIdentifier(elementIdentifier)
+            for meshGroup in meshGroups:
+                meshGroup.addElement(element)
+            elementIdentifier += 1
+
+        meshGroups = [vaginaGroup.getMeshGroup(mesh)]
+        for e in range(vaginaElementsCount):
+            element = mesh.findElementByIdentifier(elementIdentifier)
+            for meshGroup in meshGroups:
+                meshGroup.addElement(element)
+            elementIdentifier += 1
+
+        # set coordinates (outer)
+        fieldcache = fieldmodule.createFieldcache()
+        coordinates = findOrCreateFieldCoordinates(fieldmodule)
+        # need to ensure inner coordinates are at least defined:
+        cls.defineInnerCoordinates(region, coordinates, options, networkMesh, innerProportion=0.75)
+        innerCoordinates = findOrCreateFieldCoordinates(fieldmodule, "inner coordinates")
+        nodes = fieldmodule.findNodesetByFieldDomainType(Field.DOMAIN_TYPE_NODES)
+        nodeIdentifier = 1
+
+        # Calculate scales
+        fundusScalePostBodyJunction = bodyLength / fundusPostBodyJunctionElementsCount
+        cervicalScale = cervicalLength / cervixElementsCount
+        vaginalScale = vaginaLength / vaginaElementsCount
+
+        zero = [0.0, 0.0, 0.0]
+        xBodyJunction = [0.0, 0.0, 0.0]
+
+        if isPregnant:
+            aThetaCervicalEnd = math.asin(halfCervicalWidthInternalOs / halfFundusWidth)
+            aEllipse = bodyLength / math.cos(aThetaCervicalEnd)
+        cThetaCervicalEnd = math.asin(halfCervicalDepthInternalOs / halfFundusDepth)
+        cEllipse = bodyLength / math.cos(cThetaCervicalEnd)
+
+        # Oviducts
+        d1BodyJunction = []
+        d2BodyJunction = []
+        d3BodyJunction = []
+        d12BodyJunction = []
+        d13BodyJunction = []
+        id2BodyJunction = []
+        id3BodyJunction = []
+        id12BodyJunction = []
+        id13BodyJunction = []
+
+        if isRodent:
+            rC = bodyLength
+            thetaLimit = math.radians(55.0)
+            for side in (left, right):
+                rTheta = rC * thetaLimit
+                straightLength = oviductLength - rTheta
+                theta = thetaLimit * (-1.0 if side == left else 1.0)
+                xCurveEnd = add([rC * math.cos(theta), rC * math.sin(theta), 0.0], [-rC, 0.0, 0.0])
+                d1 = [(-1.0 if side == left else 1.0) * rC * math.sin(theta),
+                      (1.0 if side == left else -1.0) * rC * math.cos(theta),
+                      0.0]
+                xStart = sub(xCurveEnd, set_magnitude(d1, straightLength))
+                xEnd = xBodyJunction
+                nx = [xStart, xCurveEnd, xEnd]
+                nd1 = [set_magnitude(d1, straightLength), set_magnitude(d1, rTheta),
+                       [0.0, (1.0 if side == left else -1.0) * rTheta, 0.0]]
+                xOviduct, d1Oviduct = sampleCubicHermiteCurves(nx, nd1, oviductElementsCount)[0:2]
+
+                for i in range(oviductElementsCount + 1):
+                    x = xOviduct[i]
+                    d1 = d1Oviduct[i]
+                    d3 = [0.0, 0.0, oviductRadius]
+                    d2Direction = cross(normalize(d3), normalize([-c for c in d1]))
+                    d2 = set_magnitude(d2Direction, -oviductRadius)
+                    d12 = zero
+                    d13 = zero
+                    id2 = mult(d2, innerProportionOviducts)
+                    id3 = mult(d3, innerProportionOviducts)
+
+                    if i < oviductElementsCount:
+                        node = nodes.findNodeByIdentifier(nodeIdentifier)
+                        fieldcache.setNode(node)
+                        setNodeFieldParameters(coordinates, fieldcache, x, d1, d2, d3)
+                        setNodeFieldParameters(innerCoordinates, fieldcache, x, d1, id2, id3)
+                        nodeIdentifier += 1
+                    else:
+                        d1BodyJunction.append(d1)
+                        d2BodyJunction.append(d2)
+                        d3BodyJunction.append(d3)
+                        d12BodyJunction.append(d12)
+                        d13BodyJunction.append(d13)
+                        id2BodyJunction.append(id2)
+                        id3BodyJunction.append(id3)
+                        id12BodyJunction.append(d12)
+                        id13BodyJunction.append(d13)
+
+        else:
+            segment1Length = halfFundusWidth + oviductLength
+            segment1LengthScale = segment1Length / oviductElementsCount
+            elementsAlongHalfFundusWidth = math.ceil(halfFundusWidth / segment1LengthScale)
+
+            for side in (left, right):
+                xStart = [0.0, segment1Length * (-1.0 if side == left else 1.0), 0.0]
+                d1Oviduct = [0.0, segment1LengthScale * (1.0 if side == left else -1.0), 0.0]
+                for i in range(oviductElementsCount + 1):
+                    x = add(xStart, mult(d1Oviduct, i))
+                    if segment1LengthScale * i < oviductLength:
+                        d2 = [-oviductRadius if side == left else oviductRadius, 0.0, 0.0]
+                        d3 = [0.0, 0.0, oviductRadius]
+                        d12 = [0.0, 0.0, 0.0]
+                        d13 = [0.0, 0.0, 0.0]
+                    else: # in ellipse zone
+                        theta = math.acos(x[1] / halfFundusWidth)
+                        if abs(halfFundusDepth * math.sin(theta)) < oviductRadius:
+                            d2 = [-oviductRadius if side == left else oviductRadius, 0.0, 0.0]
+                            d3 = [0.0, 0.0, oviductRadius]
+                            d12 = [0.0, 0.0, 0.0]
+                            d13 = [0.0, 0.0, 0.0]
+                        else:
+                            d2 = [halfFundusDepth * math.sin(theta) * (-1.0 if side == left else 1.0), 0.0, 0.0]
+                            d3 = [0.0, 0.0, halfFundusDepth * math.sin(theta)]
+                            d12 = [halfFundusDepth * math.cos(theta) * (0.5 * math.pi / elementsAlongHalfFundusWidth),
+                                   0.0, 0.0]
+                            d13 = [0.0, 0.0,
+                                   halfFundusDepth * math.cos(theta) * (0.5 * math.pi / elementsAlongHalfFundusWidth) *
+                                   (-1.0 if side == left else 1.0)]
+
+                    id2 = mult(d2, innerProportionOviducts)
+                    id3 = mult(d3, innerProportionOviducts)
+                    id12 = mult(d12, innerProportionOviducts)
+                    id13 = mult(d13, innerProportionOviducts)
+
+                    if i < oviductElementsCount:
+                        node = nodes.findNodeByIdentifier(nodeIdentifier)
+                        fieldcache.setNode(node)
+                        setNodeFieldParameters(coordinates, fieldcache, x, d1Oviduct, d2, d3, d12, d13)
+                        setNodeFieldParameters(innerCoordinates, fieldcache, x, d1Oviduct, id2, id3, id12, id13)
+                        nodeIdentifier += 1
+                    else:
+                        d1BodyJunction.append(d1Oviduct)
+                        d2BodyJunction.append(d2)
+                        d3BodyJunction.append(d3)
+                        d12BodyJunction.append(d12)
+                        d13BodyJunction.append(d13)
+                        id2BodyJunction.append(id2)
+                        id3BodyJunction.append(id3)
+                        id12BodyJunction.append(id12)
+                        id13BodyJunction.append(id13)
+
+        xFundusPatchStart = [-bodyLength, 0.0, 0.0]
+        d1FundusPatch = [bodyLength / fundusPatchElementsCount, 0.0, 0.0]
+        nxPatch = []
+        nd1Patch = []
+        nd2Patch = []
+        nd3Patch = []
+
+        for i in range(fundusPatchElementsCount + 1):
+            x = [xFundusPatchStart[0] + d1FundusPatch[0] * i, 0.0, 0.0]
+            nxPatch.append(x)
+            nd1Patch.append(d1FundusPatch)
+            xi = i / fundusPatchElementsCount
+            width = xi * halfFundusWidth + (1.0 - xi) * (halfFundusWidth * 0.01 if isRodent else
+                                                         halfCervicalWidthInternalOs)
+            if isPregnant:
+                thetaA = math.acos(x[0] / aEllipse)
+                width = halfFundusWidth * math.sin(thetaA)
+            thetaC = math.acos(x[0] / cEllipse)
+            depth = halfFundusDepth * math.sin(thetaC)
+            nd2Patch.append([0.0, width, 0.0])
+            nd3Patch.append([0.0, 0.0, depth])
+        nd13Patch = smoothCurveSideCrossDerivatives(nxPatch, nd1Patch, [nd3Patch])[0]
+        nd12Patch = smoothCurveSideCrossDerivatives(nxPatch, nd1Patch, [nd2Patch])[0]
+
+        for i in range(fundusPatchElementsCount):
+            node = nodes.findNodeByIdentifier(nodeIdentifier)
+            fieldcache.setNode(node)
+            id3 = mult(nd3Patch[i], innerProportionBody)
+            id13 = mult(nd13Patch[i], innerProportionBody)
+            if isPregnant:
+                id2 = mult(nd2Patch[i], innerProportionBody)
+                id12 = mult(nd12Patch[i], innerProportionBody)
+            else:
+                xi = i / fundusPatchElementsCount
+                width = xi * halfFundusWidth * innerProportionBody + \
+                        (1.0 - xi) * (halfFundusWidth * 0.01 * innerProportionBody if isRodent else
+                                      halfCervicalWidthInternalOs * innerProportionCervix)
+
+                id2 = [0.0, width, 0.0]
+                id12 = [0.0,
+                        (halfFundusWidth * innerProportionBody -
+                         halfCervicalWidthInternalOs * innerProportionCervix) / fundusPatchElementsCount,
+                        0.0]
+            setNodeFieldParameters(coordinates, fieldcache, nxPatch[i], nd1Patch[i], nd2Patch[i], nd3Patch[i],
+                                   nd12Patch[i], nd13Patch[i])
+            setNodeFieldParameters(innerCoordinates, fieldcache, nxPatch[i], nd1Patch[i], id2, id3, id12, id13)
+            nodeIdentifier += 1
+
+        # body junction
+        node = nodes.findNodeByIdentifier(nodeIdentifier)
+        fieldcache.setNode(node)
+        for field in (coordinates, innerCoordinates):
+            field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_VALUE, 1, xBodyJunction)
+        for side in (left, right):
+            version = 1 if side == left else 2
+            d1 = d1BodyJunction[version - 1]
+            d2 = d2BodyJunction[version - 1]
+            d3 = d3BodyJunction[version - 1]
+            d12 = d12BodyJunction[version - 1]
+            d13 = d13BodyJunction[version - 1]
+            id2 = id2BodyJunction[version - 1]
+            id3 = id3BodyJunction[version - 1]
+            id12 = id12BodyJunction[version - 1]
+            id13 = id13BodyJunction[version - 1]
+            setNodeFieldVersionDerivatives(coordinates, fieldcache, version, d1, d2, d3, d12, d13)
+            setNodeFieldVersionDerivatives(innerCoordinates, fieldcache, version, d1, id2, id3, id12, id13)
+
+        version = 3
+        id3 = mult(nd3Patch[-1], innerProportionBody)
+        id13 = mult(nd13Patch[-1], innerProportionBody)
+        if isPregnant:
+            id2 = mult(nd2Patch[-1], innerProportionBody)
+            id12 = mult(nd12Patch[-1], innerProportionBody)
+        else:
+            id2 = [0.0, halfFundusWidth * innerProportionBody, 0.0]
+            id12 = [0.0,
+                    (halfCervicalWidthInternalOs * innerProportionCervix -
+                     halfFundusWidth * innerProportionBody) / fundusPatchElementsCount,
+                    0.0]
+        setNodeFieldVersionDerivatives(coordinates, fieldcache, version, nd1Patch[-1], nd2Patch[-1], nd3Patch[-1],
+                                       nd12Patch[-1], nd13Patch[-1])
+        setNodeFieldVersionDerivatives(innerCoordinates, fieldcache, version, nd1Patch[-1], id2, id3, id12, id13)
+
+        # Post body junction
+        nd12 = []
+        nd13 = []
+        nxBody = []
+        nd1Body = []
+        nd2Body = []
+        nd3Body = []
+
+        dWidth = (halfCervicalWidthInternalOs - halfFundusWidth) / fundusPostBodyJunctionElementsCount
+        dDepth = (halfCervicalDepthInternalOs - halfFundusDepth) / fundusPostBodyJunctionElementsCount
+
+        for i in range(fundusPostBodyJunctionElementsCount + 1):
+            x = [fundusScalePostBodyJunction * i, 0.0, 0.0]
+            d1 = [fundusScalePostBodyJunction, 0.0, 0.0]
+            xi = i / fundusPostBodyJunctionElementsCount
+            width = xi * halfCervicalWidthInternalOs + (1.0 - xi) * halfFundusWidth
+            if isPregnant:
+                thetaA = math.acos(x[0] / aEllipse)
+                width = halfFundusWidth * math.sin(thetaA)
+            thetaC = math.acos(x[0] / cEllipse)
+            depth = halfFundusDepth * math.sin(thetaC)
+            d2 = [0.0, width, 0.0]
+            d3 = [0.0, 0.0, depth]
+            nxBody.append(x)
+            nd1Body.append(d1)
+            nd2Body.append(d2)
+            nd3Body.append(d3)
+            nd12.append([0.0, dWidth, 0.0])
+            nd13.append([0.0, 0.0, dDepth])
+
+        if dWidth == 0.0:
+            nd12Body = nd12
+        else:
+            nd12Body = smoothCubicHermiteDerivativesLine(nd2Body, nd12)
+        if dDepth == 0.0:
+            nd13Body = nd13
+        else:
+            nd13Body = smoothCubicHermiteDerivativesLine(nd3Body, nd13)
+
+        for i in range(fundusPostBodyJunctionElementsCount + 1):
+            x = nxBody[i]
+            d1 = nd1Body[i]
+            d2 = nd2Body[i]
+            d3 = nd3Body[i]
+            d12 = nd12Body[i]
+            d13 = nd13Body[i]
+            id3 = mult(d3, innerProportionBody)
+            id13 = mult(d13, innerProportionBody)
+            if isPregnant:
+                id2 = mult(d2, innerProportionBody)
+                id12 = mult(d12, innerProportionBody)
+            else:
+                xi = i / fundusPostBodyJunctionElementsCount
+                width = xi * halfCervicalWidthInternalOs * innerProportionCervix + \
+                        (1.0 - xi) * halfFundusWidth * innerProportionBody
+                id2 = [0.0, width, 0.0]
+                id12 = [0.0,
+                        (halfCervicalWidthInternalOs * innerProportionCervix -
+                         halfFundusWidth * innerProportionBody) / fundusPatchElementsCount,
+                        0.0]
+
+            if i:
+                node = nodes.findNodeByIdentifier(nodeIdentifier)
+                fieldcache.setNode(node)
+                setNodeFieldParameters(coordinates, fieldcache, x, d1, d2, d3, d12, d13)
+                setNodeFieldParameters(innerCoordinates, fieldcache, x, d1, id2, id3, id12, id13)
+            else:
+                version = 4
+                setNodeFieldVersionDerivatives(coordinates, fieldcache, version, d1, d2, d3, d12, d13)
+                setNodeFieldVersionDerivatives(innerCoordinates, fieldcache, version, d1, id2, id3, id12, id13)
+            nodeIdentifier += 1
+
+        # Pre-calculate points in cervix and vagina
+        rotMat = axis_angle_to_rotation_matrix([0.0, 1.0, 0.0], anteversionAngleRad)
+        vaginalStartX = bodyLength + cervicalLength
+
+        turningPtIdx = 1
+        xTurningPt = [vaginalStartX + vaginalScale * turningPtIdx, 0.0, 0.0]
+        xTranslateMat = sub(xTurningPt, xBodyJunction)
+
+        li = list(range(2, vaginaElementsCount + 1))
+
+        # one node before vagina
+        nx = []
+        nd1 = []
+        nx.append([vaginalStartX - cervicalScale, 0.0, 0.0])
+        nd1.append([cervicalScale, 0.0, 0.0])
+        for i in li:
+            x = [vaginalStartX + vaginalScale * i, 0.0, 0.0]
+            d1 = [0.5 * (vaginalScale + cervicalScale) if i == 0 else vaginalScale, 0.0, 0.0]
+            if i > turningPtIdx:
+                xTranslate = sub(x, xTranslateMat)
+                xRot = [rotMat[j][0] * xTranslate[0] + rotMat[j][1] * xTranslate[1] + rotMat[j][2] * xTranslate[2]
+                        for j in range(3)]
+                x = add(xRot, xTranslateMat)
+                d1 = [rotMat[j][0] * d1[0] + rotMat[j][1] * d1[1] + rotMat[j][2] * d1[2] for j in range(3)]
+            nx.append(x)
+            nd1.append(d1)
+
+        xSampledCurve, d1SampledCurve = sampleCubicHermiteCurves(nx, nd1, vaginaElementsCount + 1,
+                                                                 arcLengthDerivatives=True)[0:2]
+        d1SmoothedCurve = smoothCubicHermiteDerivativesLine(xSampledCurve, d1SampledCurve, fixStartDirection=True)
+
+        # assign points to cervix
+        nodeIdentifier -= 1
+        cervixStartX = bodyLength
+        nxCervix = [[cervixStartX, 0.0, 0.0]] + xSampledCurve[0:2]
+        nd1Cervix = [[0.5 * (fundusScalePostBodyJunction + cervicalScale), 0.0, 0.0]] + d1SmoothedCurve[0:2]
+        nxCervixSampled, nd1CervixSampled = \
+            sampleCubicHermiteCurves(nxCervix, nd1Cervix, cervixElementsCount, arcLengthDerivatives=True)[0:2]
+        nd1CervixSampled[0] = [0.5 * (fundusScalePostBodyJunction + cervicalScale), 0.0, 0.0]
+        nd1CervixSmoothed = smoothCubicHermiteDerivativesLine(nxCervixSampled, nd1CervixSampled,
+                                                              fixStartDerivative=True,
+                                                              fixEndDerivative=True)
+        nd2Cervix = []
+        nd3Cervix = []
+        for i in range(cervixElementsCount + 1):
+            xi = i / cervixElementsCount
+            d1 = nd1CervixSmoothed[i]
+            halfCervixWidth = xi * halfCervicalWidthExternalOs + (1 - xi) * halfCervicalWidthInternalOs
+            halfCervixDepth = xi * halfCervicalDepthExternalOs + (1 - xi) * halfCervicalDepthInternalOs
+            d2 = [0.0, halfCervixWidth, 0.0]
+            nd2Cervix.append(d2)
+            d3Direction = cross(normalize(d1), normalize(d2))
+            nd3Cervix.append(set_magnitude(d3Direction, halfCervixDepth))
+
+        nxVagina = []
+        nd1Vagina = []
+        nd2Vagina = []
+        nd3Vagina = []
+
+        del xSampledCurve[0]
+        del d1SmoothedCurve[0]
+
+        for i in range(vaginaElementsCount + 1):
+            xi = i / vaginaElementsCount
+            nxVagina.append(xSampledCurve[i])
+            nd1Vagina.append(d1SmoothedCurve[i])
+            d2 = \
+                interpolateCubicHermite(
+                    [0.0, halfCervicalWidthExternalOs, 0.0],
+                    [0.0, (halfCervicalWidthExternalOs - halfCervicalWidthInternalOs) / cervixElementsCount, 0.0],
+                    [0.0, halfVaginaOrificeWidth, 0.0],
+                    [0.0, (halfVaginaOrificeWidth - halfCervicalWidthExternalOs) / vaginaElementsCount, 0.0], xi)
+
+            if i > turningPtIdx:
+                d2 = [rotMat[j][0] * d2[0] + rotMat[j][1] * d2[1] + rotMat[j][2] * d2[2] for j in range(3)]
+            nd2Vagina.append(d2)
+
+            d3 = cross(normalize(d1SmoothedCurve[i]), normalize(d2))
+            d3Interpolated = \
+                interpolateCubicHermite(
+                    [0.0, 0.0, halfCervicalDepthExternalOs],
+                    [0.0, 0.0, (halfCervicalDepthExternalOs - halfCervicalDepthInternalOs) / cervixElementsCount],
+                    [0.0, 0.0, halfVaginaOrificeDepth],
+                    [0.0, 0.0, (halfVaginaOrificeDepth - halfCervicalDepthExternalOs) / vaginaElementsCount], xi)
+            d3 = set_magnitude(d3, magnitude(d3Interpolated))
+            nd3Vagina.append(d3)
+
+        nd2 = nd2Cervix + nd2Vagina[1:]
+        nd3 = nd3Cervix + nd3Vagina[1:]
+        nd12 = [sub(nd2[c + 1], nd2[c]) for c in range(len(nd2) - 1)]
+        nd12.append(nd12[-1])
+        nd13 = [sub(nd3[c + 1], nd3[c]) for c in range(len(nd3) - 1)]
+        nd13.append(nd13[-1])
+
+        if halfCervicalWidthExternalOs - halfCervicalWidthInternalOs:
+            nd12 = smoothCubicHermiteDerivativesLine(nd2, nd12)
+        if halfCervicalDepthExternalOs - halfCervicalDepthInternalOs:
+            nd13 = smoothCubicHermiteDerivativesLine(nd3, nd13)
+
+        nd12Cervix = nd12[0:cervixElementsCount + 1]
+        nd12Vagina = nd12[cervixElementsCount:]
+        nd13Cervix = nd13[0:cervixElementsCount + 1]
+        nd13Vagina = nd13[cervixElementsCount:]
+
+        for i in range(cervixElementsCount):
+            node = nodes.findNodeByIdentifier(nodeIdentifier)
+            fieldcache.setNode(node)
+            x = nxCervixSampled[i]
+            d1 = nd1CervixSmoothed[i]
+            d2 = nd2Cervix[i]
+            d3 = nd3Cervix[i]
+            if i == 0:
+                d12 = mult(add(nd12Body[-1], nd12Cervix[i]), 0.5)
+                d13 = mult(add(nd13Body[-1], nd13Cervix[i]), 0.5)
+            else:
+                d12 = nd12Cervix[i]
+                d13 = nd13Cervix[i]
+
+            id2 = mult(d2, innerProportionCervix)
+            id3 = mult(d3, innerProportionCervix)
+            id12 = mult(d12, innerProportionCervix)
+            id13 = mult(d13, innerProportionCervix)
+            setNodeFieldParameters(coordinates, fieldcache, x, d1, d2, d3, d12, d13)
+            setNodeFieldParameters(innerCoordinates, fieldcache, x, d1, id2, id3, id12, id13)
+            nodeIdentifier += 1
+
+        for i in range(vaginaElementsCount + 1):
+            node = nodes.findNodeByIdentifier(nodeIdentifier)
+            fieldcache.setNode(node)
+            x = nxVagina[i]
+            d1 = nd1Vagina[i]
+            d2 = nd2Vagina[i]
+            d3 = nd3Vagina[i]
+            d12 = nd12Vagina[i]
+            d13 = nd13Vagina[i]
+            if i == 0:
+                innerProportion = innerProportionCervix
+            else:
+                innerProportion = innerProportionVagina
+            id2 = mult(d2, innerProportion)
+            id3 = mult(d3, innerProportion)
+            id12 = mult(d12, innerProportion)
+            id13 = mult(d13, innerProportion)
+            setNodeFieldParameters(coordinates, fieldcache, x, d1, d2, d3, d12, d13)
+            setNodeFieldParameters(innerCoordinates, fieldcache, x, d1, id2, id3, id12, id13)
+            nodeIdentifier += 1
+
+        return annotationGroups, networkMesh
+
+
+    @classmethod
+    def getInteractiveFunctions(cls):
+        """
+        Edit base class list to include only valid functions.
+        """
+        interactiveFunctions = super(MeshType_1d_uterus_network_layout1, cls).getInteractiveFunctions()
+        for interactiveFunction in interactiveFunctions:
+            if interactiveFunction[0] == "Edit structure...":
+                interactiveFunctions.remove(interactiveFunction)
+                break
+        return interactiveFunctions
 
 
 class MeshType_3d_uterus1(Scaffold_base):
@@ -79,244 +869,66 @@ class MeshType_3d_uterus1(Scaffold_base):
     wall.
     Magnitude of D2 and D3 are the radii of the uterus in the respective directions.
     """
-    parameterSetStructureStrings = {
-        'Mouse 1': ScaffoldPackage(MeshType_1d_network_layout1, {
-            'scaffoldSettings': {
-                "Structure": "1-2-3-4-5, 6-7-8-9-5.2, 5.3-10-11, 11-12-13, 13-14"
-            },
-            'meshEdits': exnode_string_from_nodeset_field_parameters(
-                ['coordinates'],
-                [Node.VALUE_LABEL_VALUE, Node.VALUE_LABEL_D_DS1, Node.VALUE_LABEL_D_DS2, Node.VALUE_LABEL_D2_DS1DS2,
-                 Node.VALUE_LABEL_D_DS3, Node.VALUE_LABEL_D2_DS1DS3], [[
-                    (1, [[24.20, 0.00, 36.56], [-2.53, 0.00, -14.75], [2.42, 0.00, -0.41], [0.13, 0.00, -0.34], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (2, [[20.88, 0.00, 23.34], [-4.10, 0.00, -11.62], [2.47, 0.00, -0.87], [-0.03, 0.00, -0.57], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (3, [[16.28, 0.00, 13.37], [-5.76, 0.00, -8.94], [2.38, 0.00, -1.54], [-0.41, 0.00, -0.56], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (4, [[9.59, 0.00, 5.63], [-8.24, 0.00, -6.86], [1.67, 0.00, -2.00], [-0.75, 0.00, -0.33], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (5, [[0.00, 0.00, 0.00], [[-10.77, 0.00, -4.33], [10.77, 0.00, -4.33], [0.00, 0.00, -4.41]],
-                         [[0.88, 0.00, -2.18], [0.83, 0.11, 2.07], [2.67, 0.00, 0.00]], [-0.32, 0.00, 0.38],
-                         [[0.00, -2.50, 0.00], [0.05, -2.50, 0.11], [0.00, -2.50, 0.00]], [0.00, 0.00, 0.00]]),
-                    (6, [[-24.20, 0.00, 36.56], [2.53, 0.00, -14.75], [2.42, 0.00, 0.41], [0.40, 0.00, 0.96], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (7, [[-20.88, 0.00, 23.34], [4.10, 0.00, -11.62], [2.47, 0.00, 0.87], [-0.03, 0.00, 0.57], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (8, [[-16.28, 0.00, 13.37], [5.76, 0.00, -8.94], [2.38, 0.00, 1.54], [-0.41, 0.00, 0.56], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (9, [[-9.59, 0.00, 5.63], [8.24, 0.00, -6.86], [1.67, 0.00, 2.00], [-0.05, 0.00, -0.59], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (10, [[0.00, 0.00, -4.05], [0.00, 0.00, -3.69], [2.50, 0.00, 0.00], [0.16, 0.00, -0.39], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (11, [[0.00, 0.00, -7.37], [0.00, 0.00, -3.46], [2.50, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (12, [[0.00, 0.00, -10.98], [0.00, 0.00, -3.31], [2.50, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (13, [[0.00, 0.00, -14.00], [0.00, 0.00, -3.51], [2.50, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]]),
-                    (14, [[0.00, 0.00, -18.00], [0.00, 0.00, -4.49], [2.50, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -2.50, 0.00], [0.00, 0.00, 0.00]])]]),
-            'userAnnotationGroups': [
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '1-4',
-                    'name': get_uterus_term('right uterine horn')[0],
-                    'ontId': get_uterus_term('right uterine horn')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '5-8',
-                    'name': get_uterus_term('left uterine horn')[0],
-                    'ontId': get_uterus_term('left uterine horn')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '9-10',
-                    'name': get_uterus_term('body of uterus')[0],
-                    'ontId': get_uterus_term('body of uterus')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '11-12',
-                    'name': get_uterus_term('uterine cervix')[0],
-                    'ontId': get_uterus_term('uterine cervix')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '13',
-                    'name': get_uterus_term('vagina')[0],
-                    'ontId': get_uterus_term('vagina')[1]
-                }]
-        }),
-        'Rat 1': ScaffoldPackage(MeshType_1d_network_layout1, {
-            'scaffoldSettings': {
-                "Structure": "1-2-3-4-5, 6-7-8-9-5.2, 5.3-10, 10-11, 11-12-13-14-15-16"
-            },
-            'meshEdits': exnode_string_from_nodeset_field_parameters(
-                ['coordinates'],
-                [Node.VALUE_LABEL_VALUE, Node.VALUE_LABEL_D_DS1, Node.VALUE_LABEL_D_DS2, Node.VALUE_LABEL_D2_DS1DS2,
-                 Node.VALUE_LABEL_D_DS3, Node.VALUE_LABEL_D2_DS1DS3], [[
-                    (1, [[25.00, 0.00, 43.53], [0.06, 0.00, -11.85], [2.27, 0.09, 0.01], [0.15, -0.06, -0.57], [0.09, -2.33, 0.00], [-0.08, 0.74, -0.01]]),
-                    (2, [[23.55, 0.00, 31.29], [-2.96, 0.00, -12.51], [2.27, 0.04, -0.54], [-0.15, -0.04, -0.53], [0.03, -1.91, -0.01], [-0.04, 0.10, 0.00]]),
-                    (3, [[19.03, 0.00, 18.76], [-6.24, 0.00, -11.75], [1.96, 0.01, -1.04], [-0.29, 0.01, -0.58], [0.01, -2.16, 0.00], [0.00, -0.20, -0.01]]),
-                    (4, [[11.24, 0.00, 8.09], [-9.64, 0.00, -9.57], [1.69, 0.05, -1.70], [-0.44, -0.01, -0.52], [0.03, -2.32, -0.03], [-0.01, -0.14, 0.01]]),
-                    (5, [[0.00, 0.00, 0.00], [[-12.65, 0.00, -6.51], [12.65, 0.00, -6.51], [0.00, 0.00, -4.00]],
-                         [[1.07, -0.02, -2.07], [1.07, 0.02, 2.07], [4.65, 0.00, 0.00]], [-0.23, -0.03, 0.15],
-                         [[-0.01, -2.44, 0.02], [0.01, -2.44, 0.02], [0.00, -2.53, 0.00]], [-0.01, -0.07, 0.04]]),
-                    (6, [[-25.00, 0.00, 43.53], [-0.06, 0.00, -11.85], [2.27, 0.09, -0.01], [0.24, -0.02, 0.84], [0.09, -2.33, 0.00], [-0.03, 0.36, 0.00]]),
-                    (7, [[-23.55, 0.00, 31.29], [2.96, 0.00, -12.51], [2.27, 0.04, 0.54], [-0.15, -0.04, 0.53], [0.03, -1.91, 0.01], [-0.04, 0.10, 0.00]]),
-                    (8, [[-19.03, 0.00, 18.76], [6.24, 0.00, -11.75], [1.96, 0.01, 1.04], [-0.29, 0.01, 0.58], [0.01, -2.16, 0.00], [0.00, -0.20, 0.01]]),
-                    (9, [[-11.24, 0.00, 8.09], [9.64, 0.00, -9.57], [1.69, 0.05, 1.70], [1.17, 0.00, -0.39], [0.03, -2.32, 0.03], [0.00, -0.21, 0.00]]),
-                    (10, [[0.00, 0.00, -4.00], [0.00, 0.00, -4.00], [4.65, 0.00, 0.00], [0.58, -0.01, -0.33], [0.00, -2.60, 0.00], [-0.01, -0.05, -0.01]]),
-                    (11, [[0.00, 0.00, -8.00], [0.00, 0.00, -4.00], [4.65, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -2.60, 0.00], [0.00, 0.00, 0.00]]),
-                    (12, [[0.00, 0.00, -12.00], [0.00, 0.00, -4.00], [4.65, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -2.60, 0.00], [0.00, 0.00, 0.00]]),
-                    (13, [[0.00, 0.00, -16.00], [0.00, 0.00, -4.00], [4.65, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -2.60, 0.00], [0.00, 0.00, 0.00]]),
-                    (14, [[0.00, 0.00, -20.00], [0.00, 0.00, -4.00], [4.65, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -2.60, 0.00], [0.00, 0.00, 0.00]]),
-                    (15, [[0.00, 0.00, -24.00], [0.00, 0.00, -4.00], [4.65, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -2.60, 0.00], [0.00, 0.00, 0.00]]),
-                    (16, [[0.00, 0.00, -28.00], [0.00, 0.00, -4.00], [4.65, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -2.60, 0.00], [0.00, 0.00, 0.00]])]]),
-            'userAnnotationGroups': [
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '1-4',
-                    'name': get_uterus_term('right uterine horn')[0],
-                    'ontId': get_uterus_term('right uterine horn')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '5-8',
-                    'name': get_uterus_term('left uterine horn')[0],
-                    'ontId': get_uterus_term('left uterine horn')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '9',
-                    'name': get_uterus_term('body of uterus')[0],
-                    'ontId': get_uterus_term('body of uterus')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '10',
-                    'name': get_uterus_term('uterine cervix')[0],
-                    'ontId': get_uterus_term('uterine cervix')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '11-15',
-                    'name': get_uterus_term('vagina')[0],
-                    'ontId': get_uterus_term('vagina')[1]
-                }]
-        }),
-        'Material': ScaffoldPackage(MeshType_1d_network_layout1, {
-            'scaffoldSettings': {
-                "Structure": "1-2-3, 4-5-3.2, 3.3-6, 6-7, 7-8-9"
-            },
-            'meshEdits': exnode_string_from_nodeset_field_parameters(
-                ['coordinates'],
-                [Node.VALUE_LABEL_VALUE, Node.VALUE_LABEL_D_DS1, Node.VALUE_LABEL_D_DS2, Node.VALUE_LABEL_D2_DS1DS2, Node.VALUE_LABEL_D_DS3, Node.VALUE_LABEL_D2_DS1DS3], [[
-                    (1, [[2.00, 0.00, 0.00], [-1.00, 0.00, 0.00], [0.00, 0.00, -0.30], [0.00, 0.00, 0.00], [0.00, -0.30, 0.00], [0.00, 0.00, 0.00]]),
-                    (2, [[1.00, 0.00, 0.00], [-1.00, 0.00, 0.00], [0.00, 0.00, -0.30], [0.00, 0.00, 0.00], [0.00, -0.30, 0.00], [0.00, 0.00, 0.00]]),
-                    (3, [[0.00, 0.00, 0.00], [[-1.00, 0.00, 0.00], [1.00, 0.00, 0.00], [0.00, 0.00, -1.00]],
-                         [[0.00, 0.00, -0.30], [0.00, 0.00, 0.30], [0.30, 0.00, 0.00]], [0.00, 0.00, 0.20],
-                         [[0.00, -0.30, 0.00], [0.00, -0.30, 0.00], [0.00, -0.30, 0.00]], [0.00, 0.00, 0.00]]),
-                    (4, [[-2.00, 0.00, 0.00], [1.00, 0.00, 0.00], [0.00, 0.00, 0.30], [0.00, 0.00, 0.20], [0.00, -0.30, 0.00], [0.00, 0.00, 0.00]]),
-                    (5, [[-1.00, 0.00, 0.00], [1.00, 0.00, 0.00], [0.00, 0.00, 0.30], [0.12, 0.00, -0.12], [0.00, -0.30, 0.00], [0.00, 0.00, 0.00]]),
-                    (6, [[0.00, 0.00, -1.00], [0.00, 0.00, -1.00], [0.30, 0.00, 0.00], [0.12, 0.00, -0.12], [0.00, -0.30, 0.00], [0.00, 0.00, 0.00]]),
-                    (7, [[0.00, 0.00, -2.00], [0.00, 0.00, -1.00], [0.30, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -0.30, 0.00], [0.00, 0.00, 0.00]]),
-                    (8, [[0.00, 0.00, -3.00], [0.00, 0.00, -1.00], [0.30, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -0.30, 0.00], [0.00, 0.00, 0.00]]),
-                    (9, [[0.00, 0.00, -4.00], [0.00, 0.00, -1.00], [0.30, 0.00, 0.00], [0.00, 0.00, 0.00], [0.00, -0.30, 0.00], [0.00, 0.00, 0.00]])]]),
-            'userAnnotationGroups': [
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '1-2',
-                    'name': get_uterus_term('right uterine horn')[0],
-                    'ontId': get_uterus_term('right uterine horn')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '3-4',
-                    'name': get_uterus_term('left uterine horn')[0],
-                    'ontId': get_uterus_term('left uterine horn')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '5',
-                    'name': get_uterus_term('body of uterus')[0],
-                    'ontId': get_uterus_term('body of uterus')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '6',
-                    'name': get_uterus_term('uterine cervix')[0],
-                    'ontId': get_uterus_term('uterine cervix')[1]
-                },
-                {
-                    '_AnnotationGroup': True,
-                    'dimension': 1,
-                    'identifierRanges': '7-8',
-                    'name': get_uterus_term('vagina')[0],
-                    'ontId': get_uterus_term('vagina')[1]
-                }]
-        })
-    }
 
-    @staticmethod
-    def getName():
+    @classmethod
+    def getName(cls):
         return '3D Uterus 1'
 
-    @staticmethod
-    def getParameterSetNames():
+    @classmethod
+    def getParameterSetNames(cls):
         return [
             'Default',
+            'Human 1',
+            'Human Pregnant 1',
             'Mouse 1',
-            'Rat 1',
-            'Material']
+            'Rat 1']
 
     @classmethod
     def getDefaultOptions(cls, parameterSetName='Default'):
-        if 'Mouse 1' in parameterSetName:
-            networkLayoutOption = cls.parameterSetStructureStrings['Mouse 1']
-        elif 'Rat 1' in parameterSetName:
-            networkLayoutOption = cls.parameterSetStructureStrings['Rat 1']
-        elif 'Material' in parameterSetName:
-            networkLayoutOption = cls.parameterSetStructureStrings['Material']
-        else:
-            networkLayoutOption = cls.parameterSetStructureStrings['Mouse 1']
+
+        useParameterSetName = "Human 1" if (parameterSetName == "Default") else parameterSetName
         options = {
-            'Network layout': copy.deepcopy(networkLayoutOption),
-            'Target element length': 6.0,
-            'Number of elements around': 8,
-            'Number of elements around horns': 8,
+            'Base parameter set': useParameterSetName,
+            'Network layout': ScaffoldPackage(MeshType_1d_uterus_network_layout1,
+                                              defaultParameterSetName=useParameterSetName),
+            'Number of elements around': 24,
+            'Number of elements around oviduct': 8,
             'Number of elements through wall': 1,
-            'Wall thickness': 1.0,
-            'Double uterus': False,
+            'Number of elements along oviduct': 6,
+            'Number of elements along body': 4,
+            'Number of elements along cervix': 2,
+            'Number of elements along vagina': 6,
+            'Target element density along longest segment': 5.5,
             'Use linear through wall': True,
-            'Use cross derivatives': False,
+            'Show trim surfaces': False,
             'Refine': False,
             'Refine number of elements along': 4,
             'Refine number of elements around': 4,
             'Refine number of elements through wall': 1
         }
-        if 'Rat' in parameterSetName:
-            options['Target element length'] = 5.0
-            options['Wall thickness'] = 1.2
-            options['Double uterus'] = True
-        if 'Material' in parameterSetName:
-            options['Target element length'] = 1.0
-            options['Wall thickness'] = 0.1
+        if 'Mouse' in parameterSetName or 'Rat' in parameterSetName:
+            options['Number of elements around'] = 12
+            options['Number of elements around oviduct'] = 8
+            options['Number of elements along oviduct'] = 12
+            options['Number of elements along body'] = 2
+            options['Number of elements along cervix'] = 1
+            options['Number of elements along vagina'] = 2
+
         return options
 
-    @staticmethod
-    def getOrderedOptionNames():
+    @classmethod
+    def getOrderedOptionNames(cls):
         optionNames = [
             'Network layout',
-            'Target element length',
             'Number of elements around',
-            'Number of elements around horns',
+            'Number of elements around oviduct',
             'Number of elements through wall',
-            'Wall thickness',
-            'Double uterus',
+            'Number of elements along oviduct',
+            'Number of elements along body',
+            'Number of elements along cervix',
+            'Number of elements along vagina',
+            'Use linear through wall',
+            'Show trim surfaces',
             'Refine',
             'Refine number of elements along',
             'Refine number of elements around',
@@ -326,17 +938,8 @@ class MeshType_3d_uterus1(Scaffold_base):
     @classmethod
     def getOptionValidScaffoldTypes(cls, optionName):
         if optionName == 'Network layout':
-            return [MeshType_1d_network_layout1]
+            return [MeshType_1d_uterus_network_layout1]
         return []
-
-    @classmethod
-    def getOptionScaffoldTypeParameterSetNames(cls, optionName, scaffoldType):
-        if optionName == 'Network layout':
-            return list(cls.parameterSetStructureStrings.keys())
-        assert scaffoldType in cls.getOptionValidScaffoldTypes(optionName), \
-            cls.__name__ + '.getOptionScaffoldTypeParameterSetNames.  ' + \
-            'Invalid option \'' + optionName + '\' scaffold type ' + scaffoldType.getName()
-        return scaffoldType.getParameterSetNames()
 
     @classmethod
     def getOptionScaffoldPackage(cls, optionName, scaffoldType, parameterSetName=None):
@@ -350,16 +953,16 @@ class MeshType_3d_uterus1(Scaffold_base):
                 ' in option ' + str(optionName) + ' of scaffold ' + cls.getName()
         if optionName == 'Network layout':
             if not parameterSetName:
-                parameterSetName = list(cls.parameterSetStructureStrings.keys())[1]
-            return copy.deepcopy(cls.parameterSetStructureStrings[parameterSetName])
+                parameterSetName = "Default"
+            return ScaffoldPackage(MeshType_1d_uterus_network_layout1, defaultParameterSetName=parameterSetName)
         assert False, cls.__name__ + '.getOptionScaffoldPackage:  Option ' + optionName + ' is not a scaffold'
 
     @classmethod
     def checkOptions(cls, options):
-        if not options['Network layout'].getScaffoldType() in cls.getOptionValidScaffoldTypes('Network layout'):
-            options['Network layout'] = cls.getOptionScaffoldPackage('Network layout', MeshType_1d_network_layout1)
+        if (options["Network layout"].getScaffoldType() not in
+                cls.getOptionValidScaffoldTypes("Network layout")):
+            options["Network layout"] = ScaffoldPackage(MeshType_1d_uterus_network_layout1)
         for key in [
-            'Number of elements around',
             'Number of elements through wall',
             'Refine number of elements along',
             'Refine number of elements around',
@@ -367,87 +970,115 @@ class MeshType_3d_uterus1(Scaffold_base):
         ]:
             if options[key] < 1:
                 options[key] = 1
-        if options['Number of elements around'] % 2 != 0:
-            options['Number of elements around'] += 1
-        if options['Number of elements around'] < options['Number of elements around horns']:
-            options['Number of elements around'] = options['Number of elements around horns']
+
+        for key in [
+            'Number of elements around',
+            'Number of elements around oviduct']:
+            if options[key] < 4:
+                options[key] = 4
+            elif (options[key] % 4) > 0:
+                options[key] += options[key] % 4
+        if options["Number of elements through wall"] < 1:
+            options["Number of elements through wall"] = 1
+
+        dependentChanges = False
+        return dependentChanges
 
     @classmethod
     def generateBaseMesh(cls, region, options):
         """
-        Generate the base tricubic Hermite mesh. See also generateMesh().
+        Generate the base tricubic hermite or bicubic hermite-linear mesh. See also generateMesh().
         :param region: Zinc region to define model in. Must be empty.
         :param options: Dict containing options. See getDefaultOptions().
         :return: None
         """
-        networkLayout = options['Network layout']
-        elementsCountAround = options['Number of elements around']
-        elementsCountAroundHorn = options['Number of elements around horns']
-        elementsCountThroughWall = options['Number of elements through wall']
-        wallThickness = options['Wall thickness']
-        doubleUterus = options['Double uterus']
-        targetElementLength = options['Target element length']
-        useCrossDerivatives = options['Use cross derivatives']
+        parameterSetName = options['Base parameter set']
+        isHuman = parameterSetName in ("Default", "Human 1")
+        isPregnant = parameterSetName in ("Human Pregnant 1")
 
-        elementsCountAroundRightHorn = elementsCountAroundHorn
-        elementsCountAroundLeftHorn = elementsCountAroundHorn
+        layoutRegion = region.createRegion()
+        networkLayout = options["Network layout"]
+        networkLayout.generate(layoutRegion)  # ask scaffold to generate to get user-edited parameters
+        layoutAnnotationGroups = networkLayout.getAnnotationGroups()
+        networkMesh = networkLayout.getConstructionObject()
 
-        materialNetworkLayout = cls.parameterSetStructureStrings['Material']
-        materialWallThickness = 0.1
-        materialTargetElementLength = 0.5
+        annotationElementsCountsAlong = []
+        annotationElementsCountsAround = []
+        for layoutAnnotationGroup in layoutAnnotationGroups:
+            elementsCountAlong = 0
+            elementsCountAround = 0
+            name = layoutAnnotationGroup.getName()
+            if "oviduct" in name or "uterine horn" in name:
+                elementsCountAlong = options['Number of elements along oviduct']
+            elif "body" in name or "fundus" in name:
+                elementsCountAlong = options['Number of elements along body']
+            elif "cervix" in name:
+                elementsCountAlong = options['Number of elements along cervix']
+            elif "vagina" in name:
+                elementsCountAlong = options['Number of elements along vagina']
+            annotationElementsCountsAlong.append(elementsCountAlong)
 
-        # Geometric coordinates
-        fm = region.getFieldmodule()
-        coordinates = findOrCreateFieldCoordinates(fm)
+            if "oviduct" in name or "uterine horn" in name or "left fundus" in name or "right fundus" in name:
+                elementsCountAround = options['Number of elements around oviduct']
+            elif "body" in name or "fundus" in name:
+                elementsCountAround = options['Number of elements around']
+            annotationElementsCountsAround.append(elementsCountAround)
 
-        geometricNetworkLayout = UterusNetworkLayout(region, networkLayout, targetElementLength)
-        rightHornLength = geometricNetworkLayout.arcLengthOfGroupsAlong[0]
-        leftHornLength = geometricNetworkLayout.arcLengthOfGroupsAlong[1]
-        bodyLength = geometricNetworkLayout.arcLengthOfGroupsAlong[2]
-        cervixLength = geometricNetworkLayout.arcLengthOfGroupsAlong[3]
-        vaginaLength = geometricNetworkLayout.arcLengthOfGroupsAlong[4]
+        uterusTubeNetworkMeshBuilder = UterusTubeNetworkMeshBuilder(
+            networkMesh,
+            targetElementDensityAlongLongestSegment=2.0, # not used
+            layoutAnnotationGroups=layoutAnnotationGroups,
+            annotationElementsCountsAlong=annotationElementsCountsAlong,
+            defaultElementsCountAround=options['Number of elements around'],
+            annotationElementsCountsAround=annotationElementsCountsAround,
+            elementsCountThroughShell=options["Number of elements through wall"],
+            useOuterTrimSurfaces=False)
+        uterusTubeNetworkMeshBuilder.build()
 
-        elementsCountInRightHorn = math.ceil(rightHornLength / targetElementLength)
-        elementsCountInLeftHorn = math.ceil(leftHornLength / targetElementLength)
-        elementsCountInBody = math.ceil(bodyLength / targetElementLength)
-        elementsCountInCervix = math.ceil(cervixLength / targetElementLength)
-        elementsCountInVagina = math.ceil(vaginaLength / targetElementLength)
+        generateData = UterusTubeNetworkMeshGenerateData(
+            region, 3,
+            isLinearThroughShell=options["Use linear through wall"],
+            isShowTrimSurfaces=options["Show trim surfaces"])
 
-        nodeIdentifier, elementIdentifier, annotationGroups = \
-            createUterusMesh3D(region, fm, coordinates, geometricNetworkLayout, elementsCountAround,
-                               elementsCountAroundRightHorn, elementsCountAroundLeftHorn, elementsCountThroughWall,
-                               elementsCountInRightHorn, elementsCountInLeftHorn, elementsCountInBody,
-                               elementsCountInCervix, elementsCountInVagina, wallThickness, doubleUterus,
-                               useCrossDerivatives)
+        uterusTubeNetworkMeshBuilder.generateMesh(generateData)
+        annotationGroups = generateData.getAnnotationGroups()
+        nodeIdentifier, elementIdentifier = generateData.getNodeElementIdentifiers()
 
-        # # Material coordinates
-        # tmp_region = region.createRegion()
-        # tmp_fm = tmp_region.getFieldmodule()
-        # with ChangeManager(tmp_fm):
-        #     tmp_uterus_coordinates = findOrCreateFieldCoordinates(tmp_fm, name="uterus coordinates")
-        #     materialNetworkLayout = UterusNetworkLayout(tmp_region, materialNetworkLayout, materialTargetElementLength)
-        #
-        #     nodeIdentifier, elementIdentifier, materialAnnotationGroups = \
-        #         createUterusMesh3D(tmp_region, tmp_fm, tmp_uterus_coordinates, materialNetworkLayout,
-        #                            elementsCountAround, elementsCountAroundRightHorn, elementsCountAroundLeftHorn,
-        #                            elementsCountThroughWall, elementsCountInRightHorn, elementsCountInLeftHorn,
-        #                            elementsCountInBody, elementsCountInCervix, elementsCountInVagina, wallThickness,
-        #                            doubleUterus, useCrossDerivatives)
-        #     # Write two coordinates
-        #     sir = tmp_region.createStreaminformationRegion()
-        #     srm = sir.createStreamresourceMemory()
-        #     tmp_region.write(sir)
-        #     result, buffer = srm.getBuffer()
-        #
-        #     sir = region.createStreaminformationRegion()
-        #     srm = sir.createStreamresourceMemoryBuffer(buffer)
-        #     region.read(sir)
-        #
-        #     del srm
-        #     del sir
-        #     del tmp_uterus_coordinates
-        # del tmp_fm
-        # del tmp_region
+        fieldmodule = region.getFieldmodule()
+        coordinates = findOrCreateFieldCoordinates(fieldmodule)
+        nodes = fieldmodule.findNodesetByFieldDomainType(Field.DOMAIN_TYPE_NODES)
+
+        mesh = fieldmodule.findMeshByDimension(3)
+
+        # add human specific annotations
+        allMarkers = {}
+        if isHuman:
+            allMarkers = {"junction of left round ligament with uterus": {"x": [-0.495368, -4.54665, 0.0]},
+                          "junction of right round ligament with uterus": {"x": [-0.495368, 4.54665, 0.0]},
+                          "junction of left uterosacral ligament with uterus": {"x": [7.0, -2.75, 0.0]},
+                          "junction of right uterosacral ligament with uterus": {"x": [7.0, 2.75, 0.0]}}
+        elif isPregnant:
+            allMarkers = {"junction of left round ligament with uterus": {"x": [-2.81269, -16.7248, 0.0]},
+                          "junction of right round ligament with uterus": {"x": [-2.81269, 16.7248, 0.0]},
+                          "junction of left uterosacral ligament with uterus": {"x": [30, -2.75, 0.0]},
+                          "junction of right uterosacral ligament with uterus": {"x": [30, 2.75, 0]}}
+
+        for key in allMarkers:
+            x = allMarkers[key]["x"]
+            group = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                        get_uterus_term(key), isMarker=True)
+            markerNode = group.createMarkerNode(nodeIdentifier, coordinates, x)
+            nodeIdentifier = markerNode.getIdentifier() + 1
+            for group in annotationGroups:
+                group.getNodesetGroup(nodes).addNode(markerNode)
+
+        uterusGroup = findOrCreateAnnotationGroupForTerm(annotationGroups, region, get_uterus_term("uterus"))
+        myometriumGroup = findOrCreateAnnotationGroupForTerm(
+            annotationGroups, region, get_uterus_term("myometrium"))
+        myometriumGroup.getMeshGroup(mesh).addElementsConditional(uterusGroup.getGroup())
+
+        upperCervixGroup = getAnnotationGroupForTerm(annotationGroups, ("upper cervix", ""))
+        uterusGroup.getMeshGroup(mesh).addElementsConditional(upperCervixGroup.getGroup())
 
         return annotationGroups, None
 
@@ -476,2471 +1107,278 @@ class MeshType_3d_uterus1(Scaffold_base):
         :param annotationGroups: List of annotation groups for top-level elements.
         New face annotation groups are appended to this list.
         """
+        parameterSetName = options['Base parameter set']
+        isHuman = parameterSetName in ("Default", "Human 1", "Human Pregnant 1")
+        isRodent = parameterSetName in ("Mouse 1", "Rat 1")
+
         # Create 2d surface mesh groups
         fm = region.getFieldmodule()
-        cervixGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("uterine cervix"))
-        rightHornGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("right uterine horn"))
-        leftHornGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("left uterine horn"))
         uterusGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("uterus"))
+        if isHuman:
+            leftOviductGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("left oviduct"))
+            rightOviductGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("right oviduct"))
+        elif isRodent:
+            leftOviductGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("left uterine horn"))
+            rightOviductGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("right uterine horn"))
+        bodyGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("body of uterus"))
+        fundusGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("fundus of uterus"))
+        cervixGroup = getAnnotationGroupForTerm(annotationGroups, ("cervix", ""))
+        upperCervixGroup = getAnnotationGroupForTerm(annotationGroups, ("upper cervix", ""))
+        lowerCervixGroup = getAnnotationGroupForTerm(annotationGroups, ("lower cervix", ""))
+        vaginaGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("vagina"))
 
+        mesh1d = fm.findMeshByDimension(1)
         mesh2d = fm.findMeshByDimension(2)
 
         is_exterior = fm.createFieldIsExterior()
         is_exterior_face_xi3_1 = fm.createFieldAnd(is_exterior, fm.createFieldIsOnFace(Element.FACE_TYPE_XI3_1))
         is_exterior_face_xi3_0 = fm.createFieldAnd(is_exterior, fm.createFieldIsOnFace(Element.FACE_TYPE_XI3_0))
-
-        is_cervix = cervixGroup.getGroup()
-        is_cervix_outer = fm.createFieldAnd(is_cervix, is_exterior_face_xi3_1)
-        is_cervix_inner = fm.createFieldAnd(is_cervix, is_exterior_face_xi3_0)
-
-        is_rightHorn = rightHornGroup.getGroup()
-        is_rightHorn_outer = fm.createFieldAnd(is_rightHorn, is_exterior_face_xi3_1)
-        is_rightHorn_inner = fm.createFieldAnd(is_rightHorn, is_exterior_face_xi3_0)
-
-        is_leftHorn = leftHornGroup.getGroup()
-        is_leftHorn_outer = fm.createFieldAnd(is_leftHorn, is_exterior_face_xi3_1)
-        is_leftHorn_inner = fm.createFieldAnd(is_leftHorn, is_exterior_face_xi3_0)
+        is_exterior_face_xi2_0 = fm.createFieldAnd(is_exterior, fm.createFieldIsOnFace(Element.FACE_TYPE_XI2_0))
+        is_exterior_face_xi2_1 = fm.createFieldAnd(is_exterior, fm.createFieldIsOnFace(Element.FACE_TYPE_XI2_1))
 
         is_uterus = uterusGroup.getGroup()
         is_uterus_outer = fm.createFieldAnd(is_uterus, is_exterior_face_xi3_1)
         is_uterus_inner = fm.createFieldAnd(is_uterus, is_exterior_face_xi3_0)
 
-        serosaOfCervix = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
-                                                            get_uterus_term("serosa of uterine cervix"))
-        serosaOfCervix.getMeshGroup(mesh2d).addElementsConditional(is_cervix_outer)
+        is_leftOviduct = leftOviductGroup.getGroup()
+        is_rightOviduct = rightOviductGroup.getGroup()
 
-        lumenOfCervix = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
-                                                           get_uterus_term("lumen of uterine cervix"))
-        lumenOfCervix.getMeshGroup(mesh2d).addElementsConditional(is_cervix_inner)
+        is_fundus = fundusGroup.getGroup()
+        is_fundus_outer = fm.createFieldAnd(is_fundus, is_exterior_face_xi3_1)
+        is_fundus_inner = fm.createFieldAnd(is_fundus, is_exterior_face_xi3_0)
 
-        serosaOfRightHorn = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
-                                                               get_uterus_term("serosa of right horn"))
-        serosaOfRightHorn.getMeshGroup(mesh2d).addElementsConditional(is_rightHorn_outer)
+        is_body = bodyGroup.getGroup()
+        is_body_outer = fm.createFieldAnd(is_body, is_exterior_face_xi3_1)
+        is_body_inner = fm.createFieldAnd(is_body, is_exterior_face_xi3_0)
 
-        lumenOfRightHorn = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
-                                                              get_uterus_term("lumen of right horn"))
-        lumenOfRightHorn.getMeshGroup(mesh2d).addElementsConditional(is_rightHorn_inner)
+        is_cervix = cervixGroup.getGroup()
+        is_cervix_outer = fm.createFieldAnd(is_cervix, is_exterior_face_xi3_1)
+        is_cervix_inner = fm.createFieldAnd(is_cervix, is_exterior_face_xi3_0)
 
-        serosaOfLeftHorn = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
-                                                              get_uterus_term("serosa of left horn"))
-        serosaOfLeftHorn.getMeshGroup(mesh2d).addElementsConditional(is_leftHorn_outer)
+        is_upperCervix = upperCervixGroup.getGroup()
+        is_upperCervix_outer = fm.createFieldAnd(is_upperCervix, is_exterior_face_xi3_1)
 
-        lumenOfLeftHorn = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
-                                                             get_uterus_term("lumen of left horn"))
-        lumenOfLeftHorn.getMeshGroup(mesh2d).addElementsConditional(is_leftHorn_inner)
+        is_lowerCervix = lowerCervixGroup.getGroup()
+        is_lowerCervix_outer = fm.createFieldAnd(is_lowerCervix, is_exterior_face_xi3_1)
+
+        is_vagina = vaginaGroup.getGroup()
+        is_vagina_outer = fm.createFieldAnd(is_vagina, is_exterior_face_xi3_1)
+        is_vagina_inner = fm.createFieldAnd(is_vagina, is_exterior_face_xi3_0)
+        is_vagina_xi2_0 = fm.createFieldAnd(is_vagina, is_exterior_face_xi2_0)
+        is_vagina_xi2_1 = fm.createFieldAnd(is_vagina, is_exterior_face_xi2_1)
+        is_vagina_xi2_01 = fm.createFieldXor(is_vagina_xi2_0, is_vagina_xi2_1)
 
         serosaOfUterus = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
                                                             get_uterus_term("serosa of uterus"))
         serosaOfUterus.getMeshGroup(mesh2d).addElementsConditional(is_uterus_outer)
 
-        lumenOfUterus = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
-                                                           get_uterus_term("lumen of uterus"))
-        lumenOfUterus.getMeshGroup(mesh2d).addElementsConditional(is_uterus_inner)
+        uterineCavity = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                           get_uterus_term("uterine cavity"))
+        uterineCavity.getMeshGroup(mesh2d).addElementsConditional(is_uterus_inner)
+        uterineCavity.getMeshGroup(mesh2d).removeElementsConditional(is_leftOviduct)
+        uterineCavity.getMeshGroup(mesh2d).removeElementsConditional(is_rightOviduct)
+        uterineCavity.getMeshGroup(mesh2d).removeElementsConditional(is_upperCervix)
+
+        serosaOfBody = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                          get_uterus_term("serosa of body of uterus"))
+        serosaOfBody.getMeshGroup(mesh2d).addElementsConditional(is_body_outer)
+        serosaOfBody.getMeshGroup(mesh2d).addElementsConditional(is_upperCervix_outer)
+
+        lumenOfFundus = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                         get_uterus_term("lumen of fundus of uterus"))
+        lumenOfFundus.getMeshGroup(mesh2d).addElementsConditional(is_fundus_inner)
+
+        serosaOfFundus = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                          get_uterus_term("serosa of fundus of uterus"))
+        serosaOfFundus.getMeshGroup(mesh2d).addElementsConditional(is_fundus_outer)
+
+        lumenOfBody = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                         get_uterus_term("lumen of body of uterus"))
+        lumenOfBody.getMeshGroup(mesh2d).addElementsConditional(is_body_inner)
+
+        lumenOfCervix = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                           get_uterus_term("uterine cervix"))
+        lumenOfCervix.getMeshGroup(mesh2d).addElementsConditional(is_cervix_inner)
+
+        serosaOfVagina = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                            get_uterus_term("serosa of vagina"))
+        serosaOfVagina.getMeshGroup(mesh2d).addElementsConditional(is_vagina_outer)
+        serosaOfVagina.getMeshGroup(mesh2d).addElementsConditional(is_lowerCervix_outer)
+
+        lumenOfVagina = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                           get_uterus_term("vaginal canal"))
+        lumenOfVagina.getMeshGroup(mesh2d).addElementsConditional(is_vagina_inner)
+
+        leftGroup = getAnnotationGroupForTerm(annotationGroups, ("left uterus", ""))
+        rightGroup = getAnnotationGroupForTerm(annotationGroups, ("right uterus", ""))
+        dorsalGroup = getAnnotationGroupForTerm(annotationGroups, ("dorsal uterus", ""))
+        ventralGroup = getAnnotationGroupForTerm(annotationGroups, ("ventral uterus", ""))
+
+        if isHuman:
+            leftOviductGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("left oviduct"))
+            is_leftOviduct = leftOviductGroup.getGroup()
+            is_leftOviduct_outer = fm.createFieldAnd(is_leftOviduct, is_exterior_face_xi3_1)
+            is_leftOviduct_inner = fm.createFieldAnd(is_leftOviduct, is_exterior_face_xi3_0)
+
+            serosaOfLeftOviduct = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                                   get_uterus_term("serosa of left oviduct"))
+            serosaOfLeftOviduct.getMeshGroup(mesh2d).addElementsConditional(is_leftOviduct_outer)
+
+            rightOviductGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("right oviduct"))
+            is_rightOviduct = rightOviductGroup.getGroup()
+            is_rightOviduct_outer = fm.createFieldAnd(is_rightOviduct, is_exterior_face_xi3_1)
+            is_rightOviduct_inner = fm.createFieldAnd(is_rightOviduct, is_exterior_face_xi3_0)
+
+            serosaOfRightOviduct = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                                         get_uterus_term("serosa of right oviduct"))
+            serosaOfRightOviduct.getMeshGroup(mesh2d).addElementsConditional(is_rightOviduct_outer)
+
+            lumenOfLeftOviduct = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                                  get_uterus_term("lumen of left oviduct"))
+            lumenOfLeftOviduct.getMeshGroup(mesh2d).addElementsConditional(is_leftOviduct_inner)
+
+            lumenOfRightOviduct = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                                        get_uterus_term("lumen of right oviduct"))
+            lumenOfRightOviduct.getMeshGroup(mesh2d).addElementsConditional(is_rightOviduct_inner)
+
+            is_pubocervical = fm.createFieldAnd(is_body_outer, is_cervix_outer)
+            pubocervical = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                            get_uterus_term("pubocervical ligament"))
+            pubocervical.getMeshGroup(mesh1d).addElementsConditional(is_pubocervical)
+
+            is_internal_os = fm.createFieldAnd(is_body_inner, is_cervix_inner)
+            internalOs = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                                get_uterus_term("internal cervical os"))
+            internalOs.getMeshGroup(mesh1d).addElementsConditional(is_internal_os)
+
+            is_external_os = fm.createFieldAnd(is_vagina_inner, is_cervix_inner)
+            externalOs = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                            get_uterus_term("external cervical os"))
+            externalOs.getMeshGroup(mesh1d).addElementsConditional(is_external_os)
+
+            is_vagina_orifice = fm.createFieldAnd(is_vagina_xi2_01, is_exterior_face_xi3_0)
+            vaginaOrifice = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                            get_uterus_term("vagina orifice"))
+            vaginaOrifice.getMeshGroup(mesh1d).addElementsConditional(is_vagina_orifice)
+
+            # ligaments
+            is_dorsalVentral = fm.createFieldAnd(dorsalGroup.getGroup(), ventralGroup.getGroup())
+            is_dorsalVentralSerosa = fm.createFieldAnd(is_dorsalVentral, is_exterior_face_xi3_1)
+            is_leftDorsalVentralSerosa = fm.createFieldAnd(leftGroup.getGroup(), is_dorsalVentralSerosa)
+            is_rightDorsalVentralSerosa = fm.createFieldAnd(rightGroup.getGroup(), is_dorsalVentralSerosa)
+            fundusGroup = getAnnotationGroupForTerm(annotationGroups, get_uterus_term("fundus of uterus"))
+            is_bodyNotFundus = fm.createFieldAnd(bodyGroup.getGroup(), fm.createFieldNot(fundusGroup.getGroup()))
+
+            # Broad ligament of uterus
+            is_leftBroadLigament = fm.createFieldAnd(is_bodyNotFundus, is_leftDorsalVentralSerosa)
+            leftBroadLigament = findOrCreateAnnotationGroupForTerm(
+                annotationGroups, region, get_uterus_term("left broad ligament of uterus"))
+            leftBroadLigament.getMeshGroup(mesh1d).addElementsConditional(is_leftBroadLigament)
+            # add connected edges from left oviduct, avoiding adding dorsal-ventral edges on the superior edge
+            leftBroadLigament.addSubelements()  # need current nodes in ligament for group_add_connected_elements
+            tmpGroup = fm.createFieldGroup()
+            tmpMeshGroup = tmpGroup.createMeshGroup(mesh1d)
+            tmpMeshGroup.addElementsConditional(fm.createFieldAnd(is_leftOviduct, is_leftDorsalVentralSerosa))
+            group_add_connected_elements(leftBroadLigament.getGroup(), tmpMeshGroup)
+            del tmpMeshGroup
+            del tmpGroup
+
+            is_rightBroadLigament = fm.createFieldAnd(is_bodyNotFundus, is_rightDorsalVentralSerosa)
+            rightBroadLigament = findOrCreateAnnotationGroupForTerm(
+                annotationGroups, region, get_uterus_term("right broad ligament of uterus"))
+            rightBroadLigament.getMeshGroup(mesh1d).addElementsConditional(is_rightBroadLigament)
+            # add connected edges from right oviduct, avoiding adding dorsal-ventral edges on the superior edge
+            rightBroadLigament.addSubelements()  # need current nodes in ligament for group_add_connected_elements
+            tmpGroup = fm.createFieldGroup()
+            tmpMeshGroup = tmpGroup.createMeshGroup(mesh1d)
+            tmpMeshGroup.addElementsConditional(fm.createFieldAnd(is_rightOviduct, is_rightDorsalVentralSerosa))
+            group_add_connected_elements(rightBroadLigament.getGroup(), tmpMeshGroup)
+            del tmpMeshGroup
+            del tmpGroup
+
+            # Transverse cervical ligament
+            is_leftTransverseCervicalLigament = fm.createFieldAnd(cervixGroup.getGroup(), is_leftDorsalVentralSerosa)
+            leftTransverseCervicalLigament = findOrCreateAnnotationGroupForTerm(
+                annotationGroups, region, get_uterus_term("left transverse cervical ligament"))
+            leftTransverseCervicalLigament.getMeshGroup(mesh1d).addElementsConditional(
+                is_leftTransverseCervicalLigament)
+
+            is_rightTransverseCervicalLigament = fm.createFieldAnd(cervixGroup.getGroup(), is_rightDorsalVentralSerosa)
+            rightTransverseCervicalLigament = findOrCreateAnnotationGroupForTerm(
+                annotationGroups, region, get_uterus_term("right transverse cervical ligament"))
+            rightTransverseCervicalLigament.getMeshGroup(mesh1d).addElementsConditional(
+                is_rightTransverseCervicalLigament)
+
+        if isRodent:
+            is_rightHorn = rightOviductGroup.getGroup()
+            is_rightHorn_outer = fm.createFieldAnd(is_rightHorn, is_exterior_face_xi3_1)
+            is_rightHorn_inner = fm.createFieldAnd(is_rightHorn, is_exterior_face_xi3_0)
+
+            is_leftHorn = leftOviductGroup.getGroup()
+            is_leftHorn_outer = fm.createFieldAnd(is_leftHorn, is_exterior_face_xi3_1)
+            is_leftHorn_inner = fm.createFieldAnd(is_leftHorn, is_exterior_face_xi3_0)
+
+            serosaOfRightHorn = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                                   get_uterus_term("serosa of right horn"))
+            serosaOfRightHorn.getMeshGroup(mesh2d).addElementsConditional(is_rightHorn_outer)
+
+            lumenOfRightHorn = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                                  get_uterus_term("lumen of right horn"))
+            lumenOfRightHorn.getMeshGroup(mesh2d).addElementsConditional(is_rightHorn_inner)
+
+            serosaOfLeftHorn = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                                  get_uterus_term("serosa of left horn"))
+            serosaOfLeftHorn.getMeshGroup(mesh2d).addElementsConditional(is_leftHorn_outer)
+
+            lumenOfLeftHorn = findOrCreateAnnotationGroupForTerm(annotationGroups, region,
+                                                                 get_uterus_term("lumen of left horn"))
+            lumenOfLeftHorn.getMeshGroup(mesh2d).addElementsConditional(is_leftHorn_inner)
+
+        annotationGroups.remove(cervixGroup)
+        annotationGroups.remove(upperCervixGroup)
+        annotationGroups.remove(lowerCervixGroup)
 
 
-def findDerivativeBetweenPoints(v1, v2):
+def setNodeFieldParameters(field, fieldcache, x, d1, d2, d3, d12=None, d13=None):
     """
-    Find vector difference between two points and rescale vector difference using cubic hermite arclength
-    between the points to derive the derivative between the points.
-    :param v1: start vector
-    :param v2: end vector
-    :return: derivative of between v1 and v2
+    Assign node field parameters x, d1, d2, d3 of field.
+    :param field: Field parameters to assign.
+    :param fieldcache: Fieldcache with node set.
+    :param x: Parameters to set for Node.VALUE_LABEL_VALUE.
+    :param d1: Parameters to set for Node.VALUE_LABEL_D_DS1.
+    :param d2: Parameters to set for Node.VALUE_LABEL_D_DS2.
+    :param d3: Parameters to set for Node.VALUE_LABEL_D_DS3.
+    :param d12: Optional parameters to set for Node.VALUE_LABEL_D2_DS1DS2.
+    :param d13: Optional parameters to set for Node.VALUE_LABEL_D2_DS1DS3.
+    :return:
     """
-    d = [v2[c] - v1[c] for c in range(3)]
-    arcLengthAround = interp.computeCubicHermiteArcLength(v1, d, v2, d, True)
-    d = [c * arcLengthAround for c in normalize(d)]
+    field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_VALUE, 1, x)
+    field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_D_DS1, 1, d1)
+    field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_D_DS2, 1, d2)
+    field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_D_DS3, 1, d3)
+    if d12:
+        field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_D2_DS1DS2, 1, d12)
+    if d13:
+        field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_D2_DS1DS3, 1, d13)
 
-    return d
 
-
-class UterusNetworkLayout:
+def setNodeFieldVersionDerivatives(field, fieldcache, version, d1, d2, d3, d12=None, d13=None):
     """
-    Generates network layout for uterus scaffold.
+    Assign node field parameters d1, d2, d3 of field.
+    :param field: Field to assign parameters of.
+    :param fieldcache: Fieldcache with node set.
+    :param version: Version of d1, d2, d3 >= 1.
+    :param d1: Parameters to set for Node.VALUE_LABEL_D_DS1.
+    :param d2: Parameters to set for Node.VALUE_LABEL_D_DS2.
+    :param d3: Parameters to set for Node.VALUE_LABEL_D_DS3.
+    :param d12: Optional parameters to set for Node.VALUE_LABEL_D2_DS1DS2.
+    :param d13: Optional parameters to set for Node.VALUE_LABEL_D2_DS1DS3.
+    :return:
     """
-    def __init__(self, region, networkLayout, targetElementLength):
-        """
-        :param region: Zinc region to define model in.
-        :param networkLayout: Network layout subscaffold from MeshType_1d_network_layout1
-        """
-
-        layoutRegion = region.createRegion()
-        layoutFieldmodule = layoutRegion.getFieldmodule()
-        layoutNodes = layoutFieldmodule.findNodesetByFieldDomainType(Field.DOMAIN_TYPE_NODES)
-        # layoutMesh = layoutFieldmodule.findMeshByDimension(1)
-        networkLayout.generate(layoutRegion)  # ask scaffold to generate to get user-edited parameters
-        # layoutAnnotationGroups = networkLayout.getAnnotationGroups()
-        layoutCoordinates = findOrCreateFieldCoordinates(layoutFieldmodule)
-        layoutFieldcache = layoutFieldmodule.createFieldcache()
-
-        networkMesh = networkLayout.getConstructionObject()
-
-        networkSegments = networkMesh.getNetworkSegments()
-
-        arcLengthOfGroupsAlong = []
-        elementsCountAlongList = []
-        cxGroups = []
-        sxGroups = []
-        for n1 in range(len(networkSegments)):
-            networkSegment = networkSegments[n1]
-            segmentNodes = networkSegment.getNetworkNodes()
-            segmentNodeCount = len(segmentNodes)
-            for n in range(segmentNodeCount):
-                segmentNode = segmentNodes[n]
-                layoutNodeIdentifier = segmentNode.getNodeIdentifier()
-                layoutNode = layoutNodes.findNodeByIdentifier(layoutNodeIdentifier)
-                layoutFieldcache.setNode(layoutNode)
-                cx, cd1, cd2, cd3 = get_nodeset_path_ordered_field_parameters(
-                    layoutNodes, layoutCoordinates, [Node.VALUE_LABEL_VALUE, Node.VALUE_LABEL_D_DS1,
-                                                     Node.VALUE_LABEL_D_DS2, Node.VALUE_LABEL_D_DS3],
-                    networkSegments[n1].getNodeIdentifiers(), networkSegments[n1].getNodeVersions())
-            cxGroup = [cx, cd1, cd2, [], cd3, []]
-            arcLength = 0.0
-            for e in range(len(cx) - 1):
-                arcLength += interp.getCubicHermiteArcLength(cx[e], cd1[e],
-                                                             cx[e + 1], cd1[e + 1])
-            arcLengthOfGroupsAlong.append(arcLength)
-            elementsAlong = math.ceil(arcLength / targetElementLength)
-            elementsCountAlongList.append(elementsAlong)
-            cxGroups.append(cxGroup)
-            # Sample
-            sx, sd1, pe, pxi, psf = interp.sampleCubicHermiteCurves(cx, cd1, elementsAlong)
-            # sd2, sd12 = interp.interpolateSampleCubicHermite(cd2, cd12, pe, pxi, psf)
-            # sd3, sd13 = interp.interpolateSampleCubicHermite(cd3, cd13, pe, pxi, psf)
-            sxGroup = [sx, sd1]
-            sxGroups.append(sxGroup)
-
-        del layoutCoordinates
-        del layoutNodes
-        del layoutFieldmodule
-        del layoutRegion
-
-        self.cxGroups = cxGroups
-        self.sxGroups = sxGroups
-        self.arcLengthOfGroupsAlong = arcLengthOfGroupsAlong
-        self.elementsCountAlongList = elementsCountAlongList
-
-
-def createUterusMesh3D(region, fm, coordinates, geometricNetworkLayout, elementsCountAround,
-                       elementsCountAroundRightHorn, elementsCountAroundLeftHorn, elementsCountThroughWall,
-                       elementsCountInRightHorn, elementsCountInLeftHorn, elementsCountInBody, elementsCountInCervix,
-                       elementsCountInVagina, wallThickness, doubleUterus, useCrossDerivatives):
-
-    mesh = fm.findMeshByDimension(3)
-
-    firstNodeIdentifier = 1
-    firstElementIdentifier = 1
-
-    cx_right_horn_group = geometricNetworkLayout.cxGroups[0]
-    cx_left_horn_group = geometricNetworkLayout.cxGroups[1]
-    cx_body_group = geometricNetworkLayout.cxGroups[2]
-    cx_cervix_group = geometricNetworkLayout.cxGroups[3]
-    cx_vagina_group = geometricNetworkLayout.cxGroups[4]
-
-    sx_right_horn_group = geometricNetworkLayout.sxGroups[0]
-    sx_left_horn_group = geometricNetworkLayout.sxGroups[1]
-    sx_body_group = geometricNetworkLayout.sxGroups[2]
-    sx_cervix_group = geometricNetworkLayout.sxGroups[3]
-
-    # Create annotation groups
-    rightHornGroup = AnnotationGroup(region, get_uterus_term("right uterine horn"))
-    leftHornGroup = AnnotationGroup(region, get_uterus_term("left uterine horn"))
-    bodyGroup = AnnotationGroup(region, get_uterus_term("body of uterus"))
-    cervixGroup = AnnotationGroup(region, get_uterus_term("uterine cervix"))
-    vaginaGroup = AnnotationGroup(region, get_uterus_term("vagina"))
-    uterusGroup = AnnotationGroup(region, get_uterus_term("uterus"))
-    annotationGroups = [cervixGroup, bodyGroup, leftHornGroup, rightHornGroup, uterusGroup, vaginaGroup]
-
-    rightHornMeshGroup = rightHornGroup.getMeshGroup(mesh)
-    leftHornMeshGroup = leftHornGroup.getMeshGroup(mesh)
-    bodyMeshGroup = bodyGroup.getMeshGroup(mesh)
-    cervixMeshGroup = cervixGroup.getMeshGroup(mesh)
-    vaginaMeshGroup = vaginaGroup.getMeshGroup(mesh)
-    uterusMeshGroup = uterusGroup.getMeshGroup(mesh)
-
-    elementsCountAcross = elementsCountAroundRightHorn - elementsCountAround // 2
-
-    # Get right horn nodes
-    rightStartRadians = -math.pi * (elementsCountAround / (2 * elementsCountAroundRightHorn))
-    rightHornCoordinates = getTubeNodes(cx_right_horn_group, elementsCountAroundRightHorn,
-                                        elementsCountInRightHorn, elementsCountThroughWall, wallThickness,
-                                        startRadian=rightStartRadians)
-
-    rhLastRingNodeCoordinates = getTargetedRingNodesCoordinates(rightHornCoordinates, elementsCountAroundRightHorn,
-                                                                elementsCountInRightHorn, elementsCountThroughWall,
-                                                                omitStartRows=0, omitEndRows=1)
-
-    rhLastRingNodeId, nodeCount = getTargetedRingNodesIds(firstNodeIdentifier, elementsCountAroundRightHorn,
-                                                          elementsCountInRightHorn, elementsCountThroughWall,
-                                                          omitStartRows=0, omitEndRows=1)
-
-    # Get left horn nodes
-    leftStartRadians = -math.pi * (elementsCountAcross / elementsCountAroundLeftHorn)
-    leftHornCoordinates = getTubeNodes(cx_left_horn_group, elementsCountAroundLeftHorn, elementsCountInLeftHorn,
-                                       elementsCountThroughWall, wallThickness, startRadian=leftStartRadians)
-
-    lhLastRingNodeCoordinates = getTargetedRingNodesCoordinates(leftHornCoordinates, elementsCountAroundLeftHorn,
-                                                                elementsCountInLeftHorn, elementsCountThroughWall,
-                                                                omitStartRows=0, omitEndRows=1)
-
-    lhLastRingNodeId, nodeCount = getTargetedRingNodesIds(nodeCount, elementsCountAroundLeftHorn,
-                                                          elementsCountInLeftHorn, elementsCountThroughWall,
-                                                          omitStartRows=0, omitEndRows=1)
-
-    if doubleUterus:
-        # Get body nodes
-        bodyInnerRightCoordinates, bodyInnerLeftCoordinates, bodyCoordinatesOuter, septumBodyCoordinates = \
-            getDoubleTubeNodes(cx_body_group, elementsCountInBody, elementsCountAround, elementsCountAroundRightHorn,
-                               elementsCountAroundLeftHorn, elementsCountThroughWall, wallThickness)
-
-        # Get cervix nodes
-        rightInnerCervixthroughWall, leftInnerCervixthroughWall, cervixCoordinatesOuter, septumCervixCoordinates = \
-            getDoubleTubeNodes(cx_cervix_group, elementsCountInCervix, elementsCountAround,
-                               elementsCountAroundRightHorn, elementsCountAroundLeftHorn, elementsCountThroughWall,
-                               wallThickness)
-
-        # Get the first layer of inner right and left body coordinates for sampling with the last layer of horns
-        xrList = []
-        d1rList = []
-        d2rList = []
-        xlList = []
-        d1lList = []
-        d2lList = []
-        for n2 in range(elementsCountInBody + 1):
-            for n3 in range(elementsCountThroughWall + 1):
-                for n1 in range(elementsCountAroundRightHorn):
-                    x = bodyInnerRightCoordinates[0][n2][n3][n1]
-                    d1 = bodyInnerRightCoordinates[1][n2][n3][n1]
-                    d2 = bodyInnerRightCoordinates[2][n2][n3][n1]
-                    xrList.append(x)
-                    d1rList.append(d1)
-                    d2rList.append(d2)
-                for n1 in range(elementsCountAroundLeftHorn):
-                    x = bodyInnerLeftCoordinates[0][n2][n3][n1]
-                    d1 = bodyInnerLeftCoordinates[1][n2][n3][n1]
-                    d2 = bodyInnerLeftCoordinates[2][n2][n3][n1]
-                    xlList.append(x)
-                    d1lList.append(d1)
-                    d2lList.append(d2)
-        newbodyInnerRightCoordinates = [xrList, d1rList, d2rList]
-        newbodyInnerLeftCoordinates = [xlList, d1lList, d2lList]
-
-        brFirstRingNodeCoordinates = \
-            getTargetedRingNodesCoordinates(newbodyInnerRightCoordinates, elementsCountAroundRightHorn,
-                                            elementsCountInBody, elementsCountThroughWall, omitStartRows=1,
-                                            omitEndRows=0)
-        blFirstRingNodeCoordinates = \
-            getTargetedRingNodesCoordinates(newbodyInnerLeftCoordinates, elementsCountAroundLeftHorn,
-                                            elementsCountInBody, elementsCountThroughWall, omitStartRows=1,
-                                            omitEndRows=0)
-
-        # Sample the curve between last horn rings and first body rings to get the inner bifurcation nodes
-        xLayersRight = []
-        d2LayersRight = []
-        xLayersLeft = []
-        d2LayersLeft = []
-        elementsCountAlongBifurcation = 2
-        for n3 in range(elementsCountThroughWall):
-            xRawRight = []
-            d2RawRight = []
-            for n1 in range(elementsCountAroundRightHorn):
-                xAlong = []
-                d2Along = []
-                xAlong.append(rhLastRingNodeCoordinates[0][n3][n1])
-                xAlong.append(brFirstRingNodeCoordinates[0][n3][n1])
-                d2Along.append(rhLastRingNodeCoordinates[2][n3][n1])
-                d2Along.append(brFirstRingNodeCoordinates[2][n3][n1])
-                xSampledAlong, d2SampledAlong = interp.sampleCubicHermiteCurves(xAlong, d2Along,
-                                                                                elementsCountAlongBifurcation,
-                                                                                arcLengthDerivatives=True)[0:2]
-                xRawRight.append(xSampledAlong)
-                d2RawRight.append(d2SampledAlong)
-            xLayersRight.append(xRawRight)
-            d2LayersRight.append(d2RawRight)
-            xRawLeft = []
-            d2RawLeft = []
-            for n1 in range(elementsCountAroundLeftHorn):
-                xAlong = []
-                d2Along = []
-                xAlong.append(lhLastRingNodeCoordinates[0][n3][n1])
-                xAlong.append(blFirstRingNodeCoordinates[0][n3][n1])
-                d2Along.append(lhLastRingNodeCoordinates[2][n3][n1])
-                d2Along.append(blFirstRingNodeCoordinates[2][n3][n1])
-                xSampledAlong, d2SampledAlong = interp.sampleCubicHermiteCurves(xAlong, d2Along,
-                                                                                elementsCountAlongBifurcation,
-                                                                                arcLengthDerivatives=True)[0:2]
-                xRawLeft.append(xSampledAlong)
-                d2RawLeft.append(d2SampledAlong)
-            xLayersLeft.append(xRawLeft)
-            d2LayersLeft.append(d2RawLeft)
-
-        # Rearrange x and d2
-        xSampledBifurcationRight = []
-        d1SampledBifurcationRight = []
-        d2SampledBifurcationRight = []
-        xSampledBifurcationLeft = []
-        d1SampledBifurcationLeft = []
-        d2SampledBifurcationLeft = []
-        for n2 in range(2 + 1):
-            xRawRight = []
-            d1RawRight = []
-            d2RawRight = []
-            xRawLeft = []
-            d1RawLeft = []
-            d2RawLeft = []
-            for n3 in range(elementsCountThroughWall):
-                xAroundRight = []
-                d1AroundRight = []
-                d2AroundRight = []
-                for n1 in range(elementsCountAroundRightHorn):
-                    x = xLayersRight[n3][n1][n2]
-                    d2 = d2LayersRight[n3][n1][n2]
-                    xAroundRight.append(x)
-                    d2AroundRight.append(d2)
-                    # Calculate d1
-                    v1 = xLayersRight[n3][n1][n2]
-                    v2 = xLayersRight[n3][n1 + 1 if n1 < elementsCountAroundRightHorn - 1 else 0][n2]
-                    d1 = findDerivativeBetweenPoints(v1, v2)
-                    d1AroundRight.append(d1)
-                d1Smoothed = interp.smoothCubicHermiteDerivativesLoop(xAroundRight, d1AroundRight)
-                xRawRight.append(xAroundRight)
-                d1RawRight.append(d1Smoothed)
-                d2RawRight.append(d2AroundRight)
-                xAroundLeft = []
-                d1AroundLeft = []
-                d2AroundLeft = []
-                for n1 in range(elementsCountAroundLeftHorn):
-                    x = xLayersLeft[n3][n1][n2]
-                    d2 = d2LayersLeft[n3][n1][n2]
-                    xAroundLeft.append(x)
-                    d2AroundLeft.append(d2)
-                    # Calculate d1
-                    v1 = xLayersLeft[n3][n1][n2]
-                    v2 = xLayersLeft[n3][n1 + 1 if n1 < elementsCountAroundLeftHorn - 1 else 0][n2]
-                    d1 = findDerivativeBetweenPoints(v1, v2)
-                    d1AroundLeft.append(d1)
-                d1Smoothed = interp.smoothCubicHermiteDerivativesLoop(xAroundLeft, d1AroundLeft)
-                xRawLeft.append(xAroundLeft)
-                d1RawLeft.append(d1Smoothed)
-                d2RawLeft.append(d2AroundLeft)
-            xSampledBifurcationRight.append(xRawRight)
-            d1SampledBifurcationRight.append(d1RawRight)
-            d2SampledBifurcationRight.append(d2RawRight)
-            xSampledBifurcationLeft.append(xRawLeft)
-            d1SampledBifurcationLeft.append(d1RawLeft)
-            d2SampledBifurcationLeft.append(d2RawLeft)
-
-        innerBifurcationRight = [xSampledBifurcationRight, d1SampledBifurcationRight, d2SampledBifurcationRight]
-        innerBifurcationLeft = [xSampledBifurcationLeft, d1SampledBifurcationLeft, d2SampledBifurcationLeft]
-
-        # Get bifurcation outer nodes
-        paCentre = sx_body_group[0][1]
-        c1Centre = sx_right_horn_group[0][-2]
-        c2Centre = sx_left_horn_group[0][-2]
-        paxList = bodyCoordinatesOuter[0][1]
-        pad2 = bodyCoordinatesOuter[2][1]
-        pad3 = bodyCoordinatesOuter[3][1]
-        c1xList = rhLastRingNodeCoordinates[0][-1]
-        c1d2 = rhLastRingNodeCoordinates[2][-1]
-        c2xList = lhLastRingNodeCoordinates[0][-1]
-        c2d2 = lhLastRingNodeCoordinates[2][-1]
-        rox, rod1, rod2, cox, cod1, cod2, paStartIndex, c1StartIndex, c2StartIndex = \
-            find_tube_bifurcation_points_converging(paCentre, paxList, pad2, c1Centre, c1xList, c1d2, c2Centre, c2xList,
-                                                    c2d2)
-        rod3 = pad3
-
-        # Get d3 for outer bifurcation (for cox nodes)
-        cod3 = []
-        for n in range(len(cox)):
-            v1 = septumBodyCoordinates[0][0][n + 1]
-            v2 = cox[n]
-            d3 = findDerivativeBetweenPoints(v1, v2)
-            cod3.append(d3)
-
-        # # Get vagina nodes
-        # vaginaInnerRightCoordinates, vaginaInnerLeftCoordinates, vaginaCoordinatesOuter, septumVaginaCoordinates = \
-        #     getDoubleTubeNodes(cx_vagina_group, elementsCountInVagina, elementsCountAround,
-        #                        elementsCountAroundRightHorn, elementsCountAroundLeftHorn, elementsCountThroughWall,
-        #                        wallThickness)
-
-
-        # # Create nodes through wall for sampled inner right bifurcation nodes
-        # xRaw = []
-        # d3Raw = []
-        # for n1 in range(elementsCountAroundRightHorn):
-        #     xAlong = []
-        #     d3Along = []
-        #     v1 = xSampledBifurcationRight[1][n1]
-        #     v1d1 = d3SampledBifurcationRight[n1]
-        #     if n1 <= elementsCountAround // 2:
-        #         v2 = rox[n1]
-        #         if n1 == 0 or elementsCountAround // 2:
-        #             v2d1 = d3SampledBifurcationRight[n1]
-        #         else:
-        #             v2d1 = rod3[n1]
-        #     else:
-        #         v2 = cox[n1 - elementsCountAround // 2 - 1]
-        #         # v2d1 = cod3[n1 - elementsCountAround // 2 - 1]
-        #         v2d1 = d3SampledBifurcationRight[n1]
-        #     xAlong.append(v1)
-        #     xAlong.append(v2)
-        #     d3Along.append(v1d1)
-        #     d3Along.append(v2d1)
-        #     xSampledWall, d3SampledWall = interp.sampleCubicHermiteCurves(xAlong, d3Along, elementsCountThroughWall,
-        #                                                                     arcLengthDerivatives=True)[0:2]
-        #     xRaw.append(xSampledWall)
-        #     d3Raw.append(d3SampledWall)
-        # print('len(xRaw)', len(xRaw))
-        # print('len(xRaw[0])', len(xRaw[0]))
-
-    else:
-        # Get body nodes
-        bodyCoordinates = getTubeNodes(cx_body_group, elementsCountAround, elementsCountInBody,
-                                       elementsCountThroughWall, wallThickness, startRadian=-math.pi / 2)
-
-        bFirstRingNodeCoordinates = getTargetedRingNodesCoordinates(bodyCoordinates, elementsCountAround,
-                                                                    elementsCountInBody, elementsCountThroughWall,
-                                                                    omitStartRows=1, omitEndRows=0)
-        # Get cervix nodes
-        cervixCoordinates = getTubeNodes(cx_cervix_group, elementsCountAround, elementsCountInCervix,
-                                         elementsCountThroughWall, wallThickness, startRadian=-math.pi / 2)
-
-        # Get vagina nodes
-        vaginaCoordinates = getTubeNodes(cx_vagina_group, elementsCountAround, elementsCountInVagina,
-                                         elementsCountThroughWall, wallThickness, startRadian=-math.pi / 2)
-
-    # Create nodes
-    # Create right horn nodes
-    nodeIdentifier = createTubeNodes(fm, coordinates, firstNodeIdentifier, rightHornCoordinates,
-                                     elementsCountInRightHorn, elementsCountAroundRightHorn, elementsCountThroughWall,
-                                     omitStartRows=0, omitEndRows=1)
-
-    # Create left horn nodes
-    nodeIdentifier = createTubeNodes(fm, coordinates, nodeIdentifier, leftHornCoordinates, elementsCountInLeftHorn,
-                                     elementsCountAroundLeftHorn, elementsCountThroughWall, omitStartRows=0,
-                                     omitEndRows=1)
-
-    if doubleUterus:
-        # Create bifurcation nodes
-        nodeIdentifier, roNodeId, coNodeId, birNodeId, bilNodeId = \
-            createDoubleTubeBifurcationNodes(fm, nodeIdentifier, rox, rod1, rod2, rod3, cox, cod1, cod2, cod3,
-                                             innerBifurcationRight, innerBifurcationLeft, elementsCountAroundRightHorn,
-                                             elementsCountAroundLeftHorn, elementsCountThroughWall)
-
-        # Create body nodes
-        nodeIdentifier, bricNodeId, blicNodeId, botNodeId, bsNodeId = \
-            createDoubleTubeNodes(fm, nodeIdentifier, bodyInnerRightCoordinates, bodyInnerLeftCoordinates,
-                                  bodyCoordinatesOuter, septumBodyCoordinates, elementsCountInBody, elementsCountAround,
-                                  elementsCountAcross, elementsCountThroughWall, omitStartRows=1, omitEndRows=1)
-
-        # Create cervix nodes
-        nodeIdentifier, cricNodeId, clicNodeId, cotNodeId, csNodeId = \
-            createDoubleTubeNodes(fm, nodeIdentifier, rightInnerCervixthroughWall, leftInnerCervixthroughWall,
-                                  cervixCoordinatesOuter, septumCervixCoordinates, elementsCountInCervix,
-                                  elementsCountAround, elementsCountAcross, elementsCountThroughWall, omitStartRows=0,
-                                  omitEndRows=0)  # omitEndRows should be 1 when we add vagina
-
-        # # Create vagina nodes
-        # nodeIdentifier, vricNodeId, vlicNodeId, votNodeId, vsNodeId = \
-        #     createDoubleTubeNodes(fm, nodeIdentifier, vaginaInnerRightCoordinates, vaginaInnerLeftCoordinates,
-        #                           vaginaCoordinatesOuter, septumVaginaCoordinates, elementsCountInVagina,
-        #                           elementsCountAround, elementsCountAcross, elementsCountThroughWall, omitStartRows=0,
-        #                           omitEndRows=0)
-
-    else:
-        # Create bifurcation nodes
-        paCentre = sx_cervix_group[0][1]
-        c1Centre = sx_right_horn_group[0][-2]
-        c2Centre = sx_left_horn_group[0][-2]
-        paxList = bFirstRingNodeCoordinates[0]
-        pad2 = bFirstRingNodeCoordinates[2]
-        c1xList = rhLastRingNodeCoordinates[0]
-        c1d2 = rhLastRingNodeCoordinates[2]
-        c2xList = lhLastRingNodeCoordinates[0]
-        c2d2 = lhLastRingNodeCoordinates[2]
-        nodeIdentifier, roNodeId, coNodeId = \
-            createTubeBifurcationNodes(fm, coordinates, nodeIdentifier, paCentre, paxList, pad2, c1Centre, c1xList,
-                                       c1d2, c2Centre, c2xList, c2d2, elementsCountThroughWall)
-
-        # Create body nodes
-        nodeCount = nodeIdentifier
-        nodeIdentifier = createTubeNodes(fm, coordinates, nodeIdentifier, bodyCoordinates, elementsCountInBody,
-                                         elementsCountAround, elementsCountThroughWall, omitStartRows=1, omitEndRows=0)
-
-        # Create cervix nodes
-        nodeIdentifier = createTubeNodes(fm, coordinates, nodeIdentifier, cervixCoordinates, elementsCountInCervix,
-                                         elementsCountAround, elementsCountThroughWall, omitStartRows=1, omitEndRows=0)
-
-        # Create vagina nodes
-        nodeIdentifier = createTubeNodes(fm, coordinates, nodeIdentifier, vaginaCoordinates, elementsCountInVagina,
-                                         elementsCountAround, elementsCountThroughWall, omitStartRows=1, omitEndRows=0)
-
-    # Create elements
-    # Create right horn elements
-    startNodeId = firstNodeIdentifier
-    elementIdentifier = \
-        make_tube_elements(fm, coordinates, startNodeId, firstElementIdentifier, elementsCountInRightHorn,
-                           elementsCountAroundRightHorn, elementsCountThroughWall, useCrossDerivatives, omitStartRows=0,
-                           omitEndRows=1, meshGroups=[rightHornMeshGroup, uterusMeshGroup])
-
-    # Create left horn elements
-    startNodeId = rhLastRingNodeId[-1][-1] + 1
-    elementIdentifier = make_tube_elements(fm, coordinates, startNodeId, elementIdentifier, elementsCountInLeftHorn,
-                                           elementsCountAroundLeftHorn, elementsCountThroughWall, useCrossDerivatives,
-                                           omitStartRows=0, omitEndRows=1,
-                                           meshGroups=[leftHornMeshGroup, uterusMeshGroup])
-
-    if doubleUterus:
-        # Create bifurcation elements
-        c1NodeId = rhLastRingNodeId
-        c2NodeId = lhLastRingNodeId
-        if elementsCountInBody < 2:
-            bricNodeId = cricNodeId[:elementsCountThroughWall * elementsCountAroundRightHorn]
-            blicNodeId = clicNodeId[:elementsCountThroughWall * elementsCountAroundLeftHorn]
-            botNodeId = cotNodeId[:elementsCountAround]
-            bsNodeId = csNodeId[:elementsCountAcross - 1]
-
-        elementIdentifier = \
-            make_double_tube_bifurcation_elements(fm, coordinates, elementIdentifier, elementsCountThroughWall,
-                                                  c1NodeId, c2NodeId, roNodeId, coNodeId, birNodeId, bilNodeId,
-                                                  bricNodeId, blicNodeId, botNodeId, bsNodeId,
-                                                  meshGroups=[bodyMeshGroup, rightHornMeshGroup, leftHornMeshGroup,
-                                                              uterusMeshGroup])
-
-        # Create body elements
-        if elementsCountInBody >= 2:
-            elementIdentifier = \
-                make_double_tube_elements(mesh, coordinates, elementIdentifier, elementsCountInBody - 1,
-                                          elementsCountAround, elementsCountAroundRightHorn,
-                                          elementsCountAroundLeftHorn, elementsCountThroughWall, bricNodeId, blicNodeId,
-                                          botNodeId, bsNodeId, useCrossDerivatives,
-                                          meshGroups=[bodyMeshGroup, uterusMeshGroup])
-
-        # Create cervix elements
-        elementIdentifier = \
-            make_double_tube_elements(mesh, coordinates, elementIdentifier, elementsCountInCervix, elementsCountAround,
-                                      elementsCountAroundRightHorn, elementsCountAroundLeftHorn,
-                                      elementsCountThroughWall, cricNodeId, clicNodeId, cotNodeId, csNodeId,
-                                      useCrossDerivatives, meshGroups=[cervixMeshGroup, uterusMeshGroup])
-
-        # # Create vagina elements
-        # elementIdentifier = \
-        #     make_double_tube_elements(mesh, coordinates, elementIdentifier, elementsCountInVagina, elementsCountAround,
-        #                               elementsCountAroundRightHorn, elementsCountAroundLeftHorn,
-        #                               elementsCountThroughWall, vricNodeId, vlicNodeId, votNodeId, vsNodeId,
-        #                               useCrossDerivatives, meshGroups=[vaginaMeshGroup])
-    else:
-        # Create bifurcation elements
-        bFirstRingNodeId, nodeCount = getTargetedRingNodesIds(nodeCount, elementsCountAround, elementsCountInBody,
-                                                              elementsCountThroughWall, omitStartRows=1, omitEndRows=0)
-        paNodeId = bFirstRingNodeId
-        c1NodeId = rhLastRingNodeId
-        c2NodeId = lhLastRingNodeId
-        elementIdentifier = make_tube_bifurcation_elements(fm, coordinates, elementIdentifier, elementsCountThroughWall,
-                                                           paNodeId, c1NodeId, c2NodeId, roNodeId, coNodeId,
-                                                           meshGroups=[bodyMeshGroup, rightHornMeshGroup,
-                                                                       leftHornMeshGroup, uterusMeshGroup])
-
-        # Create body elements
-        startNodeId = paNodeId[0][0]
-        elementIdentifier = make_tube_elements(fm, coordinates, startNodeId, elementIdentifier, elementsCountInBody,
-                                               elementsCountAround, elementsCountThroughWall, useCrossDerivatives,
-                                               omitStartRows=1, omitEndRows=0,
-                                               meshGroups=[bodyMeshGroup, uterusMeshGroup])
-
-        # Create cervix elements
-        startNodeId = paNodeId[0][0] + (elementsCountInBody - 1) * elementsCountAround * (elementsCountThroughWall + 1)
-        elementIdentifier = make_tube_elements(fm, coordinates, startNodeId, elementIdentifier, elementsCountInCervix,
-                                               elementsCountAround, elementsCountThroughWall, useCrossDerivatives,
-                                               omitStartRows=0, omitEndRows=0,
-                                               meshGroups=[cervixMeshGroup, uterusMeshGroup])
-
-        # Create vagina elements
-        startNodeId = startNodeId + elementsCountInCervix * elementsCountAround * (elementsCountThroughWall + 1)
-        elementIdentifier = make_tube_elements(fm, coordinates, startNodeId, elementIdentifier, elementsCountInVagina,
-                                               elementsCountAround, elementsCountThroughWall, useCrossDerivatives,
-                                               omitStartRows=0, omitEndRows=0, meshGroups=[vaginaMeshGroup])
-
-    return nodeIdentifier, elementIdentifier, annotationGroups
-
-
-def getTubeNodes(cx_group, elementsCountAround, elementsCountAlongTube, elementsCountThroughWall, wallThickness,
-                 startRadian):
-
-    # Create ellipses along tube around the central path
-    xEllipsesAlong = []
-    d1EllipsesAlong = []
-    for n in range(len(cx_group[0])):
-        px, pd1 = createEllipsePoints(cx_group[0][n], 2 * math.pi, cx_group[2][n], cx_group[4][n], elementsCountAround,
-                                      startRadians=startRadian)
-        xEllipsesAlong.append(px)
-        d1EllipsesAlong.append(pd1)
-
-    # Find d2
-    d2Raw = []
-    for n1 in range(elementsCountAround):
-        xAlong = []
-        d2Along = []
-        for n2 in range(len(xEllipsesAlong) - 1):
-            v1 = xEllipsesAlong[n2][n1]
-            v2 = xEllipsesAlong[n2 + 1][n1]
-            d2 = findDerivativeBetweenPoints(v1, v2)
-            xAlong.append(v1)
-            d2Along.append(d2)
-        xAlong.append(xEllipsesAlong[-1][n1])
-        d2Along.append(d2)
-        d2Smoothed = interp.smoothCubicHermiteDerivativesLine(xAlong, d2Along)
-        d2Raw.append(d2Smoothed)
-
-    # Rearrange d2
-    d2EllipsesAlong = []
-    for n2 in range(len(xEllipsesAlong)):
-        d2Around = []
-        for n1 in range(elementsCountAround):
-            d2 = d2Raw[n1][n2]
-            d2Around.append(d2)
-        d2EllipsesAlong.append(d2Around)
-
-    # Spread out elements along tube
-    xRaw = []
-    d2Raw = []
-    for n1 in range(elementsCountAround):
-        xAlong = []
-        d2Along = []
-        for n2 in range(len(xEllipsesAlong)):
-            xAlong.append(xEllipsesAlong[n2][n1])
-            d2Along.append(d2EllipsesAlong[n2][n1])
-        xSampledAlong, d2SampledAlong = interp.sampleCubicHermiteCurves(xAlong, d2Along, elementsCountAlongTube,
-                                                                        arcLengthDerivatives=True)[0:2]
-        d2Smoothed = interp.smoothCubicHermiteDerivativesLine(xSampledAlong, d2SampledAlong)
-        xRaw.append(xSampledAlong)
-        d2Raw.append(d2SampledAlong)
-
-    # Rearrange x and d2
-    xSampledTube = []
-    d1SampledTube = []
-    d2SampledTube = []
-    for n2 in range(elementsCountAlongTube + 1):
-        xAround = []
-        d1Around = []
-        d2Around = []
-        for n1 in range(elementsCountAround):
-            x = xRaw[n1][n2]
-            d2 = d2Raw[n1][n2]
-            xAround.append(x)
-            d2Around.append(d2)
-            # Calculate d1
-            v1 = xRaw[n1][n2]
-            v2 = xRaw[n1 + 1 if n1 < elementsCountAround - 1 else 0][n2]
-            d1 = findDerivativeBetweenPoints(v1, v2)
-            d1Around.append(d1)
-        d1Smoothed = interp.smoothCubicHermiteDerivativesLoop(xAround, d1Around)
-        xSampledTube.append(xAround)
-        d1SampledTube.append(d1Smoothed)
-        d2SampledTube.append(d2Around)
-
-    d3Tube = []
-    for n2 in range(elementsCountAlongTube + 1):
-        d3Around = []
-        for n1 in range(elementsCountAround):
-            d3Around.append(normalize(
-                cross(normalize(d1SampledTube[n2][n1]), normalize(d2SampledTube[n2][n1]))))
-        d3Tube.append(d3Around)
-
-    xInner = []
-    d1Inner = []
-    d2Inner = []
-    d3Inner = []
-    for n2 in range(elementsCountAlongTube + 1):
-        for n1 in range(elementsCountAround):
-            n = n2 * elementsCountAround + n1
-            xInner.append(xSampledTube[n2][n1])
-            d1Inner.append(d1SampledTube[n2][n1])
-            d2Inner.append(d2SampledTube[n2][n1])
-            d3Inner.append(d3Tube[n2][n1])
-
-    transitElementList = [0] * elementsCountAround
-    relativeThicknessList = []
-    xList, d1List, d2List, d3List, curvatureList = \
-        tubemesh.extrudeSurfaceCoordinates(xInner, d1Inner, d2Inner, d3Inner,
-                                           [wallThickness] * (elementsCountAlongTube + 1), relativeThicknessList,
-                                           elementsCountAround, elementsCountAlongTube, elementsCountThroughWall,
-                                           transitElementList, outward=False)[0:5]
-
-    coordinatesList = [xList, d1List, d2List, d3List]
-
-    return coordinatesList
-
-
-def getTargetedRingNodesCoordinates(tubeCoordinates, elementsCountAround, elementsCountAlongTube,
-                                    elementsCountThroughWall, omitStartRows, omitEndRows):
-
-    # Create tube nodes
-    lastRingsNodeId = []
-    xLastRing = []
-    d1LastRing = []
-    d2LastRing = []
-    d3LastRing = []
-    firstRingsNodeId = []
-    xFirstRing = []
-    d1FirstRing = []
-    d2FirstRing = []
-    d3FirstRing = []
-    for n2 in range(elementsCountAlongTube + 1):
-        for n3 in range(elementsCountThroughWall + 1):
-            lastRingNodeIdThroughWall = []
-            xLastRingThroughWall = []
-            d1LastRingThroughWall = []
-            d2LastRingThroughWall = []
-            # d3LastRingThroughWall = []
-            firstRingNodeIdThroughWall = []
-            xFirstRingThroughWall = []
-            d1FirstRingThroughWall = []
-            d2FirstRingThroughWall = []
-            # d3FirstRingThroughWall = []
-            for n1 in range(elementsCountAround):
-                n = n2 * elementsCountAround * (elementsCountThroughWall + 1) + n3 * elementsCountAround + n1
-                x = tubeCoordinates[0][n]
-                d1 = tubeCoordinates[1][n]
-                d2 = tubeCoordinates[2][n]
-                # d3 = tubeCoordinates[3][n]
-                if omitEndRows == 1:  # merging to the bifurcation
-                    if n2 == elementsCountAlongTube:
-                        pass
-                    else:
-                        if n2 == elementsCountAlongTube - 1:
-                            xLastRingThroughWall.append(x)
-                            d1LastRingThroughWall.append(d1)
-                            d2LastRingThroughWall.append(d2)
-                            # d3LastRingThroughWall.append(d3)
-                elif omitStartRows == 1:  # diverging from bifurcation
-                    if n2 == 0:
-                        pass
-                    else:
-                        if n2 == 1:
-                            xFirstRingThroughWall.append(x)
-                            d1FirstRingThroughWall.append(d1)
-                            d2FirstRingThroughWall.append(d2)
-                            # d3FirstRingThroughWall.append(d3)
-            if omitEndRows == 1:
-                if n2 == elementsCountAlongTube - 1:
-                    lastRingsNodeId.append(lastRingNodeIdThroughWall)
-                    xLastRing.append(xLastRingThroughWall)
-                    d1LastRing.append(d1LastRingThroughWall)
-                    d2LastRing.append(d2LastRingThroughWall)
-                    # d3LastRing.append(d3LastRingThroughWall)
-            elif omitStartRows == 1:
-                if n2 == 1:
-                    firstRingsNodeId.append(firstRingNodeIdThroughWall)
-                    xFirstRing.append(xFirstRingThroughWall)
-                    d1FirstRing.append(d1FirstRingThroughWall)
-                    d2FirstRing.append(d2FirstRingThroughWall)
-                    # d3FirstRing.append(d3FirstRingThroughWall)
-
-    if omitStartRows == 1:
-        targetedRingCoordinates = [xFirstRing, d1FirstRing, d2FirstRing, d3FirstRing]
-    elif omitEndRows == 1:
-        targetedRingCoordinates = [xLastRing, d1LastRing, d2LastRing, d3LastRing]
-    else:
-        targetedRingCoordinates = []
-
-    return targetedRingCoordinates
-
-
-def getTargetedRingNodesIds(nodeCount, elementsCountAround, elementsCountAlongTube, elementsCountThroughWall,
-                            omitStartRows, omitEndRows):
-
-    # Create tube nodes
-    lastRingsNodeId = []
-    firstRingsNodeId = []
-    for n2 in range(elementsCountAlongTube + 1):
-        for n3 in range(elementsCountThroughWall + 1):
-            lastRingNodeIdThroughWall = []
-            firstRingNodeIdThroughWall = []
-            for n1 in range(elementsCountAround):
-                if omitEndRows == 1:  # merging to the bifurcation
-                    if n2 == elementsCountAlongTube:
-                        pass
-                    else:
-                        if n2 == elementsCountAlongTube - 1:
-                            lastRingNodeIdThroughWall.append(nodeCount)
-                        nodeCount += 1
-                elif omitStartRows == 1:  # diverging from bifurcation
-                    if n2 == 0:
-                        pass
-                    else:
-                        if n2 == 1:
-                            firstRingNodeIdThroughWall.append(nodeCount)
-                        nodeCount += 1
-                else:
-                    nodeCount += 1
-            if omitEndRows == 1:
-                if n2 == elementsCountAlongTube - 1:
-                    lastRingsNodeId.append(lastRingNodeIdThroughWall)
-            elif omitStartRows == 1:
-                if n2 == 1:
-                    firstRingsNodeId.append(firstRingNodeIdThroughWall)
-
-    if omitStartRows == 1:
-        targetedRingNodeId = firstRingsNodeId
-    elif omitEndRows == 1:
-        targetedRingNodeId = lastRingsNodeId
-    else:
-        targetedRingNodeId = []
-
-    return targetedRingNodeId, nodeCount
-
-
-def createTubeNodes(fm, coordinates, nodeIdentifier, tubeCoordinates, elementsCountAlongTube, elementsCountAround,
-                    elementsCountThroughWall, omitStartRows, omitEndRows):
-
-    cache = fm.createFieldcache()
-
-    nodes = fm.findNodesetByFieldDomainType(Field.DOMAIN_TYPE_NODES)
-    nodetemplate = nodes.createNodetemplate()
-    nodetemplate.defineField(coordinates)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_VALUE, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS1, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS2, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS3, 1)
-
-    # Create tube nodes
-    lastRingsNodeId = []
-    xLastRing = []
-    d1LastRing = []
-    d2LastRing = []
-    d3LastRing = []
-    firstRingsNodeId = []
-    xFirstRing = []
-    d1FirstRing = []
-    d2FirstRing = []
-    d3FirstRing = []
-    for n2 in range(elementsCountAlongTube + 1):
-        for n3 in range(elementsCountThroughWall + 1):
-            lastRingNodeIdThroughWall = []
-            xLastRingThroughWall = []
-            d1LastRingThroughWall = []
-            d2LastRingThroughWall = []
-            d3LastRingThroughWall = []
-            firstRingNodeIdThroughWall = []
-            xFirstRingThroughWall = []
-            d1FirstRingThroughWall = []
-            d2FirstRingThroughWall = []
-            d3FirstRingThroughWall = []
-            for n1 in range(elementsCountAround):
-                n = n2 * elementsCountAround * (elementsCountThroughWall + 1) + n3 * elementsCountAround + n1
-                x = tubeCoordinates[0][n]
-                d1 = tubeCoordinates[1][n]
-                d2 = tubeCoordinates[2][n]
-                d3 = tubeCoordinates[3][n]
-                if omitEndRows == 1:  # merging to the bifurcation
-                    if n2 == elementsCountAlongTube:
-                        pass
-                    else:
-                        node = nodes.createNode(nodeIdentifier, nodetemplate)
-                        cache.setNode(node)
-                        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, x)
-                        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, d1)
-                        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, d2)
-                        # coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS3, 1, d3)
-                        if n2 == elementsCountAlongTube - 1:
-                            lastRingNodeIdThroughWall.append(nodeIdentifier)
-                            xLastRingThroughWall.append(x)
-                            d1LastRingThroughWall.append(d1)
-                            d2LastRingThroughWall.append(d2)
-                            d3LastRingThroughWall.append(d3)
-                        nodeIdentifier += 1
-                elif omitStartRows == 1:  # diverging from bifurcation
-                    if n2 == 0:
-                        pass
-                    else:
-                        node = nodes.createNode(nodeIdentifier, nodetemplate)
-                        cache.setNode(node)
-                        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, x)
-                        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, d1)
-                        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, d2)
-                        # coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS3, 1, d3)
-                        if n2 == 1:
-                            firstRingNodeIdThroughWall.append(nodeIdentifier)
-                            xFirstRingThroughWall.append(x)
-                            d1FirstRingThroughWall.append(d1)
-                            d2FirstRingThroughWall.append(d2)
-                            d3FirstRingThroughWall.append(d3)
-                        nodeIdentifier += 1
-                else:
-                    node = nodes.createNode(nodeIdentifier, nodetemplate)
-                    cache.setNode(node)
-                    coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, x)
-                    coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, d1)
-                    coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, d2)
-                    # coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS3, 1, d3)
-                    nodeIdentifier += 1
-            if omitEndRows == 1:
-                if n2 == elementsCountAlongTube - 1:
-                    lastRingsNodeId.append(lastRingNodeIdThroughWall)
-                    xLastRing.append(xLastRingThroughWall)
-                    d1LastRing.append(d1LastRingThroughWall)
-                    d2LastRing.append(d2LastRingThroughWall)
-                    d3LastRing.append(d3LastRingThroughWall)
-            elif omitStartRows == 1:
-                if n2 == 1:
-                    firstRingsNodeId.append(firstRingNodeIdThroughWall)
-                    xFirstRing.append(xFirstRingThroughWall)
-                    d1FirstRing.append(d1FirstRingThroughWall)
-                    d2FirstRing.append(d2FirstRingThroughWall)
-                    d3FirstRing.append(d3FirstRingThroughWall)
-
-    return nodeIdentifier
-
-
-def find_tube_bifurcation_points_converging(paCentre, pax, pad2, c1Centre, c1x, c1d2, c2Centre, c2x, c2d2):
-    """
-    Gets first ring of coordinates and derivatives between parent pa and
-    children c1, c2, and over the crotch between c1 and c2.
-    :return: rox, rod1, rod2, cox, cod1, cod2, paStartIndex, c1StartIndex, c2StartIndex
-    """
-    paCount = len(pax)
-    c1Count = len(c1x)
-    c2Count = len(c2x)
-    pac1Count, pac2Count, c1c2Count = get_tube_bifurcation_connection_elements_counts([paCount, c1Count, c2Count])
-    # convert to number of nodes, includes both 6-way points
-    pac1NodeCount = pac1Count + 1
-    pac2NodeCount = pac2Count + 1
-    c1c2NodeCount = c1c2Count + 1
-    paStartIndex = 0
-    c1StartIndex = 0
-    c2StartIndex = 0
-    pac1x = [None] * pac1NodeCount
-    pac1d1 = [None] * pac1NodeCount
-    pac1d2 = [None] * pac1NodeCount
-    for n in range(pac1NodeCount):
-        pan = (paStartIndex + n) % paCount
-        c1n = (c1StartIndex + n) % c1Count
-        x1, d1, x2, d2 = c1x[c1n], mult(c1d2[c1n], 2.0), pax[pan], mult(pad2[pan], 2.0)
-        pac1x[n] = interp.interpolateCubicHermite(x1, d1, x2, d2, 0.5)
-        pac1d1[n] = [0.0, 0.0, 0.0]
-        pac1d2[n] = mult(interp.interpolateCubicHermiteDerivative(x1, d1, x2, d2, 0.5), 0.5)
-    paStartIndex2 = paStartIndex + pac1Count
-    c1StartIndex2 = c1StartIndex + pac1Count
-    c2StartIndex2 = c2StartIndex + c1c2Count
-    pac2x = [None] * pac2NodeCount
-    pac2d1 = [None] * pac2NodeCount
-    pac2d2 = [None] * pac2NodeCount
-    for n in range(pac2NodeCount):
-        pan = (paStartIndex2 + n) % paCount
-        c2n = (c2StartIndex2 + n) % c2Count
-        x1, d1, x2, d2 = c2x[c2n], mult(c2d2[c2n], 2.0), pax[pan], mult(pad2[pan], 2.0)
-        pac2x[n] = interp.interpolateCubicHermite(x1, d1, x2, d2, 0.5)
-        pac2d1[n] = [0.0, 0.0, 0.0]
-        pac2d2[n] = mult(interp.interpolateCubicHermiteDerivative(x1, d1, x2, d2, 0.5), 0.5)
-    c1c2x = [None] * c1c2NodeCount
-    c1c2d1 = [None] * c1c2NodeCount
-    c1c2d2 = [None] * c1c2NodeCount
-    for n in range(c1c2NodeCount):
-        c1n = (c1StartIndex2 + n) % c1Count
-        c2n = (c2StartIndex2 - n) % c2Count  # note: reversed
-        x1, d1, x2, d2 = c1x[c1n], mult(c1d2[c1n], 2.0), c2x[c2n], mult(c2d2[c2n], -2.0)
-        c1c2x[n] = interp.interpolateCubicHermite(x1, d1, x2, d2, 0.5)
-        c1c2d1[n] = [0.0, 0.0, 0.0]
-        c1c2d2[n] = mult(interp.interpolateCubicHermiteDerivative(x1, d1, x2, d2, 0.5), 0.5)
-    # get hex triple points
-    hex1, hex1d1, hex1d2 = get_bifurcation_triple_point(
-        c2x[c1StartIndex], mult(c2d2[c2StartIndex], -1.0),
-        c1x[c1StartIndex], mult(c1d2[c1StartIndex], -1.0),
-        pax[paStartIndex], pad2[paStartIndex])
-    hex2, hex2d1, hex2d2 = get_bifurcation_triple_point(
-        c1x[c1StartIndex2], mult(c1d2[c1StartIndex2], -1.0),
-        c2x[c2StartIndex2], mult(c2d2[c2StartIndex2], -1.0),
-        pax[paStartIndex2], pad2[paStartIndex2])
-    # smooth around loops through hex points to get d1
-    loop1x = [hex2] + pac2x[1:-1] + [hex1]
-    loop1d1 = [[-d for d in hex2d2]] + pac2d1[1:-1] + [hex1d1]
-    loop2x = [hex1] + pac1x[1:-1] + [hex2]
-    loop2d1 = [[-d for d in hex1d2]] + pac1d1[1:-1] + [hex2d1]
-    loop1d1 = interp.smoothCubicHermiteDerivativesLine(loop1x, loop1d1, fixStartDirection=True, fixEndDirection=True,
-                                                       magnitudeScalingMode=interp.DerivativeScalingMode.HARMONIC_MEAN)
-    loop2d1 = interp.smoothCubicHermiteDerivativesLine(loop2x, loop2d1, fixStartDirection=True, fixEndDirection=True,
-                                                       magnitudeScalingMode=interp.DerivativeScalingMode.HARMONIC_MEAN)
-    # smooth over "crotch" between c1 and c2
-    crotchx = [hex2] + c1c2x[1:-1] + [hex1]
-    crotchd1 = [add(hex2d1, hex2d2)] + c1c2d1[1:-1] + [[(-hex1d1[c] - hex1d2[c]) for c in range(3)]]
-    crotchd1 = interp.smoothCubicHermiteDerivativesLine(crotchx, crotchd1, fixStartDerivative=True,
-                                                        fixEndDerivative=True,
-                                                        magnitudeScalingMode=interp.DerivativeScalingMode.HARMONIC_MEAN)
-    rox = [hex1] + pac1x[1:-1] + [hex2] + pac2x[1:-1]
-    rod1 = [hex1d1] + loop2d1[1:-1] + [hex2d1] + loop1d1[1:-1]
-    rod2 = [hex1d2] + pac1d2[1:-1] + [hex2d2] + pac2d2[1:-1]
-    cox = crotchx[1:-1]
-    cod1 = crotchd1[1:-1]
-    cod2 = c1c2d2[1:-1]
-    return rox, rod1, rod2, cox, cod1, cod2, paStartIndex, c1StartIndex, c2StartIndex
-
-
-def createTubeBifurcationNodes(fm, coordinates, nodeIdentifier, paCentre, paxList, pad2, c1Centre, c1xList, c1d2,
-                               c2Centre, c2xList, c2d2, elementsCountThroughWall):
-
-    cache = fm.createFieldcache()
-
-    nodes = fm.findNodesetByFieldDomainType(Field.DOMAIN_TYPE_NODES)
-    nodetemplate = nodes.createNodetemplate()
-    nodetemplate.defineField(coordinates)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_VALUE, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS1, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS2, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS3, 1)
-
-    roxList = []
-    rod1List = []
-    rod2List = []
-    coxList = []
-    cod1List = []
-    cod2List = []
-    for n3 in range(elementsCountThroughWall + 1):
-        roxOuter, rod1Outer, rod2Outer, coxOuter, cod1Outer, cod2Outer, paStartIndex, c1StartIndex, c2StartIndex = \
-            find_tube_bifurcation_points_converging(paCentre, paxList[n3], pad2[n3], c1Centre, c1xList[n3], c1d2[n3],
-                                                    c2Centre, c2xList[n3], c2d2[n3])
-        roxList.append(roxOuter)
-        rod1List.append(rod1Outer)
-        rod2List.append(rod2Outer)
-        coxList.append(coxOuter)
-        cod1List.append(cod1Outer)
-        cod2List.append(cod2Outer)
-
-    # Create bifurcation nodes
-    roNodeId = []
-    coNodeId = []
-    for n3 in range(elementsCountThroughWall + 1):
-        coNodeIdThroughWall = []
-        for n in range(len(coxOuter)):
-            node = nodes.createNode(nodeIdentifier, nodetemplate)
-            cache.setNode(node)
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, coxList[n3][n])
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, cod1List[n3][n])
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, cod2List[n3][n])
-            coNodeIdThroughWall.append(nodeIdentifier)
-            nodeIdentifier += 1
-        coNodeId.append(coNodeIdThroughWall)
-        roNodeIdThroughWall = []
-        for n in range(len(roxOuter)):
-            node = nodes.createNode(nodeIdentifier, nodetemplate)
-            cache.setNode(node)
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, roxList[n3][n])
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, rod1List[n3][n])
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, rod2List[n3][n])
-            roNodeIdThroughWall.append(nodeIdentifier)
-            nodeIdentifier += 1
-        roNodeId.append(roNodeIdThroughWall)
-
-    return nodeIdentifier, roNodeId, coNodeId
-
-
-def make_tube_elements(fm, coordinates, startNodeId, elementIdentifier, elementsCountAlongTube, elementsCountAround,
-                       elementsCountThroughWall, useCrossDerivatives, omitStartRows, omitEndRows, startNodes=None,
-                       meshGroups=None):
-
-    mesh = fm.findMeshByDimension(3)
-
-    nodes = fm.findNodesetByFieldDomainType(Field.DOMAIN_TYPE_NODES)
-    nodetemplate = nodes.createNodetemplate()
-    nodetemplate.defineField(coordinates)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_VALUE, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS1, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS2, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS3, 1)
-
-    eftfactory = eftfactory_bicubichermitelinear(mesh, useCrossDerivatives)
-    eft = eftfactory.createEftBasic()
-
-    elementtemplate = mesh.createElementtemplate()
-    elementtemplate.setElementShapeType(Element.SHAPE_TYPE_CUBE)
-    elementtemplate.defineField(coordinates, -1, eft)
-
-    # Create tube elements
-    now = elementsCountAround * (elementsCountThroughWall + 1)
-    for e2 in range(elementsCountAlongTube - 1 if omitStartRows or omitEndRows == 1 else elementsCountAlongTube):
-        for e3 in range(elementsCountThroughWall):
-            for e1 in range(elementsCountAround):
-                element = mesh.createElement(elementIdentifier, elementtemplate)
-                bni11 = e2 * now + e3 * elementsCountAround + e1 + startNodeId
-                bni12 = e2 * now + e3 * elementsCountAround + (e1 + 1) % elementsCountAround + startNodeId
-                bni21 = e2 * now + (e3 + 1) * elementsCountAround + e1 + startNodeId
-                bni22 = e2 * now + (e3 + 1) * elementsCountAround + (e1 + 1) % elementsCountAround + startNodeId
-                nodeIdentifiers = [bni11, bni12, bni11 + now, bni12 + now, bni21, bni22, bni21 + now, bni22 + now]
-                # print('nodeIdentifiers', nodeIdentifiers)
-                result = element.setNodesByIdentifier(eft, nodeIdentifiers)
-                for meshGroup in meshGroups:
-                    meshGroup.addElement(element)
-                elementIdentifier += 1
-
-    return elementIdentifier
-
-
-def make_tube_bifurcation_elements(fm, coordinates, elementIdentifier, elementsCountThroughWall, paNodeId, c1NodeId,
-                                   c2NodeId, roNodeId, coNodeId, meshGroups=None):
-
-    elementsCountAround = len(paNodeId[0])
-    elementsCountAroundRightTube = len(c1NodeId[0])
-    elementsCountAroundLeftTube = len(c2NodeId[0])
-    pac1Count, pac2Count, c1c2Count = \
-        get_tube_bifurcation_connection_elements_counts(
-            [elementsCountAround, elementsCountAroundRightTube, elementsCountAroundLeftTube])
-
-    mesh = fm.findMeshByDimension(3)
-    eftfactory = eftfactory_bicubichermitelinear(mesh, None)
-    eftStd = eftfactory.createEftBasic()
-
-    elementtemplateStd = mesh.createElementtemplate()
-    elementtemplateStd.setElementShapeType(Element.SHAPE_TYPE_CUBE)
-    elementtemplateStd.defineField(coordinates, -1, eftStd)
-
-    elementtemplateMod = mesh.createElementtemplate()
-    elementtemplateMod.setElementShapeType(Element.SHAPE_TYPE_CUBE)
-
-    # Right tube part
-    for e3 in range(elementsCountThroughWall):
-        for e1 in range(elementsCountAroundRightTube):
-            eft = eftStd
-            elementtemplate = elementtemplateStd
-            scalefactors = None
-            if e1 < elementsCountAround // 2:
-                bni1 = c1NodeId[e3][e1]
-                bni2 = c1NodeId[e3][(e1 + 1) % elementsCountAroundRightTube]
-                bni3 = roNodeId[e3][e1]
-                bni4 = roNodeId[e3][(e1 + 1) % elementsCountAroundRightTube]
-                bni5 = bni1 + elementsCountAroundRightTube
-                bni6 = bni5 + 1
-                bni7 = bni3 + len(roNodeId[0]) + len(coNodeId[0])
-                bni8 = bni7 + 1
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-            elif e1 == elementsCountAround // 2:
-                bni1 = c1NodeId[e3][e1]
-                bni2 = c1NodeId[e3][(e1 + 1) % elementsCountAroundRightTube]
-                bni3 = roNodeId[e3][e1]
-                bni4 = coNodeId[e3][e1 - elementsCountAround // 2]
-                bni5 = bni1 + elementsCountAroundRightTube
-                bni6 = bni5 + 1
-                bni7 = bni3 + len(roNodeId[0]) + len(coNodeId[0])
-                bni8 = bni4 + len(roNodeId[0]) + len(coNodeId[0])
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-            else:
-                bni1 = c1NodeId[e3][e1]
-                bni2 = c1NodeId[e3][(e1 + 1) % elementsCountAroundRightTube]
-                bni3 = coNodeId[e3][e1 - elementsCountAround // 2 - 1]
-                bni5 = bni1 + elementsCountAroundRightTube
-                if e1 == elementsCountAroundRightTube - 1:
-                    bni4 = roNodeId[e3][0]
-                    bni6 = bni5 - elementsCountAroundRightTube + 1
-                else:
-                    bni4 = bni3 + 1
-                    bni6 = bni5 + 1
-                bni7 = bni3 + len(roNodeId[0]) + len(coNodeId[0])
-                bni8 = bni4 + len(roNodeId[0]) + len(coNodeId[0])
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-            if e1 in (0, pac1Count - 1, pac1Count, elementsCountAroundRightTube - 1):
-                eft = eftfactory.createEftBasic()
-                if e1 == 0:
-                    scalefactors = [-1.0]
-                    setEftScaleFactorIds(eft, [1], [])
-                    remapEftNodeValueLabel(eft, [3, 7], Node.VALUE_LABEL_D_DS1,
-                                           [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                    remapEftNodeValueLabel(eft, [3, 7], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS1, [1])])
-                elif e1 == pac1Count - 1:
-                    remapEftNodeValueLabel(eft, [4, 8], Node.VALUE_LABEL_D_DS2,
-                                           [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                elif e1 == pac1Count:
-                    scalefactors = [-1.0]
-                    setEftScaleFactorIds(eft, [1], [])
-                    remapEftNodeValueLabel(eft, [3, 7], Node.VALUE_LABEL_D_DS2,
-                                           [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                    remapEftNodeValueLabel(eft, [3, 7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [1])])
-                elif e1 == elementsCountAroundRightTube - 1:
-                    scalefactors = [-1.0]
-                    setEftScaleFactorIds(eft, [1], [])
-                    remapEftNodeValueLabel(eft, [4, 8], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    remapEftNodeValueLabel(eft, [4, 8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [])])
-                elementtemplateMod.defineField(coordinates, -1, eft)
-                elementtemplate = elementtemplateMod
-            element = mesh.createElement(elementIdentifier, elementtemplate)
-            result2 = element.setNodesByIdentifier(eft, nodeIdentifiers)
-            if scalefactors:
-                result3 = element.setScaleFactors(eft, scalefactors)
-            else:
-                result3 = '-'
-            elementIdentifier += 1
-            for meshGroup in meshGroups:
-                if meshGroups.index(meshGroup) == 1:
-                    meshGroup.addElement(element)
-                elif meshGroups.index(meshGroup) == 3:
-                    meshGroup.addElement(element)
-
-    # Left tube part
-    for e3 in range(elementsCountThroughWall):
-        for e1 in range(elementsCountAroundLeftTube):
-            eft = eftStd
-            elementtemplate = elementtemplateStd
-            scalefactors = None
-            if e1 < c1c2Count:
-                bni1 = c2NodeId[e3][e1]
-                bni2 = c2NodeId[e3][(e1 + 1) % elementsCountAroundLeftTube]
-                if e1 == 0:
-                    bni3 = roNodeId[e3][e1]
-                    bni4 = coNodeId[e3][-1 - e1]
-                else:
-                    bni3 = coNodeId[e3][-e1]
-                    if e1 == c1c2Count - 1:
-                        bni4 = roNodeId[e3][0] + pac1Count
-                    else:
-                        bni4 = bni3 - 1
-                bni5 = bni1 + elementsCountAroundLeftTube
-                bni6 = bni2 + elementsCountAroundLeftTube
-                bni7 = bni3 + len(roNodeId[0]) + len(coNodeId[0])
-                bni8 = bni4 + len(roNodeId[0]) + len(coNodeId[0])
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-            else:
-                bni1 = c2NodeId[e3][e1]
-                bni2 = c2NodeId[e3][(e1 + 1) % elementsCountAroundLeftTube]
-                bni3 = roNodeId[e3][e1 - c1c2Count + elementsCountAround // 2]
-                bni4 = roNodeId[e3][(e1 - c1c2Count + elementsCountAround // 2 + 1) % elementsCountAround]
-                bni5 = bni1 + elementsCountAroundLeftTube
-                bni6 = bni2 + elementsCountAroundLeftTube
-                bni7 = bni3 + len(roNodeId[0]) + len(coNodeId[0])
-                bni8 = bni4 + len(roNodeId[0]) + len(coNodeId[0])
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-            if e1 <= c1c2Count:
-                eft = eftfactory.createEftBasic()
-                scalefactors = [-1.0]
-                setEftScaleFactorIds(eft, [1], [])
-                if e1 == 0:
-                    remapEftNodeValueLabel(eft, [3, 7], Node.VALUE_LABEL_D_DS2,
-                                           [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                    remapEftNodeValueLabel(eft, [3, 7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [1])])
-                    scaleEftNodeValueLabels(eft, [4, 8], [Node.VALUE_LABEL_D_DS1, Node.VALUE_LABEL_D_DS2], [1])
-                elif e1 < (c1c2Count - 1):
-                    scaleEftNodeValueLabels(eft, [3, 4, 7, 8], [Node.VALUE_LABEL_D_DS1, Node.VALUE_LABEL_D_DS2], [1])
-                elif e1 == (c1c2Count - 1):
-                    scaleEftNodeValueLabels(eft, [3, 7], [Node.VALUE_LABEL_D_DS1, Node.VALUE_LABEL_D_DS2], [1])
-                    remapEftNodeValueLabel(eft, [4, 8], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    remapEftNodeValueLabel(eft, [4, 8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [])])
-                elif e1 == c1c2Count:
-                    remapEftNodeValueLabel(eft, [3, 7], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    remapEftNodeValueLabel(eft, [3, 7], Node.VALUE_LABEL_D_DS1,
-                                           [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                elementtemplateMod.defineField(coordinates, -1, eft)
-                elementtemplate = elementtemplateMod
-            elif e1 == elementsCountAroundLeftTube - 1:
-                eft = eftfactory.createEftBasic()
-                remapEftNodeValueLabel(eft, [4, 8], Node.VALUE_LABEL_D_DS2,
-                                       [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                elementtemplateMod.defineField(coordinates, -1, eft)
-                elementtemplate = elementtemplateMod
-            element = mesh.createElement(elementIdentifier, elementtemplate)
-            result2 = element.setNodesByIdentifier(eft, nodeIdentifiers)
-            if scalefactors:
-                result3 = element.setScaleFactors(eft, scalefactors)
-            else:
-                result3 = '-'
-            elementIdentifier += 1
-            for meshGroup in meshGroups:
-                if meshGroups.index(meshGroup) == 2:
-                    meshGroup.addElement(element)
-                elif meshGroups.index(meshGroup) == 3:
-                    meshGroup.addElement(element)
-
-    # parent part
-    for e3 in range(elementsCountThroughWall):
-        for e1 in range(elementsCountAround):
-            eft = eftStd
-            elementtemplate = elementtemplateStd
-            scalefactors = None
-            bni1 = roNodeId[e3][e1]
-            bni2 = roNodeId[e3][(e1 + 1) % elementsCountAround]
-            bni3 = paNodeId[e3][e1]
-            bni4 = paNodeId[e3][(e1 + 1) % elementsCountAround]
-            bni5 = bni1 + len(roNodeId[0]) + len(coNodeId[0])
-            bni6 = bni2 + len(roNodeId[0]) + len(coNodeId[0])
-            bni7 = bni3 + elementsCountAround
-            bni8 = bni4 + elementsCountAround
-            nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-            # print('nodeIdentifiers', nodeIdentifiers)
-            if e1 in (0, pac1Count):
-                eft = eftfactory.createEftBasic()
-                if e1 in (0, pac1Count):
-                    remapEftNodeValueLabel(eft, [1, 5], Node.VALUE_LABEL_D_DS1,
-                                           [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                elementtemplateMod.defineField(coordinates, -1, eft)
-                elementtemplate = elementtemplateMod
-            element = mesh.createElement(elementIdentifier, elementtemplate)
-            result2 = element.setNodesByIdentifier(eft, nodeIdentifiers)
-            if scalefactors:
-                result3 = element.setScaleFactors(eft, scalefactors)
-            else:
-                result3 = '-'
-            elementIdentifier += 1
-            for meshGroup in meshGroups:
-                if meshGroups.index(meshGroup) == 0:
-                    meshGroup.addElement(element)
-                elif meshGroups.index(meshGroup) == 3:
-                    meshGroup.addElement(element)
-
-    return elementIdentifier
-
-
-def getDoubleTubeNodes(cx_tube_group, elementsCountAlong, elementsCountAround, elementsCountAroundRightTube,
-                       elementsCountAroundLeftTube, elementsCountThroughWall, wallThickness):
-    """
-    :return: the coordinates and derivatives of the outer tube, inner right and left tubes and the septum nodes
-    """
-    elementsCountAcross = elementsCountAroundRightTube - elementsCountAround // 2
-
-    # Get tube right and left paths
-    distance = wallThickness / 2
-    xrList = []
-    d2rList = []
-    d3rList = []
-    xlList = []
-    d2lList = []
-    d3lList = []
-    for n in range(len(cx_tube_group[0])):
-        x = cx_tube_group[0][n]
-        v2r = normalize(cx_tube_group[2][n])
-        v3r = normalize(cx_tube_group[4][n])
-        d2Mag = magnitude(cx_tube_group[2][n])
-        r2 = (d2Mag - distance - wallThickness) / 2
-        # d3Mag = magnitude(cx_tube_group[4][n])
-        # r3 = (d3Mag - wallThickness)
-        v_trans = set_magnitude(v2r, distance + r2)
-
-        x_right = [x[c] + v_trans[c] for c in range(3)]
-        xrList.append(x_right)
-        d2rList.append(set_magnitude(v2r, r2))
-        d3rList.append(set_magnitude(v3r, r2))
-
-        x_left = [x[c] - v_trans[c] for c in range(3)]
-        v2l = [-v2r[c] for c in range(3)]
-        v3l = [-v3r[c] for c in range(3)]
-        xlList.append(x_left)
-        d2lList.append(set_magnitude(v2l, r2))
-        d3lList.append(set_magnitude(v3l, r2))
-    cx_tube_group_right = [xrList, cx_tube_group[1], d2rList, [], d3rList]
-    cx_tube_group_left = [xlList, cx_tube_group[1], d2lList, [], d3lList]
-
-    # Get right inner tube nodes
-    rightStartRadians = -math.pi * (elementsCountAround / (2 * elementsCountAroundRightTube))
-    tubeInnerRightCoordinates = findNodesAlongTube2D(cx_tube_group_right, elementsCountAroundRightTube,
-                                                     elementsCountAlong, startRadian=rightStartRadians)
-
-    # Get left inner tube nodes
-    leftStartRadians = math.pi - math.pi * (elementsCountAcross / elementsCountAroundLeftTube)
-    tubeInnerLeftCoordinates = findNodesAlongTube2D(cx_tube_group_left, elementsCountAroundLeftTube,
-                                                    elementsCountAlong, startRadian=leftStartRadians)
-
-    # Get outer nodes along tube
-    startRadian = -math.pi / 2
-    xOuter, d1Outer, d2Outer = findNodesAlongTube2D(cx_tube_group, elementsCountAround, elementsCountAlong, startRadian)
-    outerCoordinates = [xOuter, d1Outer, d2Outer]
-
-    # Get coordinates across tube septum, between two inner canals
-    xAcrossSeptum = []
-    d1AcrossSeptum = []
-    for n in range(elementsCountAlong + 1):
-        oa = 0
-        ob = elementsCountAround // 2
-        v1 = xOuter[n][ob]
-        v2 = xOuter[n][oa]
-        v3 = [v1[c] / 2 + v2[c] / 2 for c in range(3)]
-        v1v2 = [v2[c] - v1[c] for c in range(3)]
-        nx = [xOuter[n][ob], v3, xOuter[n][oa]]
-        nd1 = [[d / elementsCountAcross for d in v1v2], [d / elementsCountAcross for d in v1v2],
-               [d / elementsCountAcross for d in v1v2]]
-        px, pd1, pe, pxi = interp.sampleCubicHermiteCurves(nx, nd1, elementsCountAcross)[0:4]
-        xAcrossSeptum.append(px)
-        d1AcrossSeptum.append(pd1)
-
-    # Find d2 across tube septum
-    d2Raw = []
-    for n2 in range(elementsCountAcross + 1):
-        xAlongSeptum = []
-        d2AlongSeptum = []
-        for n1 in range(elementsCountAlong):
-            v1 = xAcrossSeptum[n1][n2]
-            v2 = xAcrossSeptum[n1 + 1][n2]
-            d2 = findDerivativeBetweenPoints(v1, v2)
-            xAlongSeptum.append(v1)
-            d2AlongSeptum.append(d2)
-        xAlongSeptum.append(xAcrossSeptum[-1][n2])
-        d2AlongSeptum.append(d2)
-        d2Smoothed = interp.smoothCubicHermiteDerivativesLine(xAlongSeptum, d2AlongSeptum)
-        d2Raw.append(d2Smoothed)
-
-    # Rearrange d2
-    d2AcrossSeptum = []
-    for n2 in range(elementsCountAlong + 1):
-        d2Across = []
-        for n1 in range(elementsCountAcross + 1):
-            d2 = d2Raw[n1][n2]
-            d2Across.append(d2)
-        d2AcrossSeptum.append(d2Across)
-
-    septumCoordinates = [xAcrossSeptum, d1AcrossSeptum, d2AcrossSeptum]
-
-    # Find d3 for right inner canal nodes
-    d3InnerRight = []
-    for n2 in range(0, elementsCountAlong + 1):
-        d3Raw = []
-        for n1 in range(elementsCountAroundRightTube):
-            v1 = tubeInnerRightCoordinates[0][n2][n1]
-            if n1 <= elementsCountAround // 2:
-                v2 = outerCoordinates[0][n2][n1]
-            else:
-                v2 = septumCoordinates[0][n2][n1 - elementsCountAround // 2]
-            d3 = findDerivativeBetweenPoints(v1, v2)
-            d3Raw.append(d3)
-        d3InnerRight.append(d3Raw)
-    tubeInnerRightCoordinates.append(d3InnerRight)
-
-    # Find d3 for Left inner canal nodes
-    d3InnerLeft = []
-    for n2 in range(0, elementsCountAlong + 1):
-        d3Raw = []
-        for n1 in range(elementsCountAroundLeftTube):
-            v1 = tubeInnerLeftCoordinates[0][n2][n1]
-            if n1 == 0:
-                v2 = outerCoordinates[0][n2][n1]
-            elif 0 < n1 < elementsCountAcross:
-                v2 = septumCoordinates[0][n2][elementsCountAcross - n1]
-            elif n1 == elementsCountAcross:
-                v2 = outerCoordinates[0][n2][elementsCountAround // 2]
-            else:
-                v2 = outerCoordinates[0][n2][n1]
-            d3 = findDerivativeBetweenPoints(v1, v2)
-            d3Raw.append(d3)
-        d3InnerLeft.append(d3Raw)
-    tubeInnerLeftCoordinates.append(d3InnerLeft)
-
-    # Find d3 for outer nodes
-    d3Outer = []
-    for n2 in range(0, elementsCountAlong + 1):
-        d3Raw = []
-        for n1 in range(elementsCountAround):
-            if n1 == 0:
-                d1 = septumCoordinates[1][n2][n1]
-                d3Raw.append(d1)
-            elif 0 < n1 < elementsCountAround // 2:
-                v1 = tubeInnerRightCoordinates[0][n2][n1]
-                v2 = outerCoordinates[0][n2][n1]
-                v1v2 = findDerivativeBetweenPoints(v1, v2)
-                d3Raw.append(v1v2)
-            elif n1 == elementsCountAround // 2:
-                d = [-d1[c] for c in range(3)]
-                d3Raw.append(d)
-            else:
-                v1 = tubeInnerLeftCoordinates[0][n2][elementsCountAcross + n1 - elementsCountAround // 2]
-                v2 = outerCoordinates[0][n2][n1]
-                v1v2 = findDerivativeBetweenPoints(v1, v2)
-                d3Raw.append(v1v2)
-            d3Outer.append(d3Raw)
-
-    outerCoordinates = [outerCoordinates[0], outerCoordinates[1], outerCoordinates[2], d3Outer]
-
-    # Get all nodes through wall for right and left inner parts
-    rightInnerTubethroughWall, leftInnerTubethroughWall = \
-        getInnerDoubleTubeCoordinates(tubeInnerRightCoordinates, tubeInnerLeftCoordinates, outerCoordinates,
-                                      septumCoordinates, elementsCountAlong, elementsCountAround,
-                                      elementsCountAroundRightTube, elementsCountAroundLeftTube,
-                                      elementsCountThroughWall)
-
-    return rightInnerTubethroughWall, leftInnerTubethroughWall, outerCoordinates, septumCoordinates
-
-
-def findNodesAlongTube2D(sx_group, elementsCountAround, elementsCountAlongTube, startRadian):
-    """
-    Gets the central path nodes and return the coordinates and derivatives of a 2D tube.
-    :return: tube2dCoordinates; 2D tube coordinates and derivatives.
-    """
-    # Create ellipses along tube around the central path
-    xEllipsesAlong = []
-    d1EllipsesAlong = []
-    for n in range(len(sx_group[0])):
-        px, pd1 = createEllipsePoints(sx_group[0][n], 2 * math.pi, sx_group[2][n], sx_group[4][n], elementsCountAround,
-                                      startRadians=startRadian)
-        xEllipsesAlong.append(px)
-        d1EllipsesAlong.append(pd1)
-
-    # Find d2
-    d2Raw = []
-    for n1 in range(elementsCountAround):
-        xAlong = []
-        d2Along = []
-        for n2 in range(len(xEllipsesAlong) - 1):
-            v1 = xEllipsesAlong[n2][n1]
-            v2 = xEllipsesAlong[n2 + 1][n1]
-            d2 = findDerivativeBetweenPoints(v1, v2)
-            xAlong.append(v1)
-            d2Along.append(d2)
-        xAlong.append(xEllipsesAlong[-1][n1])
-        d2Along.append(d2)
-        d2Smoothed = interp.smoothCubicHermiteDerivativesLine(xAlong, d2Along)
-        d2Raw.append(d2Smoothed)
-
-    # Rearrange d2
-    d2EllipsesAlong = []
-    for n2 in range(len(xEllipsesAlong)):
-        d2Around = []
-        for n1 in range(elementsCountAround):
-            d2 = d2Raw[n1][n2]
-            d2Around.append(d2)
-        d2EllipsesAlong.append(d2Around)
-
-    # Spread out elements along tube
-    xRaw = []
-    d2Raw = []
-    for n1 in range(elementsCountAround):
-        xAlong = []
-        d2Along = []
-        for n2 in range(len(xEllipsesAlong)):
-            xAlong.append(xEllipsesAlong[n2][n1])
-            d2Along.append(d2EllipsesAlong[n2][n1])
-        xSampledAlong, d2SampledAlong = interp.sampleCubicHermiteCurves(xAlong, d2Along, elementsCountAlongTube,
-                                                                        arcLengthDerivatives=False)[0:2]
-        d2Smoothed = interp.smoothCubicHermiteDerivativesLine(xSampledAlong, d2SampledAlong)
-        xRaw.append(xSampledAlong)
-        d2Raw.append(d2SampledAlong)
-
-    # Rearrange x and d2
-    xSampledTube = []
-    d1SampledTube = []
-    d2SampledTube = []
-    for n2 in range(elementsCountAlongTube + 1):
-        xAround = []
-        d1Around = []
-        d2Around = []
-        for n1 in range(elementsCountAround):
-            x = xRaw[n1][n2]
-            d2 = d2Raw[n1][n2]
-            xAround.append(x)
-            d2Around.append(d2)
-            # Calculate d1
-            v1 = xRaw[n1][n2]
-            v2 = xRaw[n1 + 1 if n1 < elementsCountAround - 1 else 0][n2]
-            d1 = findDerivativeBetweenPoints(v1, v2)
-            d1Around.append(d1)
-        d1Smoothed = interp.smoothCubicHermiteDerivativesLoop(xAround, d1Around)
-        xSampledTube.append(xAround)
-        d1SampledTube.append(d1Smoothed)
-        d2SampledTube.append(d2Around)
-
-    tube2dCoordinates = [xSampledTube, d1SampledTube, d2SampledTube]
-    return tube2dCoordinates
-
-
-def getInnerDoubleTubeCoordinates(innerRightCoordinates, innerLeftCoordinates, outerCoordinates, septumCoordinates,
-                                  elementsCountAlong, elementsCountAround, elementsCountAroundRightTube,
-                                  elementsCountAroundLeftTube, elementsCountThroughWall):
-    """
-    Gets the outer layer, inner left and right layers and septum coordinates.
-    :return: the coordinates and derivatives of all nodes through wall for left and right inner tubes.
-    """
-    elementsCountAcross = elementsCountAroundRightTube - elementsCountAround // 2
-
-    # Get nodes through wall for right and left inner tubes
-    xRawRight = []
-    d3RawRight = []
-    xRawLeft = []
-    d3RawLeft = []
-    for n2 in range(elementsCountAlong + 1):
-        xtRight = []
-        d3tRight = []
-        for n1 in range(elementsCountAroundRightTube):
-            xAlong = []
-            d3Along = []
-            v1 = innerRightCoordinates[0][n2][n1]
-            if n1 <= elementsCountAround // 2:
-                v2 = outerCoordinates[0][n2][n1]
-            else:
-                v2 = septumCoordinates[0][n2][n1 - elementsCountAround // 2]
-            v1v2 = findDerivativeBetweenPoints(v1, v2)
-            xAlong.append(v1)
-            xAlong.append(v2)
-            d3Along.append(v1v2)
-            d3Along.append(v1v2)
-            xSampledWall, d3SampledWall = interp.sampleCubicHermiteCurves(xAlong, d3Along, elementsCountThroughWall,
-                                                                          arcLengthDerivatives=True)[0:2]
-            xtRight.append(xSampledWall)
-            d3tRight.append(d3SampledWall)
-        xRawRight.append(xtRight)
-        d3RawRight.append(d3tRight)
-        xtLeft = []
-        d3tLeft = []
-        for n1 in range(elementsCountAroundLeftTube):
-            xAlongLeft = []
-            d3AlongLeft = []
-            v1 = innerLeftCoordinates[0][n2][n1]
-            if n1 == 0:
-                v2 = outerCoordinates[0][n2][n1]
-            elif 0 < n1 < elementsCountAcross:
-                v2 = septumCoordinates[0][n2][elementsCountAcross - n1]
-            else:
-                v2 = outerCoordinates[0][n2][n1 - elementsCountAcross + elementsCountAround // 2]
-            v1v2 = findDerivativeBetweenPoints(v1, v2)
-            xAlongLeft.append(v1)
-            xAlongLeft.append(v2)
-            d3AlongLeft.append(v1v2)
-            d3AlongLeft.append(v1v2)
-            xSampledWall, d3SampledWall = interp.sampleCubicHermiteCurves(xAlongLeft, d3AlongLeft,
-                                                                          elementsCountThroughWall,
-                                                                          arcLengthDerivatives=True)[0:2]
-            xtLeft.append(xSampledWall)
-            d3tLeft.append(d3SampledWall)
-        xRawLeft.append(xtLeft)
-        d3RawLeft.append(d3tLeft)
-
-    # Rearrange the nodes
-    xWallRight = []
-    xWallLeft = []
-    for n2 in range(elementsCountAlong + 1):
-        xAroundRight = []
-        xAroundLeft = []
-        for n3 in range(elementsCountThroughWall + 1):
-            v1ListRight = []
-            v1ListLeft = []
-            for n1 in range(elementsCountAroundRightTube):
-                v1 = xRawRight[n2][n1][n3]
-                v1ListRight.append(v1)
-            xAroundRight.append(v1ListRight)
-            for n1 in range(elementsCountAroundLeftTube):
-                v1 = xRawLeft[n2][n1][n3]
-                v1ListLeft.append(v1)
-            xAroundLeft.append(v1ListLeft)
-        xWallRight.append(xAroundRight)
-        xWallLeft.append(xAroundLeft)
-
-    # Find d1 around tube for all nodes
-    d1WallRight = []
-    d1WallLeft = []
-    for n2 in range(elementsCountAlong + 1):
-        xRawRight = []
-        d1RawRight = []
-        xRawLeft = []
-        d1RawLeft = []
-        for n3 in range(elementsCountThroughWall + 1):
-            xAroundRight = []
-            d1AroundRight = []
-            xAroundLeft = []
-            d1AroundLeft = []
-            for n1 in range(elementsCountAroundRightTube):
-                v1 = xWallRight[n2][n3][n1]
-                v2 = xWallRight[n2][n3][(n1 + 1) % elementsCountAroundRightTube]
-                d1 = findDerivativeBetweenPoints(v1, v2)
-                xAroundRight.append(v1)
-                d1AroundRight.append(d1)
-            d1SmoothedRight = interp.smoothCubicHermiteDerivativesLoop(xAroundRight, d1AroundRight)
-            xRawRight.append(d1AroundRight)
-            d1RawRight.append(d1SmoothedRight)
-            for n1 in range(elementsCountAroundRightTube):
-                v1 = xWallLeft[n2][n3][n1]
-                v2 = xWallLeft[n2][n3][(n1 + 1) % elementsCountAroundLeftTube]
-                d1 = findDerivativeBetweenPoints(v1, v2)
-                xAroundLeft.append(v1)
-                d1AroundLeft.append(d1)
-            d1SmoothedLeft = interp.smoothCubicHermiteDerivativesLoop(xAroundLeft, d1AroundLeft)
-            xRawLeft.append(xAroundLeft)
-            d1RawLeft.append(d1SmoothedLeft)
-        d1WallRight.append(d1RawRight)
-        d1WallLeft.append(d1RawLeft)
-
-    # Find d2 along tube for nodes through wall
-    d2NewRight = []
-    d2NewLeft = []
-    for n1 in range(elementsCountAroundRightTube):
-        d2RawRight = []
-        d2RawLeft = []
-        for n3 in range(elementsCountThroughWall + 1):
-            xALongRight = []
-            d2AlongRight = []
-            xALongLeft = []
-            d2AlongLeft = []
-            for n2 in range(elementsCountAlong):
-                v1 = xWallRight[n2][n3][n1]
-                v2 = xWallRight[n2 + 1][n3][n1]
-                d2 = findDerivativeBetweenPoints(v1, v2)
-                xALongRight.append(v1)
-                d2AlongRight.append(d2)
-            xALongRight.append(v2)
-            d2AlongRight.append(d2)
-            d2SmoothedRight = interp.smoothCubicHermiteDerivativesLine(xALongRight, d2AlongRight)
-            d2RawRight.append(d2SmoothedRight)
-            for n2 in range(elementsCountAlong):
-                v1 = xWallLeft[n2][n3][n1]
-                v2 = xWallLeft[n2 + 1][n3][n1]
-                d2 = findDerivativeBetweenPoints(v1, v2)
-                xALongLeft.append(v1)
-                d2AlongLeft.append(d2)
-            xALongLeft.append(v2)
-            d2AlongLeft.append(d2)
-            d2SmoothedLeft = interp.smoothCubicHermiteDerivativesLine(xALongLeft, d2AlongLeft)
-            d2RawLeft.append(d2SmoothedLeft)
-        d2NewRight.append(d2RawRight)
-        d2NewLeft.append(d2RawLeft)
-
-    # Rearrange d2
-    d2WallRight = []
-    d2WallLeft = []
-    for n2 in range(elementsCountAlong + 1):
-        d2tRight = []
-        d2tLeft = []
-        for n3 in range(elementsCountThroughWall + 1):
-            d2AroundRight = []
-            d2AroundLeft = []
-            for n1 in range(elementsCountAroundRightTube):
-                d2 = d2NewRight[n1][n3][n2]
-                d2AroundRight.append(d2)
-            d2tRight.append(d2AroundRight)
-            for n1 in range(elementsCountAroundLeftTube):
-                d2 = d2NewLeft[n1][n3][n2]
-                d2AroundLeft.append(d2)
-            d2tLeft.append(d2AroundLeft)
-        d2WallRight.append(d2tRight)
-        d2WallLeft.append(d2tLeft)
-
-    rightInnerTubethroughWall = [xWallRight, d1WallRight, d2WallRight]
-    leftInnerTubethroughWall = [xWallLeft, d1WallLeft, d2WallLeft]
-
-    return rightInnerTubethroughWall, leftInnerTubethroughWall
-
-
-def createDoubleTubeNodes(fm, nodeIdentifier, xInnerRigh, xInnerLeft, xOuter, xAcross, elementsCountAlong,
-                          elementsCountAround, elementsCountAcross, elementsCountThroughWall, omitStartRows,
-                          omitEndRows):
-
-    """
-    Gets coordinates and derivatives of inner canals and outer of a tube and generate the nodes.
-    :return: nodeIdentifier, tricNodeId, tlicNodeId, toNodeId, tsNodeId
-    """
-    cache = fm.createFieldcache()
-    coordinates = findOrCreateFieldCoordinates(fm)
-
-    nodes = fm.findNodesetByFieldDomainType(Field.DOMAIN_TYPE_NODES)
-    nodetemplate = nodes.createNodetemplate()
-    nodetemplate.defineField(coordinates)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_VALUE, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS1, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS2, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS3, 1)
-
-    count = elementsCountAround // 2 + elementsCountAcross
-    elementsCountAroundRightHorn = count
-    elementsCountAroundLeftHorn = count
-
-    tricNodeId = []  # tube right inner canal node ids
-    tlicNodeId = []  # tube left inner canal node ids
-    toNodeId = []  # tube outer node ids
-    tsNodeId = []   # tube septum node ids
-    for n2 in range(1 if omitStartRows == 1 else 0, elementsCountAlong if omitEndRows == 1 else elementsCountAlong + 1):
-        for n3 in range(elementsCountThroughWall):
-            for n1 in range(elementsCountAroundRightHorn):
-                node = nodes.createNode(nodeIdentifier, nodetemplate)
-                cache.setNode(node)
-                coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, xInnerRigh[0][n2][n3][n1])
-                coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, xInnerRigh[1][n2][n3][n1])
-                coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, xInnerRigh[2][n2][n3][n1])
-                tricNodeId.append(nodeIdentifier)
-                nodeIdentifier += 1
-        for n3 in range(elementsCountThroughWall):
-            for n1 in range(elementsCountAroundLeftHorn):
-                node = nodes.createNode(nodeIdentifier, nodetemplate)
-                cache.setNode(node)
-                coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, xInnerLeft[0][n2][n3][n1])
-                coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, xInnerLeft[1][n2][n3][n1])
-                coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, xInnerLeft[2][n2][n3][n1])
-                tlicNodeId.append(nodeIdentifier)
-                nodeIdentifier += 1
-        for n1 in range(1, elementsCountAcross):
-            node = nodes.createNode(nodeIdentifier, nodetemplate)
-            cache.setNode(node)
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, xAcross[0][n2][n1])
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, xAcross[1][n2][n1])
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, xAcross[2][n2][n1])
-            tsNodeId.append(nodeIdentifier)
-            nodeIdentifier += 1
-        for n1 in range(elementsCountAround):
-            node = nodes.createNode(nodeIdentifier, nodetemplate)
-            cache.setNode(node)
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, xOuter[0][n2][n1])
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, xOuter[1][n2][n1])
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, xOuter[2][n2][n1])
-            if n1 in (0, elementsCountAround // 2):
-                coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS3, 1, xOuter[3][n2][n1])
-            toNodeId.append(nodeIdentifier)
-            nodeIdentifier += 1
-
-    return nodeIdentifier, tricNodeId, tlicNodeId, toNodeId, tsNodeId
-
-
-def createDoubleTubeBifurcationNodes(fm, nodeIdentifier, rox, rod1, rod2, rod3, cox, cod1, cod2, cod3,
-                                     innerBifurcationRight, innerBifurcationLeft, elementsCountAroundRightHorn,
-                                     elementsCountAroundLeftHorn, elementsCountThroughWall):
-    """
-    Gets coordinates and derivatives of bifurcation inner and outer and generate the nodes.
-    :return: nodeIdentifier, roNodeId, coNodeId, birNodeId, bilNodeId
-    """
-    cache = fm.createFieldcache()
-    coordinates = findOrCreateFieldCoordinates(fm)
-
-    nodes = fm.findNodesetByFieldDomainType(Field.DOMAIN_TYPE_NODES)
-    nodetemplate = nodes.createNodetemplate()
-    nodetemplate.defineField(coordinates)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_VALUE, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS1, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS2, 1)
-    nodetemplate.setValueNumberOfVersions(coordinates, -1, Node.VALUE_LABEL_D_DS3, 1)
-
-    # Create extra right inner nodes in bifurcation
-    birNodeId = []  # bifurcation right inner ring node ids
-    for n2 in range(2):
-        for n3 in range(elementsCountThroughWall):
-            for n1 in range(elementsCountAroundRightHorn):
-                if n2 == 1:
-                    node = nodes.createNode(nodeIdentifier, nodetemplate)
-                    cache.setNode(node)
-                    coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, innerBifurcationRight[0][n2][n3][n1])
-                    coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, innerBifurcationRight[1][n2][n3][n1])
-                    coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, innerBifurcationRight[2][n2][n3][n1])
-                    # coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS3, 1, d3SampledBifurcationRight[n1])
-                    birNodeId.append(nodeIdentifier)
-                    nodeIdentifier += 1
-
-    # Create extra left inner nodes in bifurcation
-    bilNodeId = []  # bifurcation left inner ring node ids
-    for n2 in range(2):
-        for n3 in range(elementsCountThroughWall):
-            for n1 in range(elementsCountAroundLeftHorn):
-                if n2 == 1:
-                    node = nodes.createNode(nodeIdentifier, nodetemplate)
-                    cache.setNode(node)
-                    coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, innerBifurcationLeft[0][n2][n3][n1])
-                    coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, innerBifurcationLeft[1][n2][n3][n1])
-                    coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, innerBifurcationLeft[2][n2][n3][n1])
-                    # coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS3, 1, d3SampledBifurcationRight[n1])
-                    bilNodeId.append(nodeIdentifier)
-                    nodeIdentifier += 1
-
-    # Create bifurcation nodes
-    roNodeId = []
-    for n in range(len(rox)):
-        node = nodes.createNode(nodeIdentifier, nodetemplate)
-        cache.setNode(node)
-        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, rox[n])
-        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, rod1[n])
-        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, rod2[n])
-        if n in (0, len(rox)//2):
-            coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS3, 1, rod3[n])
-        roNodeId.append(nodeIdentifier)
-        nodeIdentifier = nodeIdentifier + 1
-    coNodeId = []
-    for n in range(len(cox)):
-        node = nodes.createNode(nodeIdentifier, nodetemplate)
-        cache.setNode(node)
-        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_VALUE, 1, cox[n])
-        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS1, 1, cod1[n])
-        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS2, 1, cod2[n])
-        coordinates.setNodeParameters(cache, -1, Node.VALUE_LABEL_D_DS3, 1, cod3[n])
-        coNodeId.append(nodeIdentifier)
-        nodeIdentifier = nodeIdentifier + 1
-
-    return nodeIdentifier, roNodeId, coNodeId, birNodeId, bilNodeId
-
-
-def make_double_tube_elements(mesh, coordinates, elementIdentifier, elementsCountAlong, elementsCountAround,
-                              elementsCountAroundRightTube, elementsCountAroundLeftTube, elementsCountThroughWall,
-                              tricNodeId, tlicNodeId, toNodeId, tsNodeId, useCrossDerivatives, meshGroups=None):
-
-    """
-    Gets tube two inner canals' node ides and outer node ids and creates the elements.
-    tricNodeId, tlicNodeId: Tube's right and left inner canals node Ids.
-    toNodeId: Tube's outer node Ids.
-    tsNodeId: Tube's septum node Ids.
-    :return: elements of tube with double inner canals.
-    """
-    eftfactory = eftfactory_bicubichermitelinear(mesh, useCrossDerivatives)
-    eft = eftfactory.createEftBasic()
-
-    elementtemplate = mesh.createElementtemplate()
-    elementtemplate.setElementShapeType(Element.SHAPE_TYPE_CUBE)
-    elementtemplate.defineField(coordinates, -1, eft)
-
-    elementtemplateMod = mesh.createElementtemplate()
-    elementtemplateMod.setElementShapeType(Element.SHAPE_TYPE_CUBE)
-
-    elementsCountAcross = elementsCountAroundRightTube - elementsCountAround // 2
-
-    # Tube elements
-    for e3 in range(elementsCountThroughWall):
-        for e2 in range(elementsCountAlong):
-            # Elements around right canal in tube
-            for e1 in range(elementsCountAroundRightTube):
-                bni1 = tricNodeId[e2 * elementsCountAroundRightTube * elementsCountThroughWall + e3 * elementsCountAroundRightTube + e1]
-                bni2 = tricNodeId[e2 * elementsCountAroundRightTube * elementsCountThroughWall + e3 * elementsCountAroundRightTube + (e1 + 1) % elementsCountAroundRightTube]
-                bni3 = bni1 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                bni4 = bni2 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                if e1 < elementsCountAround // 2:
-                    if e3 == elementsCountThroughWall - 1:
-                        bni5 = toNodeId[e2 * elementsCountAround + e1]
-                    else:
-                        bni5 = bni1 + elementsCountAroundRightTube
-                    bni6 = bni5 + 1
-                    bni7 = bni5 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                    bni8 = bni7 + 1
-                elif e1 == elementsCountAround // 2:
-                    if e3 == elementsCountThroughWall - 1:
-                        bni5 = toNodeId[e2 * elementsCountAround + e1]
-                        bni6 = tsNodeId[e2 * (elementsCountAcross - 1)]
-                    else:
-                        bni5 = bni1 + elementsCountAroundRightTube
-                        bni6 = bni5 + 1
-                    bni7 = bni5 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                    bni8 = bni6 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                elif elementsCountAround // 2 < e1 < elementsCountAroundRightTube - 1:
-                    if e3 == elementsCountThroughWall - 1:
-                        bni5 = tsNodeId[e2 * (elementsCountAcross - 1) + e1 - elementsCountAround // 2 - 1]
-                        bni6 = bni5 + 1
-                    else:
-                        bni5 = bni1 + elementsCountAroundRightTube
-                        bni6 = bni2 + elementsCountAroundRightTube
-                    bni7 = bni5 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                    bni8 = bni6 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                elif e1 == elementsCountAroundRightTube - 1:
-                    if e3 == elementsCountThroughWall - 1:
-                        bni5 = tsNodeId[e2 * (elementsCountAcross - 1) + elementsCountAcross - 2]
-                        bni6 = toNodeId[e2 * elementsCountAround]
-                    else:
-                        bni5 = bni1 + elementsCountAroundRightTube
-                        bni6 = bni2 + elementsCountAroundRightTube
-                    bni7 = bni5 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                    bni8 = bni6 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-                # print('nodeIdentifiers', nodeIdentifiers)
-                # Remap derivatives for elements at the beginning and end of the septum
-                if e1 == elementsCountAround // 2 and e3 == elementsCountThroughWall - 1:
-                    scalefactors = [-1.0]
-                    eft1 = eftfactory.createEftBasic()
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [5, 7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS3, [1])])
-                    elementtemplateMod.defineField(coordinates, -1, eft1)
-                    elementtemplate1 = elementtemplateMod
-                elif e1 == elementsCountAroundRightTube - 1 and e3 == elementsCountThroughWall - 1:
-                    eft1 = eftfactory.createEftBasic()
-                    remapEftNodeValueLabel(eft1, [6, 8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS3, [])])
-                    elementtemplateMod.defineField(coordinates, -1, eft1)
-                    elementtemplate1 = elementtemplateMod
-                else:
-                    scalefactors = None
-                    eft1 = eft
-                    elementtemplate1 = elementtemplate
-                element = mesh.createElement(elementIdentifier, elementtemplate1)
-                result = element.setNodesByIdentifier(eft1, nodeIdentifiers)
-                if scalefactors:
-                    result3 = element.setScaleFactors(eft1, scalefactors)
-                else:
-                    result3 = '-'
-                for meshGroup in meshGroups:
-                    meshGroup.addElement(element)
-                elementIdentifier += 1
-
-            # Elements around Left canal in tube
-            for e1 in range(elementsCountAroundLeftTube):
-                if 0 <= e1 < elementsCountAcross:
-                    bni1 = tlicNodeId[e2 * elementsCountAroundLeftTube * elementsCountThroughWall + e3 * elementsCountAroundRightTube + e1]
-                    bni2 = tlicNodeId[e2 * elementsCountAroundLeftTube * elementsCountThroughWall + e3 * elementsCountAroundRightTube + (e1 + 1) % elementsCountAroundLeftTube]
-                    bni3 = bni1 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                    bni4 = bni2 + 2 * elementsCountAroundRightTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                    if e1 == 0:
-                        if e3 == elementsCountThroughWall - 1:
-                            bni5 = toNodeId[e2 * elementsCountAround + e1]
-                            bni6 = tsNodeId[e2 * (elementsCountAcross - 1) + elementsCountAcross - 2]
-                        else:
-                            bni5 = bni1 + elementsCountAroundLeftTube
-                            bni6 = bni2 + elementsCountAroundLeftTube
-                    elif e1 == elementsCountAcross - 1:
-                        if e3 == elementsCountThroughWall - 1:
-                            bni5 = tsNodeId[e2 * (elementsCountAcross - 1) + elementsCountAcross - 1 - e1]
-                            bni6 = toNodeId[e2 * elementsCountAround + elementsCountAround // 2]
-                        else:
-                            bni5 = bni1 + elementsCountAroundLeftTube
-                            bni6 = bni2 + elementsCountAroundLeftTube
-                    else:
-                        if e3 == elementsCountThroughWall - 1:
-                            bni5 = tsNodeId[e2 * (elementsCountAcross - 1) + elementsCountAcross - 1 - e1]
-                            bni6 = bni5 - 1
-                        else:
-                            bni5 = bni1 + elementsCountAroundLeftTube
-                            bni6 = bni2 + elementsCountAroundLeftTube
-                    bni7 = bni5 + 2 * elementsCountAroundLeftTube * elementsCountThroughWall + elementsCountAround+ elementsCountAcross - 1
-                    bni8 = bni6 + 2 * elementsCountAroundLeftTube * elementsCountThroughWall + elementsCountAround+ elementsCountAcross - 1
-                    nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-                    # print('nodeIdentifiers', nodeIdentifiers)
-                else:
-                    bni1 = tlicNodeId[e2 * elementsCountAroundLeftTube * elementsCountThroughWall + e3 * elementsCountAroundLeftTube + e1]
-                    bni2 = tlicNodeId[e2 * elementsCountAroundLeftTube * elementsCountThroughWall + e3 * elementsCountAroundLeftTube + (e1 + 1) % elementsCountAroundLeftTube]
-                    bni3 = bni1 + 2 * elementsCountAroundLeftTube * elementsCountThroughWall + elementsCountAround + elementsCountAcross - 1
-                    bni4 = bni2 + 2 * elementsCountAroundLeftTube * elementsCountThroughWall + elementsCountAround+ elementsCountAcross - 1
-                    if e3 == elementsCountThroughWall - 1:
-                        bni5 = toNodeId[e2 * elementsCountAround + e1 - elementsCountAcross + elementsCountAround // 2]
-                        if e1 == elementsCountAroundLeftTube - 1:
-                            bni6 = toNodeId[e2 * elementsCountAround]
-                        else:
-                            bni6 = toNodeId[e2 * elementsCountAround + (
-                                        e1 + 1) % elementsCountAroundLeftTube - elementsCountAcross + elementsCountAround // 2]
-                    else:
-                        bni5 = bni1 + elementsCountAroundLeftTube
-                        bni6 = bni2 + elementsCountAroundLeftTube
-                    bni7 = bni5 + 2 * elementsCountAroundLeftTube * elementsCountThroughWall + elementsCountAround+ elementsCountAcross - 1
-                    bni8 = bni6 + 2 * elementsCountAroundLeftTube * elementsCountThroughWall + elementsCountAround+ elementsCountAcross - 1
-                    nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-                    # print('nodeIdentifiers', nodeIdentifiers)
-                # Remap derivatives for elements across septum
-                if e3 == elementsCountThroughWall - 1 and e1 == 0:
-                    scalefactors = [-1.0]
-                    eft1 = eftfactory.createEftBasic()
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [6, 8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    remapEftNodeValueLabel(eft1, [5, 7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS3, [1])])
-                    elementtemplateMod.defineField(coordinates, -1, eft1)
-                    elementtemplate1 = elementtemplateMod
-                elif e3 == elementsCountThroughWall - 1 and 0 < e1 < elementsCountAcross - 1:
-                    scalefactors = [-1.0]
-                    eft1 = eftfactory.createEftBasic()
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [5, 6, 7, 8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    elementtemplateMod.defineField(coordinates, -1, eft1)
-                    elementtemplate1 = elementtemplateMod
-                elif e3 == elementsCountThroughWall - 1 and e1 == elementsCountAcross - 1:
-                    scalefactors = [-1.0]
-                    eft1 = eftfactory.createEftBasic()
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [5, 7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    remapEftNodeValueLabel(eft1, [6, 8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS3, [])])
-                    elementtemplateMod.defineField(coordinates, -1, eft1)
-                    elementtemplate1 = elementtemplateMod
-                else:
-                    scalefactors = None
-                    eft1 = eft
-                    elementtemplate1 = elementtemplate
-
-                element = mesh.createElement(elementIdentifier, elementtemplate1)
-                result = element.setNodesByIdentifier(eft1, nodeIdentifiers)
-                if scalefactors:
-                    result3 = element.setScaleFactors(eft1, scalefactors)
-                else:
-                    result3 = '-'
-                for meshGroup in meshGroups:
-                    meshGroup.addElement(element)
-                elementIdentifier += 1
-    return elementIdentifier
-
-
-def make_double_tube_bifurcation_elements(fm, coordinates, elementIdentifier, elementsCountThroughWall, c1NodeId,
-                                          c2NodeId, roNodeId, coNodeId, birNodeId, bilNodeId, pricNodeId, plicNodeId,
-                                          poNodeId, psNodeId, meshGroups=None):
-    """
-    Gets child1, child2, parent, and inner bifurcation node Ids.
-    c1NodeId, c2NodeId: Child1 and child2 node Ids.
-    roNodeId, coNodeId: 2D bifurcation node Ids which are in outer bifurcation surface.
-    birNodeId, bilNodeId: Right and left bifurcation inner node Ids.
-    pricNodeId, plicNodeId: Parent's right and left inner canals node Ids.
-    poNodeId: Parent's outer node Ids.
-    psNodeId: parent's septum node Ids.
-    :return: elements of bifurcation with double inner tube.
-    """
-    elementsCountAround = len(poNodeId)
-    elementsCountAroundRightTube = len(c1NodeId[0])
-    elementsCountAroundLeftTube = len(c2NodeId[0])
-    pac1Count, pac2Count, c1c2Count = \
-        get_tube_bifurcation_connection_elements_counts(
-            [elementsCountAround, elementsCountAroundRightTube, elementsCountAroundLeftTube])
-
-    elementsCountAcross = elementsCountAroundRightTube - elementsCountAround // 2
-
-    mesh = fm.findMeshByDimension(3)
-    eftfactory = eftfactory_bicubichermitelinear(mesh, None)
-    eftStd = eftfactory.createEftBasic()
-
-    elementtemplateStd = mesh.createElementtemplate()
-    elementtemplateStd.setElementShapeType(Element.SHAPE_TYPE_CUBE)
-    elementtemplateStd.defineField(coordinates, -1, eftStd)
-
-    elementtemplateMod = mesh.createElementtemplate()
-    elementtemplateMod.setElementShapeType(Element.SHAPE_TYPE_CUBE)
-
-    # Right tube part
-    for e3 in range(elementsCountThroughWall):
-        for e1 in range(elementsCountAroundRightTube):
-            eft = eftStd
-            elementtemplate = elementtemplateStd
-            scalefactors = None
-            if e1 < elementsCountAround // 2:
-                bni1 = c1NodeId[e3][e1]
-                bni2 = c1NodeId[e3][(e1 + 1) % elementsCountAroundRightTube]
-                bni3 = birNodeId[e3 * elementsCountAroundRightTube + e1]
-                bni4 = birNodeId[e3 * elementsCountAroundRightTube + (e1 + 1) % elementsCountAroundRightTube]
-                bni5 = bni1 + elementsCountAroundRightTube
-                bni6 = bni5 + 1
-                if e3 == elementsCountThroughWall - 1:
-                    bni7 = roNodeId[e1]
-                    bni8 = bni7 + 1
-                else:
-                    bni7 = bni3 + elementsCountAroundRightTube
-                    bni8 = bni4 + elementsCountAroundRightTube
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-                # print('nodeIdentifiers', nodeIdentifiers)
-            elif e1 == elementsCountAround // 2:
-                bni1 = c1NodeId[e3][e1]
-                bni2 = c1NodeId[e3][(e1 + 1) % elementsCountAroundRightTube]
-                bni3 = birNodeId[e3 * elementsCountAroundRightTube + e1]
-                bni4 = birNodeId[e3 * elementsCountAroundRightTube + (e1 + 1) % elementsCountAroundRightTube]
-                bni5 = bni1 + elementsCountAroundRightTube
-                bni6 = bni5 + 1
-                if e3 == elementsCountThroughWall - 1:
-                    bni7 = roNodeId[e1]
-                    bni8 = coNodeId[0]
-                else:
-                    bni7 = bni3 + elementsCountAroundRightTube
-                    bni8 = bni4 + elementsCountAroundRightTube
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-                # print('nodeIdentifiers', nodeIdentifiers)
-            else:
-                bni1 = c1NodeId[e3][e1]
-                bni2 = c1NodeId[e3][(e1 + 1) % elementsCountAroundRightTube]
-                bni3 = birNodeId[e3 * elementsCountAroundRightTube + e1]
-                bni5 = bni1 + elementsCountAroundRightTube
-                if e3 == elementsCountThroughWall - 1:
-                    bni7 = coNodeId[e1 - elementsCountAround // 2 - 1]
-                else:
-                    bni7 = bni3 + elementsCountAroundRightTube
-                if e1 == elementsCountAroundRightTube - 1:
-                    bni4 = birNodeId[0 + e3 * elementsCountAroundRightTube]
-                    bni6 = bni5 - elementsCountAroundRightTube + 1
-                    if e3 == elementsCountThroughWall - 1:
-                        bni8 = roNodeId[0]
-                    else:
-                        bni8 = bni4 + elementsCountAroundRightTube
-                else:
-                    bni4 = bni3 + 1
-                    bni6 = bni5 + 1
-                    bni8 = bni7 + 1
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-                # print('nodeIdentifiers', nodeIdentifiers)
-            # Remap the derivatives
-            if e3 == elementsCountThroughWall - 1 and e1 in (0, pac1Count - 1, pac1Count, elementsCountAroundRightTube - 1):
-                eft1 = eftfactory.createEftBasic()
-                if e1 == 0:
-                    scalefactors = [-1.0]
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS1,
-                                           [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS1, [1])])
-                elif e1 == pac1Count - 1:
-                    remapEftNodeValueLabel(eft1, [8], Node.VALUE_LABEL_D_DS2,
-                                           [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                elif e1 == pac1Count:
-                    scalefactors = [-1.0]
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS2,
-                                           [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [1])])
-                elif e1 == elementsCountAroundRightTube - 1:
-                    scalefactors = [-1.0]
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [8], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    remapEftNodeValueLabel(eft1, [8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [])])
-                elementtemplateMod.defineField(coordinates, -1, eft1)
-                elementtemplate1 = elementtemplateMod
-            else:
-                scalefactors = None
-                eft1 = eft
-                elementtemplate1 = elementtemplate
-            element = mesh.createElement(elementIdentifier, elementtemplate1)
-            result2 = element.setNodesByIdentifier(eft1, nodeIdentifiers)
-            if scalefactors:
-                result3 = element.setScaleFactors(eft1, scalefactors)
-            else:
-                result3 = '-'
-            elementIdentifier += 1
-            for meshGroup in meshGroups:
-                if meshGroups.index(meshGroup) == 1:
-                    meshGroup.addElement(element)
-                elif meshGroups.index(meshGroup) == 3:
-                    meshGroup.addElement(element)
-
-    # Left tube part
-    for e3 in range(elementsCountThroughWall):
-        for e1 in range(elementsCountAroundLeftTube):
-            eft = eftStd
-            elementtemplate = elementtemplateStd
-            scalefactors = None
-            bni1 = c2NodeId[e3][e1]
-            bni2 = c2NodeId[e3][(e1 + 1) % elementsCountAroundLeftTube]
-            bni3 = bilNodeId[e3 * elementsCountAroundLeftTube + e1]
-            bni4 = bilNodeId[e3 * elementsCountAroundLeftTube + (e1 + 1) % elementsCountAroundLeftTube]
-            bni5 = bni1 + elementsCountAroundLeftTube
-            bni6 = bni2 + elementsCountAroundLeftTube
-            if 0 <= e1 < elementsCountAcross:
-                if e3 != elementsCountThroughWall - 1:
-                    bni7 = bni3 + elementsCountAroundLeftTube
-                    bni8 = bni4 + elementsCountAroundLeftTube
-                else:
-                    if e1 == 0:
-                        bni7 = roNodeId[0]
-                        bni8 = coNodeId[-1]
-                    elif e1 == elementsCountAcross - 1:
-                        bni7 = coNodeId[elementsCountAcross - 1 - e1]
-                        bni8 = roNodeId[elementsCountAround // 2]
-                    else:
-                        bni7 = coNodeId[elementsCountAcross - 1 - e1]
-                        bni8 = bni7 - 1
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-                # print('nodeIdentifiers', nodeIdentifiers)
-            else:
-                if e3 == elementsCountThroughWall - 1:
-                    bni7 = roNodeId[elementsCountAround // 2 + e1 - elementsCountAcross]
-                    if e1 == elementsCountAroundLeftTube - 1:
-                        bni8 = roNodeId[0]
-                    else:
-                        bni8 = bni7 + 1
-                else:
-                    bni7 = bni3 + elementsCountAroundLeftTube
-                    bni8 = bni4 + elementsCountAroundLeftTube
-                nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-                # print('nodeIdentifiers', nodeIdentifiers)
-            # Remap the derivatives
-            if e3 == elementsCountThroughWall - 1 and 0 <= e1 < elementsCountAcross + 1:
-                eft1 = eftfactory.createEftBasic()
-                if e1 == 0:
-                    scalefactors = [-1.0]
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [1])])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS2,
-                                           [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                    remapEftNodeValueLabel(eft1, [8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    remapEftNodeValueLabel(eft1, [8], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS2, [1])])
-                elif 0 < e1 < elementsCountAcross - 1:
-                    scalefactors = [-1.0]
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [7, 8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    remapEftNodeValueLabel(eft1, [7, 8], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS2, [1])])
-                elif e1 == elementsCountAcross - 1:
-                    scalefactors = [-1.0]
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS2, [1])])
-                    remapEftNodeValueLabel(eft1, [8], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS1, [1])])
-                    remapEftNodeValueLabel(eft1, [8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [])])
-                elif e1 == elementsCountAcross:
-                    scalefactors = [-1.0]
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS1, [1])])
-                elementtemplateMod.defineField(coordinates, -1, eft1)
-                elementtemplate1 = elementtemplateMod
-            elif e3 == elementsCountThroughWall - 1 and e1 == elementsCountAroundLeftTube - 1:
-                eft1 = eftfactory.createEftBasic()
-                remapEftNodeValueLabel(eft1, [8], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                elementtemplateMod.defineField(coordinates, -1, eft1)
-                elementtemplate1 = elementtemplateMod
-            else:
-                scalefactors = None
-                eft1 = eft
-                elementtemplate1 = elementtemplate
-            element = mesh.createElement(elementIdentifier, elementtemplate1)
-            result2 = element.setNodesByIdentifier(eft1, nodeIdentifiers)
-            if scalefactors:
-                result3 = element.setScaleFactors(eft1, scalefactors)
-            else:
-                result3 = '-'
-            elementIdentifier += 1
-            for meshGroup in meshGroups:
-                if meshGroups.index(meshGroup) == 2:
-                    meshGroup.addElement(element)
-                elif meshGroups.index(meshGroup) == 3:
-                    meshGroup.addElement(element)
-
-    # parent right part
-    for e3 in range(elementsCountThroughWall):
-        for e1 in range(elementsCountAroundRightTube):
-            eft = eftStd
-            elementtemplate = elementtemplateStd
-            scalefactors = None
-            bni1 = birNodeId[e3 * elementsCountAroundRightTube + e1]
-            bni2 = birNodeId[e3 * elementsCountAroundRightTube + (e1 + 1) % elementsCountAroundRightTube]
-            bni3 = pricNodeId[e3 * elementsCountAroundRightTube + e1]
-            bni4 = pricNodeId[e3 * elementsCountAroundRightTube + (e1 + 1) % elementsCountAroundRightTube]
-            if e3 == elementsCountThroughWall - 1:
-                if e1 < elementsCountAround // 2:
-                        bni5 = roNodeId[e1]
-                        bni6 = roNodeId[(e1 + 1) % elementsCountAround]
-                        bni7 = poNodeId[e1]
-                        bni8 = poNodeId[(e1 + 1) % elementsCountAround]
-                elif e1 == elementsCountAround // 2:
-                    bni5 = roNodeId[e1]
-                    bni6 = coNodeId[e1 - elementsCountAround // 2]
-                    bni7 = poNodeId[e1]
-                    bni8 = psNodeId[e1 - elementsCountAround // 2]
-                elif elementsCountAround // 2 < e1 < elementsCountAroundRightTube - 1:
-                    bni5 = coNodeId[e1 - elementsCountAround // 2 - 1]
-                    bni6 = bni5 + 1
-                    bni7 = psNodeId[e1 - elementsCountAround // 2 - 1]
-                    bni8 = bni7 + 1
-                elif e1 == elementsCountAroundRightTube - 1:
-                    bni5 = coNodeId[-1]
-                    bni6 = roNodeId[0]
-                    bni8 = poNodeId[0]
-                    bni7 = bni8 - 1
-            else:
-                bni5 = bni1 + elementsCountAroundRightTube
-                bni6 = bni2 + elementsCountAroundRightTube
-                bni7 = bni3 + elementsCountAroundRightTube
-                bni8 = bni4 + elementsCountAroundRightTube
-            nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-            # print('nodeIdentifiers', nodeIdentifiers)
-            # Remap the derivatives
-            if e3 == elementsCountThroughWall - 1:
-                if e1 == 0:
-                    eft1 = eftfactory.createEftBasic()
-                    remapEftNodeValueLabel(eft1, [5], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                    elementtemplateMod.defineField(coordinates, -1, eft1)
-                    elementtemplate1 = elementtemplateMod
-                elif e1 == elementsCountAround // 2:
-                    scalefactors = [-1.0]
-                    eft1 = eftfactory.createEftBasic()
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [5], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [1])])
-                    remapEftNodeValueLabel(eft1, [6], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS3, [1])])
-                    remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS3, [1])])
-                    elementtemplateMod.defineField(coordinates, -1, eft1)
-                    elementtemplate1 = elementtemplateMod
-                elif elementsCountAround // 2 < e1 < elementsCountAroundRightTube - 1:
-                    scalefactors = [-1.0]
-                    eft1 = eftfactory.createEftBasic()
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [5, 6], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS3, [1])])
-                    elementtemplateMod.defineField(coordinates, -1, eft1)
-                    elementtemplate1 = elementtemplateMod
-                elif e1 == elementsCountAroundRightTube - 1:
-                    scalefactors = [-1.0]
-                    eft1 = eftfactory.createEftBasic()
-                    setEftScaleFactorIds(eft1, [1], [])
-                    remapEftNodeValueLabel(eft1, [5], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS3, [1])])
-                    remapEftNodeValueLabel(eft1, [6], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [])])
-                    remapEftNodeValueLabel(eft1, [8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS3, [])])
-                    elementtemplateMod.defineField(coordinates, -1, eft1)
-                    elementtemplate1 = elementtemplateMod
-                else:
-                    scalefactors = None
-                    eft1 = eft
-                    elementtemplate1 = elementtemplate
-            else:
-                scalefactors = None
-                eft1 = eft
-                elementtemplate1 = elementtemplate
-            element = mesh.createElement(elementIdentifier, elementtemplate1)
-            result = element.setNodesByIdentifier(eft1, nodeIdentifiers)
-            if scalefactors:
-                result3 = element.setScaleFactors(eft1, scalefactors)
-            else:
-                result3 = '-'
-            elementIdentifier += 1
-            for meshGroup in meshGroups:
-                if meshGroups.index(meshGroup) == 0:
-                    meshGroup.addElement(element)
-                elif meshGroups.index(meshGroup) == 3:
-                    meshGroup.addElement(element)
-
-    # parent left part
-    for e3 in range(elementsCountThroughWall):
-        for e1 in range(elementsCountAroundLeftTube):
-            eft = eftStd
-            elementtemplate = elementtemplateStd
-            scalefactors = None
-            bni1 = bilNodeId[e3 * elementsCountAroundRightTube + e1]
-            bni2 = bilNodeId[e3 * elementsCountAroundRightTube + (e1 + 1) % elementsCountAroundLeftTube]
-            bni3 = plicNodeId[e3 * elementsCountAroundRightTube + e1]
-            bni4 = plicNodeId[e3 * elementsCountAroundRightTube + (e1 + 1) % elementsCountAroundLeftTube]
-            if e3 == elementsCountThroughWall - 1:
-                if e1 == 0:
-                    bni5 = roNodeId[0]
-                    bni6 = coNodeId[-1]
-                    bni7 = poNodeId[e1]
-                    bni8 = psNodeId[e1 + elementsCountAcross - 2]
-                elif 0 < e1 < elementsCountAcross - 1:
-                    bni5 = coNodeId[-e1 + elementsCountAcross - 1]
-                    bni6 = bni5 - 1
-                    bni7 = psNodeId[-e1 + elementsCountAcross - 1]
-                    bni8 = bni7 - 1
-                elif e1 == elementsCountAcross - 1:
-                    bni5 = coNodeId[0]
-                    bni6 = roNodeId[elementsCountAround // 2]
-                    bni7 = psNodeId[0]
-                    bni8 = poNodeId[elementsCountAround // 2]
-                else:
-                    bni5 = roNodeId[elementsCountAround // 2 - elementsCountAcross + e1]
-                    bni7 = poNodeId[elementsCountAround // 2 - elementsCountAcross + e1]
-                    if e1 == elementsCountAroundLeftTube - 1:
-                        bni6 = roNodeId[0]
-                        bni8 = poNodeId[0]
-                    else:
-                        bni6 = bni5 + 1
-                        bni8 = bni7 + 1
-            else:
-                bni5 = bni1 + elementsCountAroundRightTube
-                bni6 = bni2 + elementsCountAroundRightTube
-                bni7 = bni3 + elementsCountAroundRightTube
-                bni8 = bni4 + elementsCountAroundRightTube
-            nodeIdentifiers = [bni1, bni2, bni3, bni4, bni5, bni6, bni7, bni8]
-            # print('nodeIdentifiers', nodeIdentifiers)
-            # Remap the derivatives
-            if e3 == elementsCountThroughWall - 1:
-                if 0 <= e1 < elementsCountAcross + 1:
-                    eft1 = eftfactory.createEftBasic()
-                    if e1 == 0:
-                        scalefactors = [-1.0]
-                        setEftScaleFactorIds(eft1, [1], [])
-                        remapEftNodeValueLabel(eft1, [5], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [1])])
-                        remapEftNodeValueLabel(eft1, [7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS3, [1])])
-                        remapEftNodeValueLabel(eft1, [6, 8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, [1])])
-                        remapEftNodeValueLabel(eft1, [6], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS3, [1])])
-                    elif 0 < e1 < elementsCountAcross - 1:
-                        scalefactors = [-1.0]
-                        setEftScaleFactorIds(eft1, [1], [])
-                        remapEftNodeValueLabel(eft1, [5, 6, 7, 8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, [1])])
-                        remapEftNodeValueLabel(eft1, [5, 6], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS3, [1])])
-                    elif e1 == elementsCountAcross - 1:
-                        scalefactors = [-1.0]
-                        setEftScaleFactorIds(eft1, [1], [])
-                        remapEftNodeValueLabel(eft1, [5, 7], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, [1])])
-                        remapEftNodeValueLabel(eft1, [5], Node.VALUE_LABEL_D_DS2, [(Node.VALUE_LABEL_D_DS3, [1])])
-                        remapEftNodeValueLabel(eft1, [6], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS2, [])])
-                        remapEftNodeValueLabel(eft1, [8], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS3, [])])
-                    elif e1 == elementsCountAcross:
-                        remapEftNodeValueLabel(eft1, [5], Node.VALUE_LABEL_D_DS1, [(Node.VALUE_LABEL_D_DS1, []), (Node.VALUE_LABEL_D_DS2, [])])
-                    elementtemplateMod.defineField(coordinates, -1, eft1)
-                    elementtemplate1 = elementtemplateMod
-                else:
-                    scalefactors = None
-                    eft1 = eft
-                    elementtemplate1 = elementtemplate
-            else:
-                scalefactors = None
-                eft1 = eft
-                elementtemplate1 = elementtemplate
-            element = mesh.createElement(elementIdentifier, elementtemplate1)
-            result = element.setNodesByIdentifier(eft1, nodeIdentifiers)
-            if scalefactors:
-                result3 = element.setScaleFactors(eft1, scalefactors)
-            else:
-                result3 = '-'
-            elementIdentifier += 1
-            for meshGroup in meshGroups:
-                if meshGroups.index(meshGroup) == 0:
-                    meshGroup.addElement(element)
-                elif meshGroups.index(meshGroup) == 3:
-                    meshGroup.addElement(element)
-
-    return elementIdentifier
+    field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_D_DS1, version, d1)
+    field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_D_DS2, version, d2)
+    field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_D_DS3, version, d3)
+    if d12:
+        field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_D2_DS1DS2, version, d12)
+    if d13:
+        field.setNodeParameters(fieldcache, -1, Node.VALUE_LABEL_D2_DS1DS3, version, d13)
